@@ -6,47 +6,53 @@ description: A new Flutter application.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.2+8
+# F:\flutter_sdks\sdk\flutter_2.5.3\bin\flutter.bat --no-color build apk --target-platform android-arm,android-arm64,android-x64 --split-per-abi --no-tree-shake-icons
+version: 2.0.0+60
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_ml_vision: ^0.9.5
-  firebase_core: ^0.4.5
-  camera: ^0.6.4+5
-  provider: ^4.3.2
-  localstorage: ^3.0.0
-  shared_preferences: ^0.5.12+4
-  rounded_loading_button: ^1.0.18
-  permission_handler: ^5.0.1+1
-  sliding_up_panel: ^1.0.2
-  firebase_messaging: ^6.0.16
-  pin_code_text_field: ^1.7.1
-  carousel_slider: ^2.3.1
-  cached_network_image: ^2.5.0
-  pull_to_refresh: 1.6.3
-  image_picker: ^0.6.7+21
-  geolocator: ^6.1.14
-  intl: ^0.16.1
-  flutter_date_pickers: 0.1.4
-  google_maps_flutter: ^1.1.1
-  flutter_slidable: ^0.5.7
-  syncfusion_flutter_pdfviewer: ^18.4.44-beta
-  flutter_nfc_reader: ^0.1.0
-  flutter_colorpicker: 0.3.5
-  image_cropper: 1.3.1
-  audioplayers: 0.17.4
+  google_ml_kit: ^0.7.3
+  camera: ^0.9.4+16
+  provider: ^6.0.1
+  localstorage: ^4.0.0+1
+  shared_preferences: ^2.0.8
+  rounded_loading_button: ^2.0.9
+  permission_handler: ^8.3.0
+  sliding_up_panel: ^2.0.0+1
+  firebase_messaging: ^11.1.0
+  pin_code_text_field: ^1.8.0
+  carousel_slider: ^4.0.0
+  cached_network_image: ^3.1.0+1
+  pull_to_refresh: ^2.0.0
+  image_picker: ^0.8.4+4
+  geolocator: ^7.7.1
+  intl: ^0.17.0
+  flutter_date_pickers: ^0.2.6
+  google_maps_flutter: ^2.1.1
+  flutter_slidable: ^1.1.0
+  syncfusion_flutter_pdfviewer: ^19.3.54-beta
+  flutter_colorpicker: ^1.0.2
   flutter_localizations:
     sdk: flutter
-  wakelock: 0.2.1+1
+  wakelock: ^0.5.6
+  url_launcher: ^6.0.13
+  path_provider: ^2.0.7
+  path: ^1.8.0
+  nfc_manager: ^3.1.0
+  image_cropper: ^1.4.1
+  device_info: ^2.0.3
+  assets_audio_player: ^3.0.4+1
+  loader_overlay: ^2.0.5
+  dropdown_search: ^2.0.1
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  path_provider: ^1.6.11
-  path: ^1.6.2
+
 flutter:
   uses-material-design: true
 
