@@ -1,540 +1,66 @@
-import 'package:facepunch/widgets/calendar_strip/date-utils.dart';
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:http/http.dart' as http;
 import 'dart:convert';
-import 'app_const.dart';
-import 'package:localstorage/localstorage.dart';
-import 'package:firebase_messaging/firebase_messaging.dart';
-
+import 'package:collection/collection.dart';
 import 'work_model.dart';
+import '/widgets/utils.dart';
+import '/lang/l10n.dart';
+import 'revision_model.dart';
+import '/widgets/calendar_strip/date-utils.dart';
+import '/config/app_const.dart';
+import '/providers/base_provider.dart';
 
+class User with HttpRequest{
+  int? id;
+  String? name;
+  String? email;
+  String? emailVerifiedAt;
+  String? firstName;
+  String? lastName;
+  String? phone;
+  String? pin;
+  String? employeeCode;
+  String? start;
+  String? salary;
+  String? birthday;
+  String? address1;
+  String? address2;
+  String? city;
+  String? country;
+  String? state;
+  String? postalCode;
+  String? language;
+  String? avatar;
+  String? avatarUrl;
+  String? role;
+  String? type;
+  String? firebaseToken;
+  String? emailVerifyNumber;
+  int? companyId;
+  String? nfc;
+  String? createdAt;
+  String? updatedAt;
+  bool? active = true;
+  BreakSetting? breakSetting;
 
-class UserModel with ChangeNotifier{
-  User user;
-  final LocalStorage storage = LocalStorage('face_punch_user');
-  final FirebaseMessaging _firebaseMessaging = FirebaseMessaging();
-  String locale = 'en';
-
-  Future<User> getUserFromLocal()async{
-    try{
-      bool storageReady = await storage.ready;
-      if(storageReady){
-        var json = await storage.getItem('user');
-        if(json!=null){
-          user = User.fromJson(json);
-          if(user.language=='Spanish'){
-            locale = 'es';
-          }else if(user.language=='French'){
-            locale = 'fr';
-          }
-          GlobalData.token = user.token;
-        }
-      }
-    }catch(e){
-      print("[UserModel.getUserFromLocal] $e");
-    }
-    notifyListeners();
-    return user;
-  }
-
-  saveUserToLocal()async{
-    try{
-      bool storageReady = await storage.ready;
-      if(storageReady)
-        await storage.setItem('user', user.toJson());
-      if(user.language=='Spanish'){
-        locale = 'es';
-      }else if(user.language=='French'){
-        locale = 'fr';
-      }else{
-        locale = 'en';
-      }
-      notifyListeners();
-    }catch(e){
-      print("[UserModel.saveUserToLocal] $e");
-    }
-  }
-
-  Future<String> adminLogin({String email,String password, bool isRememberMe})async{
-    String result = 'Oops, Unknown Errors!';
-    try{
-      String deviceToken = await _firebaseMessaging.getToken();
-      var res = await http.post(
-          AppConst.adminLogin,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-          },
-          body: {
-            'email':email,
-            'password':password,
-            'firebase_token': deviceToken??''
-          }
-      );
-      print("[UserModel.adminLogin] ${res.body}");
-      if(res.statusCode==200){
-        user = User.fromJson(jsonDecode(res.body));
-        GlobalData.token = user.token;
-        if(isRememberMe)await saveUserToLocal();
-        result = null;
-      }else{
-        result =  jsonDecode(res.body)['message'].toString();
-      }
-    }catch(e){
-      print("[UserModel.adminLogin] $e");
-    }
-    return result;
-  }
-
-  Future<String> adminRegister(String email,String password,String fName,String lName)async{
-    String result = 'Oops, Unknown Errors!';
-    try{
-      String deviceToken  = await _firebaseMessaging.getToken();
-      var res = await http.post(
-          AppConst.adminRegister,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-          },
-          body: {
-            'email':email,
-            'password':password,
-            'first_name':fName,
-            'last_name':lName,
-            'firebase_token':deviceToken??''
-          }
-      );
-      print("[UserModel.adminRegister] ${res.body}");
-      if(res.statusCode==200){
-        user = User.fromJson(jsonDecode(res.body));
-        GlobalData.token = user.token;
-        await saveUserToLocal();
-        result =  null;
-      }else{
-        result =  jsonDecode(res.body)['message'].toString();
-      }
-    }catch(e){
-      print("[UserModel.adminRegister] $e");
-    }
-    return result;
-  }
-
-  logOut()async{
-    try{
-      bool storageReady = await storage.ready;
-      if(storageReady){
-        await storage.deleteItem('user');
-        user = null;
-      }
-    }catch(e){
-      print("[UserModel.logOut] $e");
-    }
-  }
-
-  Future<String> recoverPassword(String email)async{
-    try{
-      var res = await http.post(
-          AppConst.recoverPassword,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-          },
-          body: {
-            'email':email
-          }
-      );
-      print("[UserModel.recoverPassword] ${res.body}");
-      if(res.statusCode==200){
-        return "We have sent new password to your email.";
-      }else{
-        return jsonDecode(res.body)['message'].toString();
-      }
-    }catch(e){
-      print("[UserModel.recoverPassword] $e");
-      return 'Oops, Unknown Errors!';
-    }
-  }
-
-  Future<String> verifyEmailAddress(String number)async{
-    String result = 'Oops, Unknown Errors!';
-    try{
-      var res = await http.post(
-          AppConst.emailVerify,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-            'Authorization':'Bearer '+user.token
-          },
-          body: {
-            'email' : user.email,
-            'verify_number' : number,
-          }
-      );
-      print("[UserModel.verifyEmailAddress] ${res.body}");
-      if(res.statusCode==200){
-        user = User.fromJson(jsonDecode(res.body));
-        GlobalData.token = user.token;
-        await saveUserToLocal();
-        result = null;
-      }else{
-        result =  jsonDecode(res.body)['message'].toString();
-      }
-    }catch(e){
-      print("[UserModel.verifyEmailAddress] $e");
-    }
-    return result;
-  }
-
-  Future<String> sendVerifyEmailAgain()async{
-    String result = 'Oops, Unknown Errors!';
-    try{
-      var res = await http.get(
-          AppConst.sendVerifyAgain,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-            'Authorization':'Bearer '+user.token
-          }
-      );
-      print("[UserModel.sendVerifyEmailAgain] ${res.body}");
-      if(res.statusCode==200){
-        user.emailVerifyNumber = jsonDecode(res.body)['number'];
-        await saveUserToLocal();
-        result = null;
-      }else{
-        result =  jsonDecode(res.body)['message'].toString();
-      }
-    }catch(e){
-      print("[UserModel.sendVerifyEmailAgain] $e");
-    }
-    return result;
-  }
-
-  Future<String> updateAdmin({String email, String newPassword, String oldPassword, String fName, String lName})async{
-    String result = 'Oops, Unknown Errors!';
-    try{
-      var res = await http.post(
-          AppConst.updateAdmin,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-            'Authorization':'Bearer '+user.token
-          },
-          body: {
-            'email':email,
-            'new_password' : newPassword,
-            'old_password' : oldPassword,
-            'first_name' : fName,
-            'last_name' : lName
-          }
-      );
-      print("[UserModel.updateUser] ${res.body}");
-      if(res.statusCode==200){
-        user = User.fromJson(jsonDecode(res.body));
-        GlobalData.token = user.token;
-        await saveUserToLocal();
-        result =  null;
-      }else{
-        result =  jsonDecode(res.body)['message'].toString();
-      }
-    }catch(e){
-      print("[UserModel.updateUser] $e");
-    }
-    return result;
-  }
-
-  Future<String> loginWithFace(String photo)async{
-    String result = 'Oops, Unknown Errors!';
-    try{
-      String deviceToken = await _firebaseMessaging.getToken();
-      print("[deviceToken] $deviceToken");
-      var res = await http.post(
-          AppConst.loginWithFace,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-          },
-          body: {
-            'photo':photo,
-            'firebase_token': deviceToken??''
-          }
-      );
-      print("[UserModel.loginWithFace] ${res.body}");
-      if(res.statusCode==200){
-        user = User.fromJson(jsonDecode(res.body));
-        GlobalData.token = user.token;
-        await saveUserToLocal();
-        result = null;
-      }else{
-        result = jsonDecode(res.body)['message'];
-      }
-    }catch(e){
-      print("[UserModel.loginWithFace] $e");
-    }
-    return result;
-  }
-
-  Future<dynamic> punchWithFace({String photo,double longitude, double latitude})async{
-    String result = 'Oops, Unknown Errors!';
-    try{
-      var res = await http.post(
-          AppConst.punchWithFace,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/json',
-          },
-          body: jsonEncode({
-            'photo':photo,
-            'longitude':longitude,
-            'latitude':latitude
-          })
-      );
-      print("[UserModel.punchWithFace] ${res.body}");
-      if(res.statusCode==200){
-        return jsonDecode(res.body);
-      }else{
-        result = jsonDecode(res.body)['message'];
-      }
-    }catch(e){
-      print("[UserModel.punchWithFace] $e");
-    }
-    return result;
-  }
-
-  Future<String> getUserPunches()async{
-    try{
-      var res = await http.get(
-          AppConst.getUserPunches,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/json',
-            'Authorization':'Bearer ${user.token}'
-          }
-      );
-      print("[UserModel.getUserPunches] ${res.body}");
-      if(res.statusCode==200){
-        List<Punch> punches = [];
-        for(var punch in jsonDecode(res.body)){
-          punches.add(Punch.fromJson(punch));
-        }
-        user.punches = punches;
-        notifyListeners();
-        return null;
-      }else{
-        return jsonDecode(res.body)['message'];
-      }
-    }catch(e){
-      print("[UserModel.getUserPunches] $e");
-      return e.toString();
-    }
-  }
-
-  Future<String> getUserWorkHistory()async{
-    try{
-      var res = await http.get(
-          AppConst.getUserWorks,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/json',
-            'Authorization':'Bearer ${user.token}'
-          }
-      );
-      print("[UserModel.getUserWorkHistory] ${res.body}");
-      if(res.statusCode==200){
-        List<WorkHistory> works = [];
-        for(var work in jsonDecode(res.body)){
-          works.add(WorkHistory.fromJson(work));
-        }
-        user.works = works;
-        notifyListeners();
-        return null;
-      }else{
-        return jsonDecode(res.body)['message'];
-      }
-    }catch(e){
-      print("[UserModel.getUserWorkHistory] $e");
-      return e.toString();
-    }
-  }
-
-  Future<List<Punch>> getEmployeePunches(int userId)async{
-    try{
-      var res = await http.post(
-          AppConst.getEmployeePunches,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/json',
-            'Authorization':'Bearer '+user.token
-          },
-        body: jsonEncode({'user_id':userId}),
-      );
-      print("[UserModel.getEmployeePunches] ${res.body}");
-      if(res.statusCode==200){
-        List<Punch> punches = [];
-        for(var punch in jsonDecode(res.body)){
-          punches.add(Punch.fromJson(punch));
-        }
-        return punches;
-      }
-    }catch(e){
-      print("[UserModel.getEmployeePunches] $e");
-    }
-    return [];
-  }
-
-  Future<List<WorkHistory>> getEmployeeWorks(int userId)async{
-    try{
-      var res = await http.post(
-          AppConst.getEmployeeWorks,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/json',
-            'Authorization':'Bearer '+user.token
-          },
-        body: jsonEncode({'user_id':userId}),
-      );
-      print("[UserModel.getEmployeeWorks] ${res.body}");
-      if(res.statusCode==200){
-        List<WorkHistory> works = [];
-        for(var work in jsonDecode(res.body)){
-          works.add(WorkHistory.fromJson(work));
-        }
-        return works;
-      }
-    }catch(e){
-      print("[UserModel.getEmployeeWorks] $e");
-    }
-    return [];
-  }
-
-  Future<String> notificationSetting()async{
-    try{
-      String token = "disabled";
-      if(user.firebaseToken==null || user.firebaseToken=="disabled"){
-        token = await _firebaseMessaging.getToken();
-      }
-      var res = await http.post(
-        AppConst.notificationSetting,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/json',
-          'Authorization':'Bearer '+user.token
-        },
-        body: jsonEncode({'token':token}),
-      );
-      print("[UserModel.notificationSetting] ${res.body}");
-      if(res.statusCode==200){
-        user.firebaseToken = token;
-        saveUserToLocal();
-        return null;
-      }else{
-        return jsonDecode(res.body)['message'];
-      }
-    }catch(e){
-      print("[UserModel.notificationSetting]");
-      return e.toString();
-    }
-  }
-
-  changeAppLanguage(String lang){
-    locale = lang;
-    notifyListeners();
-  }
-
-  Future<String> startWork({User employee, int taskId, int projectId})async{
-    try{
-      var res = await http.post(
-          AppConst.startWork,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/json',
-            'Authorization':'Bearer '+employee.token
-          },
-          body: jsonEncode({'task_id':taskId,'project_id':projectId})
-      );
-      print("[UserModel.startWork] ${res.body}");
-      if(res.statusCode!=200){
-        return jsonDecode(res.body)['message'];
-      }
-    }catch(e){
-      print("[UserModel.startWork] $e");
-    }
-    return null;
-  }
-  Future punchOut({User employee, double longitude, double latitude})async{
-    try{
-      var res = await http.post(
-          AppConst.punchOut,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/json',
-            'Authorization':'Bearer '+employee.token
-          },
-          body: jsonEncode({
-            'longitude':longitude,
-            'latitude':latitude,
-          })
-      );
-      print("[UserModel.punchOut] ${res.body}");
-      if(res.statusCode==200){
-        return Punch.fromJson(jsonDecode(res.body));
-      }else{
-        return jsonDecode(res.body)['message'];
-      }
-    }catch(e){
-      print("[UserModel.punchOut] $e");
-    }
-  }
-}
-
-class User {
-  int id;
-  String name;
-  String email;
-  String emailVerifiedAt;
-  String firstName;
-  String lastName;
-  String phone;
-  String pin;
-  String function;
-  String start;
-  String salary;
-  String birthday;
-  String address1;
-  String address2;
-  String city;
-  String country;
-  String state;
-  String postalCode;
-  String language;
-  String avatar;
-  String role;
-  String type;
-  String firebaseToken;
-  String emailVerifyNumber;
-  int companyId;
-  String employeeCode;
-  int lunchTime=0;
-  String nfc;
-  String token;
-  String createdAt;
-  String updatedAt;
   List<Punch> punches = [];
   List<WorkHistory> works = [];
-  Punch lastPunch;
-  bool canNTCTracking = true;
-  bool sendScheduleNotification = true;
+  List<EmployeeBreak> breaks = [];
+  List<WorkSchedule> schedules = [];
+  List<EmployeeCall> calls = [];
+  Punch? lastPunch;
+  String? token;
+
+  bool? canNTCTracking = true;
 
   User({
     this.id,
     this.name,
     this.email,
-    this.emailVerifiedAt,
     this.firstName,
     this.lastName,
     this.phone,
     this.pin,
-    this.function,
     this.start,
     this.salary,
     this.birthday,
@@ -548,20 +74,11 @@ class User {
     this.avatar,
     this.role,
     this.type,
-    this.firebaseToken,
-    this.emailVerifyNumber,
     this.companyId,
     this.employeeCode,
-    this.token,
     this.nfc,
-    this.punches,
-    this.lastPunch,
-    this.lunchTime,
-    this.createdAt,
-    this.updatedAt,
-    this.works,
-    this.canNTCTracking,
-    this.sendScheduleNotification
+    this.token,
+    this.active
   });
 
   User.fromJson(Map<String, dynamic> json) {
@@ -574,7 +91,6 @@ class User {
       lastName = json['last_name'];
       phone = json['phone'];
       pin = json['pin'];
-      function = json['function'];
       start = json['start'];
       salary = json['salary'];
       birthday = json['birthday'];
@@ -589,9 +105,7 @@ class User {
         language = 'English';
       }
       avatar = json['avatar'];
-      if(avatar==null){
-        avatar = 'user_avatar.png';
-      }
+      avatarUrl = json['avatar_url'];
       role = json['role'];
       type = json['type'];
       firebaseToken = json['firebase_token'];
@@ -602,14 +116,16 @@ class User {
       if(json['last_punch']!=null){
         lastPunch = Punch.fromJson(json['last_punch']);
       }
-      lunchTime = json['lunch_time'];
       nfc = json['nfc'];
-      canNTCTracking = json['can_nfc_tracking']!=null && json['can_nfc_tracking']==1;
-      sendScheduleNotification = json['send_schedule_notification']==1 && json['send_schedule_notification']==1;
       createdAt = json['created_at'];
       updatedAt = json['updated_at'];
+      active = json['active']??false;
+      canNTCTracking = json['can_nfc_tracking']??false;
+      if(json['break'] != null){
+        breakSetting = BreakSetting.fromJson(json['break']);
+      }
     }catch(e){
-      print("[User.fromJson] $e");
+      Tools.consoleLog("[User.fromJson.err] $e");
     }
   }
 
@@ -623,7 +139,6 @@ class User {
     data['last_name'] = this.lastName;
     data['phone'] = this.phone;
     data['pin'] = this.pin;
-    data['function'] = this.function;
     data['start'] = this.start;
     data['salary'] = this.salary;
     data['birthday'] = this.birthday;
@@ -634,7 +149,8 @@ class User {
     data['state'] = this.state;
     data['postal_code'] = this.postalCode;
     if(language!=null)data['language'] = this.language;
-    if(avatar!=null)data['avatar'] = this.avatar;
+    data['avatar'] = this.avatar;
+    data['avatar_url'] = this.avatarUrl;
     data['role'] = this.role;
     data['type'] = this.type;
     data['firebase_token'] = this.firebaseToken;
@@ -642,16 +158,14 @@ class User {
     data['company_id'] = this.companyId;
     data['employee_code'] = this.employeeCode;
     data['token'] = this.token;
-    data['lunch_time'] = this.lunchTime;
     data['nfc'] = this.nfc;
-    if(this.canNTCTracking!=null){
-      data['can_nfc_tracking'] = this.canNTCTracking?1:0;
-    }
-    if(this.sendScheduleNotification!=null){
-      data['send_schedule_notification'] = this.sendScheduleNotification?1:0;
-    }
+    data['can_nfc_tracking'] = this.canNTCTracking;
     data['created_at'] = this.createdAt;
     data['updated_at'] = this.updatedAt;
+    data['active'] = this.active;
+    if(this.breakSetting != null){
+      data['break'] = this.breakSetting!.toJson();
+    }
     return data;
   }
 
@@ -659,8 +173,8 @@ class User {
     return '$firstName $lastName';
   }
 
-  Punch getTodayPunch(){
-    return punches.lastWhere((p) =>isSameDatePunch(p.createdAt, DateTime.now().toString()),orElse: ()=>null);
+  Punch? getTodayPunch(){
+    return punches.lastWhereOrNull((p) =>(p.createdAt != null && isSameDatePunch(p.createdAt!, DateTime.now().toString())));
   }
 
   bool isSameDatePunch(String punchDate, String date){
@@ -674,31 +188,35 @@ class User {
   Map<String, List<Punch>> getPunchesGroupByDate(){
     Map<String, List<Punch>> punchGroup = {};
     punches.forEach((p) {
-      if(punchGroup[getDateString(p.createdAt)]==null){
-        punchGroup[getDateString(p.createdAt)] = [p];
-      }else{
-        punchGroup[getDateString(p.createdAt)].add(p);
+      if(p.createdAt != null){
+        if(punchGroup[getDateString(p.createdAt!)] == null){
+          punchGroup[getDateString(p.createdAt!)] = [p];
+        }else{
+          punchGroup[getDateString(p.createdAt!)]!.add(p);
+        }
       }
     });
     return punchGroup;
   }
 
   List<Punch> getPunchesOfDate(DateTime date){
-    return punches.where((p) => isSameDatePunch(p.createdAt, date.toString())).toList();
+    return punches.where((p) =>p.createdAt != null && isSameDatePunch(p.createdAt!, date.toString()) && p.isIn()).toList();
   }
 
   List<Punch> getPunchesOfWeek(DateTime startDate){
-    DateTime endDate = startDate.add(Duration(days: 6));
-    return punches.where((p) => (DateTime.parse(p.createdAt).isAfter(startDate) && DateTime.parse(p.createdAt).isBefore(endDate))).toList();
+    DateTime endDate = startDate.add(Duration(days: 7));
+    return punches.where((p) => (p.createdAt != null && DateTime.parse(p.createdAt!).isAfter(startDate) && DateTime.parse(p.createdAt!).isBefore(endDate))).toList();
   }
 
   Map<String, List<Punch>> getPunchesGroupOfWeek(DateTime startDate){
     Map<String, List<Punch>> punchGroup = {};
     getPunchesOfWeek(startDate).forEach((p) {
-      if(punchGroup[getDateString(p.createdAt)]==null){
-        punchGroup[getDateString(p.createdAt)] = [p];
-      }else{
-        punchGroup[getDateString(p.createdAt)].add(p);
+      if(p.punch == "In" && p.createdAt != null){
+        if(punchGroup[getDateString(p.createdAt!)]==null){
+          punchGroup[getDateString(p.createdAt!)] = [p];
+        }else{
+          punchGroup[getDateString(p.createdAt!)]!.add(p);
+        }
       }
     });
     return punchGroup;
@@ -706,45 +224,38 @@ class User {
 
   double getHoursOfDate(DateTime date){
     List<Punch> punchesOfDate = getPunchesOfDate(date);
-    if(punchesOfDate.length<2)return 0.0;
     double hours = 0.0;
-    for(int i=0; i<punchesOfDate.length; i++ ){
-      if(punchesOfDate[i].punch=="In"){
-        if(i+1<punchesOfDate.length && punchesOfDate[i+1].punch=="Out"){
-          hours += DateTime.parse(punchesOfDate[i+1].createdAt).difference(DateTime.parse(punchesOfDate[i].createdAt)).inMinutes/60;
-        }else if(i+2<punchesOfDate.length && punchesOfDate[i+2].punch=="Out"){
-          hours += DateTime.parse(punchesOfDate[i+2].createdAt).difference(DateTime.parse(punchesOfDate[i].createdAt)).inMinutes/60;
-        }
+    punchesOfDate.forEach((punchIn) {
+      Punch? punchOut = getPunchOut(punchIn);
+      if(punchOut != null && punchOut.createdAt != null){
+        hours += DateTime.parse(punchOut.createdAt!).difference(DateTime.parse(punchIn.createdAt!)).inMinutes/60;
       }
-    }
+    });
     return hours;
   }
 
-  double getLunchBreakTime(DateTime date){
-    List<Punch> punchesOfDate = getPunchesOfDate(date);
-    Punch lunch = punchesOfDate.firstWhere((p) => p.punch=="Lunch",orElse: ()=>null);
-    if(lunch==null)return 0;
-    return DateTime.parse(lunch.updatedAt).difference(DateTime.parse(lunch.createdAt)).inMinutes/60;
+  double getBreakTime(DateTime date){
+    double hours = 0;
+    breaks.where((b) => b.start != null && isSameDatePunch(b.start!, date.toString())).forEach((b) {
+      if(b.length != null){
+        hours += b.length!/60;
+      }
+    });
+    return hours;
   }
 
-  double getTotalHoursOfYear(){
-    Map<String, List<Punch>> punchGroup = getPunchesGroupByDate();
-    double totalHours = 0.0;
-    punchGroup.forEach((key, v) {
-      totalHours +=getHoursOfDate(DateTime.parse(key));
-    });
-    return totalHours;
+  double calculateHoursOfDate(DateTime date){
+    return getHoursOfDate(date) - getBreakTime(date);
   }
 
   double getTotalHoursOfWeek(DateTime startDate){
     DateTime endDate = startDate.add(Duration(days: 6));
     Map<String, List<Punch>> punchGroup = getPunchesGroupByDate();
     double totalHours = 0.0;
-    punchGroup.forEach((key, v) {
-      DateTime date = DateTime.parse(key);
+    punchGroup.forEach((d, v) {
+      DateTime date = DateTime.parse(d);
       if(date.isAfter(startDate) && date.isBefore(endDate)){
-        totalHours +=getHoursOfDate(date);
-        totalHours -=getLunchBreakTime(date);
+        totalHours += calculateHoursOfDate(date);
       }
     });
     return totalHours;
@@ -760,166 +271,386 @@ class User {
     return getTotalHoursOfWeek(startOfWeek.subtract(Duration(days: 7)));
   }
 
-  Future<String> editPunch(int punchId, String value)async{
+  Punch? getPunchOut(Punch punchIn){
+    return punches.firstWhereOrNull((p) => (p.inId != null && p.inId == punchIn.id));
+  }
+
+  List<WorkHistory> worksOfPunch(Punch punch){
+    return works.where((w) => (w.punchId != null && w.punchId == punch.id),).toList();
+  }
+
+  List<EmployeeBreak> breaksOfPunch(Punch punch){
+    return breaks.where((b) => (b.punchId!=null && b.punchId == punch.id),).toList();
+  }
+
+  Future<String?> editPunch(int? punchId, String value)async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
         AppConst.editPunch,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/json',
-          'Authorization':'Bearer '+GlobalData.token
-        },
-        body: jsonEncode({'punch_id':punchId,'value':value}),
+        GlobalData.token,
+        {'punch_id':punchId,'value':value},
       );
-      print("[User.editPunch] ${res.body}");
+      Tools.consoleLog("[User.editPunch.res] ${res.body}");
       if(res.statusCode==200){
         return null;
       }else{
         return jsonDecode(res.body)['message'];
       }
     }catch(e){
-      print("[User.editPunch] $e");
+      Tools.consoleLog("[User.editPunch.err] $e");
       return e.toString();
     }
   }
 
-  Future<String> deletePunch(int punchId)async{
+  Future<String?> deletePunch(int? punchId)async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
         AppConst.deletePunch,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/json',
-          'Authorization':'Bearer '+GlobalData.token
-        },
-        body: jsonEncode({'punch_id':punchId}),
+        GlobalData.token,
+        {'punch_id':punchId},
       );
-      print("[User.deletePunch] ${res.body}");
+      Tools.consoleLog("[User.deletePunch.res] ${res.body}");
       if(res.statusCode==200){
         return null;
       }else{
         return jsonDecode(res.body)['message'];
       }
     }catch(e){
-      print("[User.deletePunch] $e");
+      Tools.consoleLog("[User.deletePunch.err] $e");
       return e.toString();
     }
   }
 
-  Future<String> editWork(WorkHistory work)async{
+  Future<String?> editWork(WorkHistory work)async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
         AppConst.editWork,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/json',
-          'Authorization':'Bearer '+GlobalData.token
-        },
-        body: jsonEncode(work.toJson()),
+        GlobalData.token,
+        work.toJson(),
       );
-      print("[User.editWork] ${res.body}");
+      Tools.consoleLog("[User.editWork.res] ${res.body}");
       if(res.statusCode==200){
         return null;
       }else{
         return jsonDecode(res.body)['message'];
       }
     }catch(e){
-      print("[User.editWork] $e");
+      Tools.consoleLog("[User.editWork.err] $e");
       return e.toString();
     }
   }
 
-  Future<String> deleteWork(int workId)async{
+  Future<String?> deleteWork(int? workId)async{
     try{
-      var res = await http.post(
+      var res = await sendPostRequest(
         AppConst.deleteWork,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/json',
-          'Authorization':'Bearer '+GlobalData.token
-        },
-        body: jsonEncode({'id':workId}),
+        GlobalData.token,
+        {'id':workId},
       );
-      print("[User.deleteWork] ${res.body}");
+      Tools.consoleLog("[User.deleteWork.res] ${res.body}");
       if(res.statusCode==200){
         return null;
       }else{
         return jsonDecode(res.body)['message'];
       }
     }catch(e){
-      print("[User.deleteWork] $e");
+      Tools.consoleLog("[User.deleteWork.err] $e");
       return e.toString();
     }
   }
 
   bool isPunchIn(){
-    return lastPunch!=null && lastPunch.punch=='In';
+    return lastPunch != null && lastPunch?.punch=='In';
   }
 
-  bool isTracking(){
+  bool hasTracking(){
     return ['shop_tracking','call_shop_tracking'].contains(type);
   }
 
+  bool hasSchedule(){
+    return ['shop_daily','call_shop_daily'].contains(type);
+  }
 
-  String pdfUrl(DateTime startDate){
+  bool hasCall(){
+    return ['call', 'call_shop_daily', 'call_shop_tracking'].contains(type);
+  }
+
+  bool checkType(String v){
+    return type == v;
+  }
+
+  bool hasNTCTracking(){
+    return canNTCTracking??false;
+  }
+
+  bool hasCode(){
+    return employeeCode != null && employeeCode!.isNotEmpty;
+  }
+
+  String pdfUrl(DateTime? startDate){
     DateTime pdfDate = startDate??PunchDateUtils.getStartOfCurrentWeek(DateTime.now());
     final pdfLink = "$firstName $lastName (${pdfDate.toString().split(" ")[0]} ~ ${pdfDate.add(Duration(days: 6)).toString().split(" ")[0]}).pdf";
-    print(pdfLink);
-    return Uri.encodeFull('https://facepunch.app/punch-pdfs/$companyId/$pdfLink');
+    return Uri.encodeFull('${AppConst.domainURL}punch-pdfs/$companyId/$pdfLink');
   }
 
-  List<WorkHistory> worksOfPunch(Punch punchIn, Punch punchOut){
-    try{
-      final inTime = DateTime.parse(punchIn.createdAt);
-      final outTime = punchOut!=null?DateTime.parse(punchOut.createdAt):null;
-      return works.where((w){
-        final workTime = DateTime.parse(w.createdAt);
-        if(workTime.day != inTime.day || workTime.month !=inTime.month)return false;
-        if(workTime.isBefore(inTime))return false;
-        if(outTime!=null && workTime.isAfter(outTime))return false;
-        return true;
-      }).toList();
-    }catch(e){
-      print('[User.worksOfPunch] $e');
-    }
-    return [];
+  String harvestReportUrl(){
+    return Uri.encodeFull('${AppConst.domainURL}harvest-reports/$companyId/Harvest_Report_${DateTime.now().toString().split(' ')[0]}.png');
   }
 
-  Future<List<WorkSchedule>> getDailySchedule(String date)async{
-    List<WorkSchedule> schedules = [];
+  Widget userAvatarImage(){
+    return CachedNetworkImage(
+      imageUrl: '$avatarUrl',
+      alignment: Alignment.center,
+      placeholder: (_,__)=>Image.asset("assets/images/person.png"),
+      errorWidget: (_,__,___)=>Image.asset("assets/images/person.png"),
+      fit: BoxFit.cover,
+    );
+  }
+
+  Future<String?> getDailyTasks(String date)async{
     try{
-      var res = await http.post(
-          AppConst.getDailySchedule,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-            'Authorization':'Bearer '+token
-          },
-          body: {
-            'date':date
-          }
+      var res = await sendPostRequest(
+          AppConst.getDailyTasks,
+          token,
+          {'date':date}
       );
-      print('[WorkModel.getDailySchedule]${res.body}');
+      Tools.consoleLog('[WorkModel.getDailyTasks.res]${res.body}');
+      var body = jsonDecode(res.body);
       if(res.statusCode==200){
-        for(var json in jsonDecode(res.body))
-          schedules.add(WorkSchedule.fromJson(json));
+        schedules.clear();
+        calls.clear();
+        works.clear();
+        for(var s in body['schedules']){
+          schedules.add(WorkSchedule.fromJson(s));
+        }
+        for(var c in body['calls']){
+          calls.add(EmployeeCall.fromJson(c));
+        }
+        for(var w in body['works']){
+          works.add(WorkHistory.fromJson(w));
+        }
+        return null;
+      }else{
+        return body['message']??'Something went wrong.';
       }
     }catch(e){
-      print('[WorkModel.getDailySchedule]$e');
+      Tools.consoleLog('[WorkModel.getDailyTasks.err]$e');
+      return e.toString();
     }
-    return schedules;
+  }
+
+  Future<List<Revision>> getMyRevisionNotifications()async{
+    var res = await sendGetRequest(
+        AppConst.getRevisionNotifications,
+        token
+    );
+    Tools.consoleLog('[User.getMyRevisionNotifications.res]${res.body}');
+    final json = jsonDecode(res.body);
+    if(res.statusCode==200){
+      List<Revision> revisions = [];
+      for(var revision in json){
+        revisions.add(Revision.fromJson(revision));
+      }
+      return revisions;
+    }else{
+      throw handleError(json);
+    }
+  }
+
+  Future<List<Revision>> getTeamRevisionNotifications(String date, bool isWeek)async{
+    var res = await sendPostRequest(
+        AppConst.getTeamRevisionNotifications,
+        token,
+      {
+        'date': date,
+        'is_week': isWeek
+      }
+    );
+    Tools.consoleLog('[User.getTeamRevisionNotifications.res]${res.body}');
+    final json = jsonDecode(res.body);
+    if(res.statusCode==200){
+      List<Revision> revisions = [];
+      for(var revision in json){
+        revisions.add(Revision.fromJson(revision));
+      }
+      return revisions;
+    }else{
+      throw handleError(json);
+    }
+  }
+
+  Future<String?> getDailyCall(String date)async{
+    try{
+      var res = await sendPostRequest(
+          AppConst.getEmployeeCall,
+          GlobalData.token,
+          {
+            'date' : date,
+            'id' : id.toString()
+          }
+      );
+      Tools.consoleLog('[WorkModel.getDailyCall.res]${res.body}');
+      var body = jsonDecode(res.body);
+      if(res.statusCode == 200){
+        calls.clear();
+        for(var c in body){
+          calls.add(EmployeeCall.fromJson(c));
+        }
+        return null;
+      }else{
+        return body['message']??'Something went wrong.';
+      }
+    }catch(e){
+      Tools.consoleLog('[WorkModel.getDailyCall.err]$e');
+      return e.toString();
+    }
+  }
+
+  Future<String?> deleteBreak(int? breakId)async{
+    try{
+      var res = await sendPostRequest(
+        AppConst.deleteBreak,
+        GlobalData.token,
+        {'id':breakId},
+      );
+      Tools.consoleLog("[User.deleteBreak.res] ${res.body}");
+      if(res.statusCode==200){
+        return null;
+      }else{
+        return jsonDecode(res.body)['message'];
+      }
+    }catch(e){
+      Tools.consoleLog("[User.deleteBreak.err] $e");
+      return e.toString();
+    }
+  }
+
+  bool isAdmin(){
+    return role == 'admin';
+  }
+
+  bool isEmployee(){
+    return role == 'employee';
+  }
+
+  bool isManager(){
+    return role == 'manager';
+  }
+
+  bool isSubAdmin(){
+    return role == 'sub_admin';
+  }
+
+  bool canManageDispatch(){
+    return role != 'employee';
+  }
+
+  bool isManualBreak(){
+    return breakSetting != null && breakSetting!.type == "manual";
+  }
+
+  Future<String?> startShopTracking({int? projectId, int? taskId, double? latitude, double? longitude})async{
+    try{
+      var res = await sendPostRequest(
+          AppConst.startShopTracking,
+          token,
+          {
+            'task_id': taskId,
+            'project_id': projectId,
+            'latitude': latitude,
+            'longitude': longitude,
+          }
+      );
+      Tools.consoleLog('[User.startShopTracking.res]${res.body}');
+      if(res.statusCode==200){
+        return null;
+      }else{
+        return jsonDecode(res.body)['message'];
+      }
+    }catch(e){
+      Tools.consoleLog('[User.startShopTracking.err]$e');
+      return e.toString();
+    }
+  }
+
+  Future<String?> startManualBreak()async{
+    try{
+      var res = await sendGetRequest(AppConst.startManualBreak, token);
+      Tools.consoleLog('[User.startManualBreak.res]${res.body}');
+      if(res.statusCode==200){
+        return null;
+      }else{
+        return jsonDecode(res.body)['message'];
+      }
+    }catch(e){
+      Tools.consoleLog('[User.startManualBreak.err]$e');
+      return e.toString();
+    }
+  }
+
+  Future<String?> endManualBreak()async{
+    try{
+      var res = await sendGetRequest(AppConst.endManualBreak, token);
+      Tools.consoleLog('[User.endManualBreak.res]${res.body}');
+      if(res.statusCode==200){
+        return null;
+      }else{
+        return jsonDecode(res.body)['message'];
+      }
+    }catch(e){
+      Tools.consoleLog('[User.endManualBreak.err]$e');
+      return e.toString();
+    }
+  }
+
+  Future<String?> delete()async{
+    String result = 'Oops, Unknown Errors!';
+    try{
+      var res = await sendPostRequest(AppConst.deleteEmployee, GlobalData.token, {'id' : id});
+      Tools.consoleLog("[User.delete.res] ${res.body}");
+      if(res.statusCode==200){
+        return null;
+      }else{
+        return jsonDecode(res.body)['message'];
+      }
+    }catch(e){
+      Tools.consoleLog("[User.delete.err] $e");
+    }
+    return result;
+  }
+
+  Future<String?> punchOut({double? latitude, double? longitude})async{
+    try{
+      var res = await sendPostRequest(
+          AppConst.punchOut,
+          token,
+          {
+            'latitude':latitude,
+            'longitude':longitude
+          }
+      );
+      if(res.statusCode==200){
+        return null;
+      }else{
+        return jsonDecode(res.body)['message'];
+      }
+    }catch(e){
+      Tools.consoleLog("[User.punchOut.err] $e");
+      return e.toString();
+    }
   }
 }
 
 class Punch{
-  int id;
-  int userId;
-  String punch;
-  double longitude;
-  double latitude;
-  String createdAt;
-  String updatedAt;
-  int paid;
+  int? id;
+  int? userId;
+  int? inId;
+  String? punch;
+  double? longitude;
+  double? latitude;
+  String? status;
+  String? createdAt;
+  String? updatedAt;
 
   Punch({
     this.id,
@@ -927,7 +658,6 @@ class Punch{
     this.punch,
     this.longitude,
     this.latitude,
-    this.paid,
     this.createdAt,
     this.updatedAt
   });
@@ -936,27 +666,134 @@ class Punch{
     try{
       id = json['id'];
       userId = json['user_id'];
+      inId = json['in_id'];
       punch = json['punch'];
       longitude = json['longitude']==null?null:double.parse(json['longitude'].toString());
       latitude = json['latitude']==null?null:double.parse(json['latitude'].toString());
-      paid = json['paid'];
+      status = json['status'];
       createdAt = json['created_at'];
       updatedAt = json['updated_at'];
     }catch(e){
-      print("[Punch.fromJson] $e");
+      Tools.consoleLog("[Punch.fromJson.err] $e");
     }
+  }
+
+  bool isIn(){
+    return punch == "In";
+  }
+
+  bool isOut(){
+    return punch == "Out";
+  }
+
+  bool isSent(){
+    return status == 'Sent';
+  }
+
+  String title(BuildContext context){
+    if(createdAt == null){
+      return "${S.of(context).punch} $punch ${S.of(context).at} --:--";
+    }
+    return "${S.of(context).punch} $punch ${S.of(context).at} ${PunchDateUtils.getTimeString(DateTime.parse(createdAt!))}";
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = Map<String, dynamic>();
     data['id'] = this.id;
+    data['in_id'] = this.inId;
     data['user_id'] = this.userId;
     data['punch'] = this.punch;
     data['longitude'] = this.longitude;
     data['latitude'] = this.latitude;
-    data['paid'] = this.paid;
     data['created_at'] = this.createdAt;
     data['updated_at'] = this.updatedAt;
+    return data;
+  }
+
+  String getTime(){
+    if(createdAt == null) return '--:--';
+    return PunchDateUtils.getTimeString(DateTime.parse(createdAt!));
+  }
+
+  bool hasLocation(){
+    return latitude != null && longitude != null;
+  }
+
+}
+
+class FacePunchData{
+  late User employee;
+  Punch? punch;
+  List<EmployeeCall> calls = [];
+  List<WorkSchedule> schedules = [];
+  List<Project> projects = [];
+  List<ScheduleTask> tasks = [];
+  String? message;
+  WorkHistory? work;
+  double? latitude;
+  double? longitude;
+  bool isInManualBreak = false;
+
+  FacePunchData.fromJson(Map<String, dynamic> json){
+    try{
+      employee = User.fromJson(json['employee']);
+      message = json['message'];
+      if(json['punch'] != null){
+        punch = Punch.fromJson(json['punch']);
+      }
+      if(json['calls'] != null){
+        for(var call in json['calls']){
+          calls.add(EmployeeCall.fromJson(call));
+        }
+      }
+      if(json['schedules'] != null){
+        for(var schedule in json['schedules']){
+          schedules.add(WorkSchedule.fromJson(schedule));
+        }
+      }
+      if(json['projects'] != null){
+        for(var project in json['projects']){
+          projects.add(Project.fromJson(project));
+        }
+      }
+      if(json['tasks'] != null){
+        for(var task in json['tasks']){
+          tasks.add(ScheduleTask.fromJson(task));
+        }
+      }
+      if(json['work'] != null){
+        work = WorkHistory.fromJson(json['work']);
+      }
+      isInManualBreak = json['in_manual_break']??false;
+    }catch(e){
+      Tools.consoleLog("[FacePunchData.fromJson.err] $e");
+    }
+  }
+}
+
+class BreakSetting{
+  String? type;
+  String? start;
+  int? length;
+  bool? calculate;
+
+  BreakSetting.fromJson(Map<String, dynamic> json){
+    try{
+      type = json['type'];
+      start = json['start'];
+      length = json['length'];
+      calculate = json['calculate'];
+    }catch(e){
+      Tools.consoleLog("[BreakSetting.fromJson.err] $e");
+    }
+  }
+
+  Map<String, dynamic> toJson(){
+    final Map<String, dynamic> data = new Map<String, dynamic>();
+    data['type'] = type;
+    data['start'] = start;
+    data['length'] = length;
+    data['calculate'] = calculate;
     return data;
   }
 }
