@@ -1,102 +1,20 @@
 import 'dart:convert';
-
 import 'package:flutter/material.dart';
-import 'package:http/http.dart' as http;
 
-import 'app_const.dart';
+import '/lang/l10n.dart';
+import '/widgets/calendar_strip/date-utils.dart';
+import '/config/app_const.dart';
+import '/widgets/utils.dart';
+import '/providers/base_provider.dart';
 
-class WorkModel with ChangeNotifier{
-  List<Project> projects = [];
-  List<ScheduleTask> tasks = [];
-
-  Future<void> getProjectsAndTasks()async{
-    try{
-      var res = await http.get(
-        AppConst.getProjectsAndTasks,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/x-www-form-urlencoded',
-          'Authorization':'Bearer '+GlobalData.token
-        },
-      );
-      print('[WorkModel.getProjects]${res.body}');
-      if(res.statusCode==200){
-        projects.clear();
-        for(var project in jsonDecode(res.body)['projects']){
-          projects.add(Project.fromJson(project));
-        }
-        tasks.clear();
-        for(var project in jsonDecode(res.body)['tasks']){
-          tasks.add(ScheduleTask.fromJson(project));
-        }
-        notifyListeners();
-      }
-    }catch(e){
-      print('[WorkModel.getProjects]$e');
-    }
-  }
-
-  Future<String> submitRevision({WorkSchedule newSchedule, WorkSchedule oldSchedule})async{
-    try{
-      var res = await http.post(
-        AppConst.sendTimeRevisionRequest,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/json',
-          'Authorization':'Bearer '+GlobalData.token
-        },
-        body: jsonEncode({
-          'schedule_id':oldSchedule.id,
-          'new_value':newSchedule.toJson(),
-          'old_value':oldSchedule.toJson()
-        }),
-      );
-      print('[WorkSchedule.submitRevision]${res.body}');
-      if(res.statusCode==200){
-        return null;
-      }else{
-        return jsonDecode(res.body)['message'];
-      }
-    }catch(e){
-      print('[WorkModel.submitRevision]$e');
-      return e.toString();
-    }
-  }
-
-  Future<List<WorkSchedule>> getEmployeeSchedule({String date, int userId})async{
-    List<WorkSchedule> schedules = [];
-    try{
-      var res = await http.post(
-          AppConst.getEmployeeSchedule,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-            'Authorization':'Bearer '+GlobalData.token
-          },
-          body: {
-            'date':date,
-            'id':userId.toString()
-          }
-      );
-      print('[WorkModel.getEmployeeSchedule]${res.body}');
-      if(res.statusCode==200){
-        for(var json in jsonDecode(res.body))
-          schedules.add(WorkSchedule.fromJson(json));
-      }
-    }catch(e){
-      print('[WorkModel.getEmployeeSchedule]$e');
-    }
-    return schedules;
-  }
-}
 
 class Project{
-  int id;
-  int companyId;
-  String name;
-  String companyName;
-  String code;
-
+  int? id;
+  int? companyId;
+  String? name;
+  String? companyName;
+  String? code;
+  String? address;
 
   Project.fromJson(Map<String, dynamic> json){
     try{
@@ -105,18 +23,23 @@ class Project{
       companyName = json['company_name'];
       name = json['name'];
       code = json['code'];
+      address = json['address'];
     }catch(e){
-      print('[Project.fromJson]$e');
+      Tools.consoleLog('[Project.fromJson.err]$e');
     }
+  }
+
+  bool hasCode(){
+    return code != null && code!.isNotEmpty;
   }
 
 }
 
 class ScheduleTask{
-  int id;
-  String name;
-  String code;
-  int companyId;
+  int? id;
+  String? name;
+  String? code;
+  int? companyId;
 
   ScheduleTask.fromJson(Map<String, dynamic> json){
     try{
@@ -125,23 +48,33 @@ class ScheduleTask{
       name = json['name'];
       code = json['code'];
     }catch(e){
-      print('[ScheduleTask.fromJson]$e');
+      Tools.consoleLog('[ScheduleTask.fromJson.err]$e');
     }
+  }
+
+  bool hasCode(){
+    return code != null && code!.isNotEmpty;
   }
 }
 
 class WorkHistory{
-  int id;
-  int userId;
-  int taskId;
-  int projectId;
-  String start;
-  String end;
-  String createdAt;
-  String updatedAt;
-  String taskName;
-  String projectName;
-
+  int? id;
+  int? userId;
+  int? punchId;
+  int? taskId;
+  int? projectId;
+  String? projectAddress;
+  int? callId;
+  int? scheduleId;
+  String? start;
+  String? end;
+  String? taskName;
+  String? taskCode;
+  String? projectName;
+  String? projectCode;
+  String? type;
+  String? createdAt;
+  String? updatedAt;
   WorkHistory({
     this.id,
     this.userId,
@@ -150,80 +83,135 @@ class WorkHistory{
     this.start,
     this.end,
     this.taskName,
-    this.projectName
+    this.projectName,
+    this.taskCode,
+    this.projectCode
   });
 
   WorkHistory.fromJson(Map<String, dynamic> json){
     try{
       id = json['id'];
       userId = json['user_id'];
+      punchId = json['punch_id'];
       projectId = json['project_id'];
       taskId = json['task_id'];
+      callId = json['call_id'];
+      scheduleId = json['schedule_id'];
       taskName = json['task_name'];
+      taskCode = json['task_code'];
       projectName = json['project_name'];
+      projectCode = json['project_code'];
+      projectAddress = json['project_address'];
       start = json['start'];
       end = json['end'];
+      type = json['type'];
       createdAt = json['created_at'];
       updatedAt = json['updated_at'];
     }catch(e){
-      print('[WorkHistory.fromJson]$e');
+      Tools.consoleLog('[WorkHistory.fromJson.err]$e');
     }
   }
 
   double workHour(){
     try{
-      if(end==null || end.isEmpty)return 0.0;
-      final workDate = createdAt.split(' ')[0];
-      final startTime = DateTime.parse('$workDate $start');
-      final endTime = DateTime.parse('$workDate $end');
+      final startTime = DateTime.tryParse(start!);
+      final endTime = DateTime.tryParse(end!);
+      if(startTime == null)return 0.0;
+      if(endTime == null)return 0.0;
       return endTime.difference(startTime).inMinutes/60;
     }catch(e){
-      print('[WorkHistory.workHour]$e');
+      Tools.consoleLog('[WorkHistory.workHour.err]$e');
     }
     return 0.0;
   }
 
-  DateTime getStartTime(){
-    final workDate = createdAt.split(' ')[0];
-    return DateTime.parse('$workDate $start');
+  DateTime? getStartTime(){
+    if(start != null){
+      return DateTime.tryParse(start!);
+    }
   }
 
-  DateTime getEndTime(){
-    final workDate = createdAt.split(' ')[0];
-    return DateTime.parse('$workDate $end');
+  DateTime? getEndTime(){
+    if(end==null || end!.isEmpty) return null;
+    return DateTime.tryParse(end!);
+  }
+
+  String title(){
+    return '${projectName??''} - ${taskName??''}';
+  }
+
+  String projectTitle(){
+    return '$projectName - $projectCode \n $projectAddress'.trim();
+  }
+
+  String taskTitle(){
+    return '$taskName - $taskCode';
   }
 
   toJson(){
     return {
       'id':id,
       'user_id':userId,
+      'punch_id':punchId,
       'start':start,
       'end':end,
       'task_id' : taskId,
       'project_id' : projectId,
+      'call_id' : callId,
+      'schedule_id' : scheduleId,
       'task_name':taskName,
+      'task_code':taskCode,
       'project_name':projectName,
+      'project_code':projectCode,
+      'project_address':projectAddress,
+      'type':type,
       'created_at':createdAt,
       'updated_at':updatedAt
     };
   }
+
+  bool isEnd(){
+    return end != null && end!.isNotEmpty;
+  }
+
+  bool isWorkingOn(){
+    return start != null && end == null;
+  }
+
+  String startTime(){
+    if(start != null && start!.length > 16){
+      return start!.substring(11, 16);
+    }
+    return '--:--';
+  }
+
+  String endTime(){
+    if(end != null && end!.length > 16){
+      return end!.substring(11, 16);
+    }
+    return '--:--';
+  }
+
 }
 
-class WorkSchedule{
-  int id;
-  int userId;
-  int projectId;
-  int taskId;
-  String projectName;
-  String taskName;
-  String start;
-  String end;
-  String type;
-  String todo;
-  String note;
-  int priority;
-  String createdAt;
-  String updatedAt;
+class WorkSchedule with HttpRequest{
+  int? id;
+  int? userId;
+  int? projectId;
+  int? taskId;
+  String? projectName;
+  String? projectCode;
+  String? projectAddress;
+  String? taskName;
+  String? taskCode;
+  String? start;
+  String? end;
+  Shift? shift;
+  String? color;
+  String? noAvailable;
+  String? status;
+  String? createdAt;
+  String? updatedAt;
 
   WorkSchedule({
     this.id,
@@ -231,13 +219,15 @@ class WorkSchedule{
     this.projectId,
     this.taskId,
     this.projectName,
+    this.projectCode,
+    this.projectAddress,
     this.taskName,
+    this.taskCode,
     this.start,
     this.end,
-    this.type,
-    this.todo,
-    this.note,
-    this.priority,
+    this.shift,
+    this.color,
+    this.noAvailable,
     this.createdAt,
     this.updatedAt
   });
@@ -249,17 +239,22 @@ class WorkSchedule{
       projectId = json['project_id'];
       taskId = json['task_id'];
       projectName = json['project_name'];
+      projectCode = json['project_code'];
+      projectAddress = json['project_address'];
       taskName = json['task_name'];
+      taskCode = json['task_code'];
       start = json['start'];
       end = json['end'];
-      type = json['type'];
-      todo = json['todo'];
-      note = json['note'];
-      priority = json['priority'];
+      if(json['shift'] != null){
+        shift = Shift.fromJson(json['shift']);
+      }
+      noAvailable = json['no_available'];
+      color = json['color'];
+      status = json['status'];
       createdAt = json['created_at'];
       updatedAt = json['updated_at'];
     }catch(e){
-      print('[WorkSchedule.fromJson]$e');
+      Tools.consoleLog('[WorkSchedule.fromJson.err]$e');
     }
   }
 
@@ -270,145 +265,462 @@ class WorkSchedule{
       'project_id':projectId,
       'task_id':taskId,
       'project_name':projectName,
+      'project_code':projectCode,
+      'project_address':projectAddress,
       'task_name':taskName,
+      'task_code':taskCode,
       'start':start,
       'end':end,
-      'type':type,
+      'shift':shift,
+      'no_available':noAvailable,
+      'color': color,
+      'status' : status,
+      'created_at':createdAt,
+      'updated_at':updatedAt
+    };
+  }
+
+  bool isWorked(){
+    return status == 'done';
+  }
+
+  bool isWorkingOn(){
+    return status == "working";
+  }
+
+  DateTime? getStartTime(){
+    return DateTime.tryParse(start!);
+  }
+
+  DateTime? getEndTime(){
+    return DateTime.tryParse(end!);
+  }
+
+  String startTime(){
+    if(start != null && start!.length > 16){
+      return start!.substring(11, 16);
+    }
+    return '--:--';
+  }
+
+  String endTime(){
+    if(end != null && end!.length > 16){
+      return end!.substring(11, 16);
+    }
+    return '--:--';
+  }
+
+  String projectTitle(){
+    if(isNoAvailable()) return '';
+    return '$projectName - ${projectCode??''} \n $projectAddress'.trim();
+  }
+
+  String taskTitle(){
+    if(isNoAvailable()) return '$taskName';
+    return '$taskName - ${taskCode??''}';
+  }
+
+  bool isNoAvailable(){
+    return noAvailable != null && noAvailable!.isNotEmpty;
+  }
+
+  String? isValid(){
+    final startTime = getStartTime();
+    if(startTime != null){
+      int startDiff = startTime.difference(DateTime.now()).inMinutes;
+      if(startDiff > 5){
+        return 'Why do you start this schedule early?';
+      }else if(startDiff < -5){
+        return 'Why do you start this schedule late?';
+      }
+    }
+    final endTime = getEndTime();
+    if(endTime != null){
+      int endDiff = endTime.difference(DateTime.now()).inMinutes;
+      if(endDiff > 5){
+        return 'Why do you finish this schedule early?';
+      }else if(endDiff < -5){
+        return 'Why do you finish this schedule late?';
+      }
+    }
+  }
+
+  Future<String?> startSchedule({String? token, double? latitude, double? longitude})async{
+    try{
+      var res = await sendPostRequest(
+        AppConst.startSchedule,
+        token??GlobalData.token,
+        {
+          'id' : id,
+          'latitude' : latitude,
+          'longitude' : longitude
+        }
+      );
+      Tools.consoleLog('[WorkSchedule.startSchedule.res]${res.body}');
+      if(res.statusCode==200){
+        return null;
+      }else{
+        return jsonDecode(res.body)['message'];
+      }
+    }catch(e){
+      Tools.consoleLog('[WorkModel.startSchedule.err]$e');
+      return e.toString();
+    }
+  }
+
+  Future<String?> endSchedule()async{
+    try{
+      var res = await sendPostRequest(
+        AppConst.endSchedule,
+        GlobalData.token,
+        {'id':id}
+      );
+      Tools.consoleLog('[WorkSchedule.endSchedule.res]${res.body}');
+      if(res.statusCode==200){
+        return null;
+      }else{
+        return jsonDecode(res.body)['message'];
+      }
+    }catch(e){
+      Tools.consoleLog('[WorkModel.endSchedule.err]$e');
+      return e.toString();
+    }
+  }
+
+  Future<String?> deleteSchedule()async{
+    try{
+      var res = await sendPostRequest(
+          AppConst.deleteSchedule,
+          GlobalData.token,
+          { 'id':id }
+      );
+      Tools.consoleLog('[WorkSchedule.deleteSchedule.res]${res.body}');
+      if(res.statusCode==200){
+        return null;
+      }else{
+        return jsonDecode(res.body)['message'];
+      }
+    }catch(e){
+      Tools.consoleLog('[WorkModel.deleteSchedule.err]$e');
+      return e.toString();
+    }
+  }
+
+  Future<String?> editSchedule()async{
+    try{
+      var res = await sendPostRequest(
+          AppConst.editSchedule,
+          GlobalData.token,
+          toJson(),
+      );
+      Tools.consoleLog('[WorkSchedule.editSchedule.res]${res.body}');
+      if(res.statusCode==200){
+        return null;
+      }else{
+        return jsonDecode(res.body)['message'];
+      }
+    }catch(e){
+      Tools.consoleLog('[WorkModel.editSchedule.err]$e');
+      return e.toString();
+    }
+  }
+
+  Future<String?> addSchedule()async{
+    try{
+      var res = await sendPostRequest(
+          AppConst.addSchedule,
+          GlobalData.token,
+          toJson()
+      );
+      Tools.consoleLog('[WorkSchedule.addSchedule.res]${res.body}');
+      if(res.statusCode==200){
+        return null;
+      }else{
+        return jsonDecode(res.body)['message'];
+      }
+    }catch(e){
+      Tools.consoleLog('[WorkModel.addSchedule.err]$e');
+      return e.toString();
+    }
+  }
+}
+
+class EmployeeCall with HttpRequest{
+  int? id;
+  int? userId;
+  int? projectId;
+  int? taskId;
+  String? projectName;
+  String? projectCode;
+  String? projectAddress;
+  String? taskName;
+  String? taskCode;
+  String? date;
+  String? start;
+  String? end;
+  String? todo;
+  String? note;
+  int? priority;
+  String? status;
+  String? createdAt;
+  String? updatedAt;
+
+  EmployeeCall({
+    this.id,
+    this.userId,
+    this.projectId,
+    this.taskId,
+    this.projectName,
+    this.projectCode,
+    this.projectAddress,
+    this.taskCode,
+    this.taskName,
+    this.date,
+    this.start,
+    this.end,
+    this.todo,
+    this.note,
+    this.priority,
+    this.createdAt,
+    this.updatedAt
+  });
+
+  EmployeeCall.fromJson(Map<String, dynamic> json){
+    try{
+      id = json['id'];
+      userId = json['user_id'];
+      projectId = json['project_id'];
+      taskId = json['task_id'];
+      projectName = json['project_name'];
+      projectCode = json['project_code'];
+      projectAddress = json['project_address'];
+      taskName = json['task_name'];
+      taskCode = json['task_code'];
+      date = json['date'];
+      start = json['start'];
+      end = json['end'];
+      todo = json['todo']??'';
+      note = json['note']??'';
+      priority = json['priority'];
+      status = json['status'];
+      createdAt = json['created_at'];
+      updatedAt = json['updated_at'];
+    }catch(e){
+      Tools.consoleLog('[EmployeeCall.fromJson.err]$e');
+    }
+  }
+
+  Map<String, dynamic> toJson(){
+    return {
+      'id':id,
+      'user_id':userId,
+      'project_id':projectId,
+      'task_id':taskId,
+      'project_name':projectName,
+      'project_code':projectCode,
+      'project_address':projectAddress,
+      'task_name':taskName,
+      'task_code':taskCode,
+      'date':date,
+      'start':start,
+      'end':end,
       'todo':todo,
       'note':note,
+      'status':status,
       'priority':priority,
       'created_at':createdAt,
       'updated_at':updatedAt
     };
   }
 
-  bool isStarted(){
-    return type=='call' && (start !=null && start.isNotEmpty) && (end==null || end.isEmpty);
+  bool isWorked(){
+    return status == 'done';
   }
 
-  bool isEnded(){
-    return type=='call' && (start !=null && start.isNotEmpty) && (end !=null && end.isNotEmpty);
+  bool isWorkingOn(){
+    return status == 'working';
   }
 
-  DateTime getStartTime(){
-    final workDate = createdAt.split(' ')[0];
-    return DateTime.parse('$workDate $start');
-  }
-  DateTime getEndTime(){
-    final workDate = createdAt.split(' ')[0];
-    return DateTime.parse('$workDate $end');
+  DateTime? getStartTime(){
+    if(start == null) return null;
+    return DateTime.tryParse(start!);
   }
 
-  Future<String> startSchedule()async{
+  DateTime? getEndTime(){
+    if(end == null) return null;
+    return DateTime.tryParse(end!);
+  }
+
+  String startTime(){
+    if(start != null && start!.length > 16){
+      return start!.substring(11, 16);
+    }
+    return '--:--';
+  }
+
+  String endTime(){
+    if(end != null && end!.length > 16){
+      return end!.substring(11, 16);
+    }
+    return '--:--';
+  }
+
+  Future<String?> startCall({String? token, double? latitude, double? longitude})async{
     try{
-      var res = await http.post(
-        AppConst.startSchedule,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/x-www-form-urlencoded',
-          'Authorization':'Bearer '+GlobalData.token
-        },
-        body: {'id':id.toString()}
+      var res = await sendPostRequest(
+          AppConst.startCall,
+          token,
+          {
+            'id':id,
+            'latitude':latitude,
+            'longitude':longitude
+          }
       );
-      print('[WorkSchedule.startSchedule]${res.body}');
+      Tools.consoleLog('[EmployeeCall.startCall.res]${res.body}');
       if(res.statusCode==200){
         return null;
       }else{
         return jsonDecode(res.body)['message'];
       }
     }catch(e){
-      print('[WorkModel.startSchedule]$e');
+      Tools.consoleLog('[EmployeeCall.startSchedule.err]$e');
       return e.toString();
     }
   }
 
-  Future<String> endSchedule()async{
+  Future<String?> addEditCall()async{
     try{
-      var res = await http.post(
-        AppConst.endSchedule,
-        headers: {
-          'Accept':'application/json',
-          'Content-Type':'application/x-www-form-urlencoded',
-          'Authorization':'Bearer '+GlobalData.token
-        },
-        body: {'id':id.toString()}
+      var res = await sendPostRequest(
+        AppConst.addEditCall,
+        GlobalData.token,
+        toJson(),
       );
-      print('[WorkSchedule.endSchedule]${res.body}');
+      Tools.consoleLog('[EmployeeCall.addEditCall.res]${res.body}');
       if(res.statusCode==200){
         return null;
       }else{
         return jsonDecode(res.body)['message'];
       }
     }catch(e){
-      print('[WorkModel.endSchedule]$e');
+      Tools.consoleLog('[EmployeeCall.addEditCall.err]$e');
       return e.toString();
     }
   }
 
-  Future<String> deleteSchedule()async{
+  Future<String?> delete()async{
     try{
-      var res = await http.post(
-          AppConst.deleteSchedule,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/x-www-form-urlencoded',
-            'Authorization':'Bearer '+GlobalData.token
-          },
-          body: {'id':id.toString()}
+      var res = await sendPostRequest(
+          AppConst.deleteCall,
+          GlobalData.token,
+          {'id':id}
       );
-      print('[WorkSchedule.deleteSchedule]${res.body}');
+      Tools.consoleLog('[EmployeeCall.delete.res]${res.body}');
       if(res.statusCode==200){
         return null;
       }else{
         return jsonDecode(res.body)['message'];
       }
     }catch(e){
-      print('[WorkModel.deleteSchedule]$e');
+      Tools.consoleLog('[EmployeeCall.startSchedule.err]$e');
       return e.toString();
     }
   }
-  Future<String> editSchedule()async{
+
+  Color color(){
+    List<Color> colors = [
+      Colors.blue,
+      Colors.green,
+      Colors.teal,
+      Colors.amber,
+      Colors.brown,
+      Colors.cyan,
+      Colors.indigo,
+      Colors.orange,
+      Colors.purple,
+      Colors.teal,
+    ];
+    if(id==null) return Colors.blue;
+    return colors[id!%10];
+  }
+
+  String projectTitle(){
+    return '$projectName - ${projectCode??''} \n $projectAddress'.trim();
+  }
+
+  String taskTitle(){
+    return '$taskName - $taskCode';
+  }
+
+  bool hasTime(){
+    return start != null && end != null;
+  }
+}
+
+class EmployeeBreak{
+  int? id;
+  int? userId;
+  int? punchId;
+  String? title;
+  String? start;
+  int? length;
+  bool? calculate;
+  String? createdAt;
+  String? updatedAt;
+
+  EmployeeBreak.fromJson(Map<String, dynamic> json){
     try{
-      var res = await http.post(
-          AppConst.editSchedule,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/json',
-            'Authorization':'Bearer '+GlobalData.token
-          },
-          body: jsonEncode(toJson()),
-      );
-      print('[WorkSchedule.editSchedule]${res.body}');
-      if(res.statusCode==200){
-        return null;
-      }else{
-        return jsonDecode(res.body)['message'];
-      }
+      id = json['id'];
+      userId = json['user_id'];
+      punchId = json['punch_id'];
+      title = json['title'];
+      start = json['start'];
+      length = json['length'];
+      calculate = json['calculate']==1;
+      createdAt = json['created_at'];
+      updatedAt = json['updated_at'];
     }catch(e){
-      print('[WorkModel.editSchedule]$e');
-      return e.toString();
+      Tools.consoleLog('[EmployeeBreak.fromJson.err]$e');
     }
   }
-  Future<String> addSchedule()async{
+
+  String getTitle(BuildContext context){
+    if(start == null){
+      return '$title ${S.of(context).at} --:--, $length Minutes';
+    }
+    return '$title ${S.of(context).at} ${PunchDateUtils.getTimeString(DateTime.tryParse(start!))}, $length Minutes';
+  }
+
+  Map<String, dynamic> toJson(){
+    return {
+      'id':id,
+      'user_id':userId,
+      'punch_id':punchId,
+      'start':start,
+      'title':title,
+      'length':length,
+      'calculate':(calculate!= null && calculate!)?1:0,
+      'created_at':createdAt,
+      'updated_at':updatedAt
+    };
+  }
+}
+
+class Shift{
+  int? id;
+  String? name;
+  String? start;
+  String? end;
+  String? color;
+
+  Shift.fromJson(Map<String, dynamic> json){
     try{
-      var res = await http.post(
-          AppConst.addSchedule,
-          headers: {
-            'Accept':'application/json',
-            'Content-Type':'application/json',
-            'Authorization':'Bearer '+GlobalData.token
-          },
-          body: jsonEncode(toJson()),
-      );
-      print('[WorkSchedule.addSchedule]${res.body}');
-      if(res.statusCode==200){
-        return null;
-      }else{
-        return jsonDecode(res.body)['message'];
-      }
+      id = json['id'];
+      name = json['name'];
+      start = json['start'];
+      end = json['end'];
+      color = json['color'];
     }catch(e){
-      print('[WorkModel.addSchedule]$e');
-      return e.toString();
+      Tools.consoleLog('[Shift.fromJson.err]$e');
     }
   }
 }
