@@ -1,29 +1,32 @@
 import 'dart:convert';
 import 'dart:io';
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:facepunch/lang/l10n.dart';
-import 'package:facepunch/widgets/calendar_strip/date-utils.dart';
-import 'package:image_cropper/image_cropper.dart';
-import '../../models/company_model.dart';
-import '../../widgets/address_picker/country_state_city_picker.dart';
-import '../../models/user_model.dart';
-import '../../models/app_const.dart';
+import 'package:nfc_manager/nfc_manager.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:provider/provider.dart';
-import 'package:flutter_nfc_reader/flutter_nfc_reader.dart';
+import 'package:image_cropper/image_cropper.dart';
+import 'package:loader_overlay/loader_overlay.dart';
+
+import '/providers/company_provider.dart';
+import '/widgets/utils.dart';
+import '/lang/l10n.dart';
+import '/widgets/calendar_strip/date-utils.dart';
+import '/widgets/address_picker/country_state_city_picker.dart';
+import '/models/user_model.dart';
+import '/config/app_const.dart';
 
 class CreateEditEmployee extends StatefulWidget {
-  final User employee;
-  CreateEditEmployee({this.employee});
+  final User? employee;
+  final PageController? pageController;
+  CreateEditEmployee({this.employee, this.pageController});
   @override
   _CreateEditEmployeeState createState() => _CreateEditEmployeeState();
 }
 
 class _CreateEditEmployeeState extends State<CreateEditEmployee> {
   final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
-  final _picker = ImagePicker();
-  File _photoFile;
+  File? _photoFile;
   TextEditingController _fName = TextEditingController();
   TextEditingController _lName = TextEditingController();
   TextEditingController _email = TextEditingController();
@@ -33,55 +36,49 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
   TextEditingController _postal = TextEditingController();
   TextEditingController _phone = TextEditingController();
   TextEditingController _employeeCode = TextEditingController();
-  TextEditingController _function = TextEditingController();
   TextEditingController _salary = TextEditingController();
   TextEditingController _nfc = TextEditingController();
 
 
-  String _fNameError,_lNameError,_emailError,_passwordError,_addressError,
-      _postalError, _codeError,_functionError, _salaryError, _startDateError,_birthDayError;
-  String country,state,city, language;
-  int lunchTime = 30;
-  DateTime _startDate;
-  DateTime _birthDay;
-  bool isLoading = false;
-
-
+  String? _fNameError,_lNameError,_emailError,_passwordError,_addressError,
+      _postalError, _codeError, _salaryError, _startDateError,_birthDayError;
+  String? country,state,city, language;
+  DateTime? _startDate;
+  DateTime? _birthDay;
 
   @override
   void initState() {
     super.initState();
-    if(widget.employee!=null){
+    if(widget.employee != null){
       try{
-        _fName = TextEditingController(text: widget.employee.firstName);
-        _lName = TextEditingController(text: widget.employee.lastName);
-        _email = TextEditingController(text: widget.employee.email);
-        _password = TextEditingController(text: widget.employee.pin);
-        _address1 = TextEditingController(text: widget.employee.address1);
-        _address2 = TextEditingController(text: widget.employee.address2);
-        _postal = TextEditingController(text: widget.employee.postalCode);
-        _phone = TextEditingController(text: widget.employee.phone);
-        _employeeCode = TextEditingController(text: widget.employee.employeeCode);
-        _function = TextEditingController(text: widget.employee.function);
-        _salary = TextEditingController(text: widget.employee.salary);
-        _nfc = TextEditingController(text: widget.employee.nfc);
-        country = widget.employee.country;
-        state = widget.employee.state;
-        city = widget.employee.city;
-        language = widget.employee.language;
-        lunchTime = widget.employee.lunchTime;
-        if(widget.employee.start!=null)_startDate = DateTime.parse(widget.employee.start);
-        if(widget.employee.birthday!=null)_birthDay = DateTime.parse(widget.employee.birthday);
+        _fName = TextEditingController(text: widget.employee?.firstName);
+        _lName = TextEditingController(text: widget.employee?.lastName);
+        _email = TextEditingController(text: widget.employee?.email);
+
+        _password = TextEditingController(text: widget.employee?.pin);
+        _address1 = TextEditingController(text: widget.employee?.address1);
+        _address2 = TextEditingController(text: widget.employee?.address2);
+        _postal = TextEditingController(text: widget.employee?.postalCode);
+        _phone = TextEditingController(text: widget.employee?.phone);
+        _employeeCode = TextEditingController(text: widget.employee?.employeeCode);
+        _salary = TextEditingController(text: widget.employee?.salary);
+        _nfc = TextEditingController(text: widget.employee?.nfc);
+        country = widget.employee?.country;
+        state = widget.employee?.state;
+        city = widget.employee?.city;
+        language = widget.employee?.language;
+        if(widget.employee?.start != null)_startDate = DateTime.tryParse(widget.employee!.start!);
+        if(widget.employee?.birthday != null)_birthDay = DateTime.tryParse(widget.employee!.birthday!);
       }catch(e){
-       print("[CreateEditEmployee.initState] $e");
+       Tools.consoleLog("[CreateEditEmployee.initState] $e");
       }
     }
   }
 
   _pickUserPhoto(ImageSource source)async{
-    PickedFile image = await _picker.getImage(source: source,maxHeight: 800,maxWidth: 800);
-    if(image!=null){
-      File _cropFile = await ImageCropper.cropImage(
+    XFile? image = await ImagePicker().pickImage(source: source, maxHeight: 800, maxWidth: 800);
+    if(image != null){
+      File? _cropFile = await ImageCropper.cropImage(
           sourcePath: image.path,
           aspectRatioPresets: [
             CropAspectRatioPreset.square,
@@ -103,20 +100,20 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
           )
       );
       if(_cropFile!=null){
-        setState(() {_photoFile = _cropFile;});
+        if(mounted)setState(() {_photoFile = _cropFile;});
         return null;
       }
     }
-    setState(() {_photoFile = null;});
+    if(mounted)setState(() {_photoFile = null;});
   }
 
   Widget getAvatarWidget(){
     Widget image;
     if(_photoFile!=null){
-      image = Image.file(_photoFile,width: 120,height: 120,fit: BoxFit.cover,);
-    }else if(widget.employee!=null){
+      image = Image.file(_photoFile!, width: 120,height: 120,fit: BoxFit.cover,);
+    }else if(widget.employee != null){
       image = CachedNetworkImage(
-        imageUrl: "${AppConst.domainURL}images/user_avatars/${widget.employee.avatar}",
+        imageUrl: '${widget.employee!.avatarUrl}',
         height: 120,
         width: 120,
         alignment: Alignment.center,
@@ -126,13 +123,14 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
       );
     }else{
       image = Container(
-        color: Colors.green,
+        color: Color(primaryColor),
           child: Padding(
             padding: const EdgeInsets.all(10.0),
             child: Image.asset("assets/images/person.png",width: 100,height: 100,),
           )
       );
     }
+
     return ClipOval(
       child: image,
     );
@@ -146,42 +144,30 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
   }
 
   _selectStartDate(BuildContext context) async {
-    final DateTime picked = await showDatePicker(
+    final DateTime? picked = await showDatePicker(
         context: context,
-        initialDate: _startDate==null?DateTime.now():_startDate,
+        initialDate: _startDate == null? DateTime.now() : _startDate!,
         initialDatePickerMode: DatePickerMode.day,
         firstDate: DateTime(1970),
         lastDate: DateTime(2101));
-    if (picked != null)
+    if (picked != null && mounted)
       setState(() {_startDate = picked;});
   }
 
   _selectBirthDate(BuildContext context) async {
-    final DateTime picked = await showDatePicker(
+    final DateTime? picked = await showDatePicker(
         context: context,
-        initialDate: _birthDay==null?DateTime.now():_birthDay,
+        initialDate: _birthDay == null ? DateTime.now() : _birthDay!,
         initialDatePickerMode: DatePickerMode.day,
         firstDate: DateTime(1970),
         lastDate: DateTime(2101));
-    if (picked != null)
+    if (picked != null && mounted)
       setState(() {_birthDay = picked;});
-  }
-
-  showMessage(String message){
-    _scaffoldKey.currentState.hideCurrentSnackBar();
-    _scaffoldKey.currentState.showSnackBar(
-        SnackBar(
-          content: Text(message),
-          backgroundColor: message.toLowerCase().contains("success")?Colors.green:Colors.red,
-          duration: Duration(seconds: 2),
-          action: SnackBarAction(onPressed: (){},label: S.of(context).close,textColor: Colors.white,),
-        )
-    );
   }
 
   bool validator(){
     _fNameError=null;_lNameError=null;_emailError=null;_passwordError=null;_addressError=null;
-    _postalError=null;_codeError=null;_functionError=null;_salaryError=null;_startDateError=null;_birthDayError=null;
+    _postalError=null;_codeError=null;_salaryError=null;_startDateError=null;_birthDayError=null;
     if(_fName.text.isEmpty){
       _fNameError = S.of(context).firstNameIsRequired;
       return false;
@@ -198,18 +184,23 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
       _passwordError = S.of(context).passwordIsRequired;
       return false;
     }
+    if(_password.text.length < 8){
+      _passwordError = S.of(context).passwordTooShort;
+      return false;
+    }
     return true;
   }
 
   createEditEmployee()async{
     try{
+      context.loaderOverlay.show();
       User user = User(
         id: widget.employee?.id,
         name: '${_fName.text} ${_lName.text}'.toLowerCase(),
         avatar: widget.employee?.avatar,
         firstName: _fName.text,
         lastName: _lName.text,
-        email: _email.text.isNotEmpty?_email.text:'${_fName.text}${_lName.text}@${context.read<CompanyModel>().myCompany.name}.com'.toLowerCase().replaceAll(' ', ''),
+        email: _email.text,
         pin: _password.text,
         address1: _address1.text,
         address2: _address2.text,
@@ -219,66 +210,64 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
         postalCode: _postal.text,
         phone: _phone.text,
         employeeCode: _employeeCode.text,
-        function: _function.text,
         start: _startDate!=null?_startDate.toString().split(" ").first:null,
         salary: _salary.text,
         birthday: _startDate!=null?_birthDay.toString().split(" ").first:null,
         nfc: _nfc.text,
         language: language,
-        lunchTime: lunchTime,
-        role: "employee",
-        canNTCTracking: widget.employee?.canNTCTracking,
+        role: (widget.employee?.role)??'employee',
         companyId: widget.employee?.companyId,
-        createdAt: widget.employee?.createdAt,
-        emailVerifiedAt: widget.employee?.emailVerifiedAt,
-        emailVerifyNumber: widget.employee?.emailVerifyNumber,
-        firebaseToken: widget.employee?.firebaseToken,
-        sendScheduleNotification: widget.employee?.sendScheduleNotification,
         token: widget.employee?.token,
-        lastPunch: widget.employee?.lastPunch,
-        punches: widget.employee?.punches,
         type: widget.employee?.type??'shop_daily',
-        updatedAt: widget.employee?.updatedAt,
-        works: widget.employee?.works
+        active: widget.employee?.active??true,
       );
-      String base64Image;
-      if(_photoFile!=null){
-        base64Image = base64Encode(_photoFile.readAsBytesSync());
+      String? base64Image;
+      if(_photoFile != null){
+        base64Image = base64Encode(_photoFile!.readAsBytesSync());
       }
-      String result = await context.read<CompanyModel>().createEditEmployee(user,base64Image);
-      if(result==null){
-        showMessage(S.of(context).success);
-        Future.delayed(Duration(seconds: 1)).whenComplete((){
+      String? result = await context.read<CompanyProvider>().createEditEmployee(user, base64Image);
+      context.loaderOverlay.hide();
+      if(result == null){
+        Tools.showSuccessMessage(context, S.of(context).success);
+        if(widget.pageController != null){
+          widget.pageController!.jumpToPage(0);
+        }else if(widget.employee != null){
           Navigator.pop(context);
-        });
+        }
       }else{
-        showMessage(result);
+        Tools.showErrorMessage(context, result);
       }
     }catch(e){
-      print("[createEditEmployee] $e");
-      showMessage(e.toString());
+      context.loaderOverlay.hide();
+      Tools.consoleLog("[createEditEmployee] $e");
+      Tools.showErrorMessage(context, e.toString());
     }
   }
 
   @override
+  void dispose() {
+    NfcManager.instance.stopSession().catchError((e){
+      Tools.consoleLog('[CreateEditEmployee.NfcManager.instance.stopSession]$e');
+    });
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
+    final settings = context.watch<CompanyProvider>().myCompanySettings;
     return Scaffold(
       key: _scaffoldKey,
       appBar: AppBar(
-        title: Text(widget.employee==null?S.of(context).createNewEmployee:S.of(context).editEmployee,style: TextStyle(color: Colors.black87,fontSize: 25,fontWeight: FontWeight.bold),),
-        leading: IconButton(
-          icon: Icon(Icons.arrow_back,color: Colors.black87,size: 35,),
-          onPressed: ()=>Navigator.pop(context),
-        ),
+        title: Text(widget.employee==null?S.of(context).createNewEmployee:S.of(context).editEmployee,),
         backgroundColor: Color(primaryColor),
         elevation: 0,
+        centerTitle: true,
       ),
       body: SingleChildScrollView(
         child: Container(
           decoration: BoxDecoration(
             color: Colors.white,
-            border: Border.all(color: Colors.grey),
-            borderRadius: BorderRadius.circular(20),
+            borderRadius: BorderRadius.circular(10),
           ),
           width: MediaQuery.of(context).size.width,
           margin: EdgeInsets.all(8),
@@ -294,53 +283,51 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                       Positioned(
                         right: 0,
                         bottom: 0,
-                        child: ButtonTheme(
-                          height: 0,
-                          minWidth: 50,
-                          splashColor: Color(primaryColor),
-                          child: RaisedButton(
-                              child: Icon(Icons.camera_enhance,),
-                              color: Colors.white,
-                              shape: CircleBorder(),
-                              padding: EdgeInsets.all(4),
-                              materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                              onPressed: (){
-                                FocusScope.of(context).requestFocus(FocusNode());
-                                showModalBottomSheet(
-                                    context: context,
-                                    shape: RoundedRectangleBorder(borderRadius: BorderRadius.vertical(top: Radius.circular(8))),
-                                    clipBehavior: Clip.hardEdge,
-                                    builder: (c){
-                                      return Container(
-                                        height: 104,
-                                        padding: EdgeInsets.symmetric(vertical: 4),
-                                        child: Column(
-                                          children: [
-                                            ListTile(
-                                              dense: true,
-                                              leading: Icon(Icons.camera_alt_outlined),
-                                              title: Text(S.of(context).camera),
-                                              onTap: (){
-                                                Navigator.pop(c);
-                                                _pickUserPhoto(ImageSource.camera);
-                                              },
-                                            ),
-                                            ListTile(
-                                              dense: true,
-                                              leading: Icon(Icons.photo_library),
-                                              title: Text(S.of(context).gallery),
-                                              onTap: (){
-                                                Navigator.pop(c);
-                                                _pickUserPhoto(ImageSource.gallery);
-                                              },
-                                            )
-                                          ],
-                                        ),
-                                      );
-                                    },
-                                );
-                              }
-                          ),
+                        child: MaterialButton(
+                            height: 0,
+                            minWidth: 50,
+                            splashColor: Color(primaryColor),
+                            child: Icon(Icons.camera_enhance,),
+                            color: Colors.white,
+                            shape: CircleBorder(),
+                            padding: EdgeInsets.all(4),
+                            materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
+                            onPressed: (){
+                              FocusScope.of(context).requestFocus(FocusNode());
+                              showModalBottomSheet(
+                                  context: context,
+                                  shape: RoundedRectangleBorder(borderRadius: BorderRadius.vertical(top: Radius.circular(8))),
+                                  clipBehavior: Clip.hardEdge,
+                                  builder: (c){
+                                    return Container(
+                                      height: 104,
+                                      padding: EdgeInsets.symmetric(vertical: 4),
+                                      child: Column(
+                                        children: [
+                                          ListTile(
+                                            dense: true,
+                                            leading: Icon(Icons.camera_alt_outlined),
+                                            title: Text(S.of(context).camera),
+                                            onTap: (){
+                                              Navigator.pop(c);
+                                              _pickUserPhoto(ImageSource.camera);
+                                            },
+                                          ),
+                                          ListTile(
+                                            dense: true,
+                                            leading: Icon(Icons.photo_library),
+                                            title: Text(S.of(context).gallery),
+                                            onTap: (){
+                                              Navigator.pop(c);
+                                              _pickUserPhoto(ImageSource.gallery);
+                                            },
+                                          )
+                                        ],
+                                      ),
+                                    );
+                                  },
+                              );
+                            }
                         ),
                       ),
                     ],
@@ -355,12 +342,11 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                     suffixIcon: Icon(Icons.person,color: Colors.black87,),
                     isDense: true,
                     labelText: S.of(context).firstName,
-                    labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                     suffixIconConstraints: BoxConstraints(maxHeight: 20),
                     contentPadding: EdgeInsets.zero,
                     errorText: _fNameError
                 ),
-                enabled: !isLoading,
+                enabled: settings?.useOwnData??false,
                 textCapitalization: TextCapitalization.words,
                 textInputAction: TextInputAction.next,
                 keyboardType: TextInputType.name,
@@ -379,12 +365,11 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                     suffixIcon: Icon(Icons.person,color: Colors.black87,),
                     isDense: true,
                     labelText: S.of(context).lastName,
-                    labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                     suffixIconConstraints: BoxConstraints(maxHeight: 20),
                     contentPadding: EdgeInsets.zero,
                     errorText: _lNameError
                 ),
-                enabled: !isLoading,
+                enabled: settings?.useOwnData??false,
                 textCapitalization: TextCapitalization.words,
                 textInputAction: TextInputAction.next,
                 keyboardType: TextInputType.name,
@@ -403,12 +388,11 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                     suffixIcon: Icon(Icons.mail,color: Colors.black87,),
                     isDense: true,
                     labelText: S.of(context).email,
-                    labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                     suffixIconConstraints: BoxConstraints(maxHeight: 20),
                     contentPadding: EdgeInsets.zero,
                     errorText: _emailError
                 ),
-                enabled: !isLoading,
+                enabled: settings?.useOwnData??false,
                 keyboardType: TextInputType.emailAddress,
                 textInputAction: TextInputAction.next,
                 maxLines: 1,
@@ -423,21 +407,19 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                     suffixIcon: Icon(Icons.lock,color: Colors.black87,),
                     isDense: true,
                     labelText: S.of(context).passwordPin,
-                    labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                     suffixIconConstraints: BoxConstraints(maxHeight: 20),
                     contentPadding: EdgeInsets.zero,
                     errorText: _passwordError
                 ),
-                enabled: !isLoading,
-                keyboardType: TextInputType.number,
-                maxLength: 4,
+                enabled: settings?.useOwnData??false,
+                keyboardType: TextInputType.visiblePassword,
                 textInputAction: TextInputAction.next,
                 maxLines: 1,
                 controller: _password,
               ),
               Padding(
                 padding: const EdgeInsets.only(top: 20,bottom: 5),
-                child: Text(S.of(context).address,style: TextStyle(fontWeight: FontWeight.bold,fontSize: 18,color: Colors.grey),),
+                child: Text(S.of(context).address,style: TextStyle(fontWeight: FontWeight.bold,color: Colors.grey),),
               ),
               TextField(
                 decoration: InputDecoration(
@@ -446,11 +428,10 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                     focusedBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.black87)),
                     isDense: true,
                     labelText: S.of(context).streetAddress,
-                    labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                     contentPadding: EdgeInsets.zero,
                     errorText: _addressError
                 ),
-                enabled: !isLoading,
+                enabled: settings?.useOwnData??false,
                 keyboardType: TextInputType.streetAddress,
                 textInputAction: TextInputAction.next,
                 maxLines: 1,
@@ -464,10 +445,9 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                     focusedBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.black87)),
                     isDense: true,
                     labelText: S.of(context).aptSuiteBuilding,
-                    labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                     contentPadding: EdgeInsets.zero,
                 ),
-                enabled: !isLoading,
+                enabled: settings?.useOwnData??false,
                 keyboardType: TextInputType.streetAddress,
                 textInputAction: TextInputAction.next,
                 maxLines: 1,
@@ -478,6 +458,7 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                 initCity: city,
                 initState: state,
                 initCountry: country,
+                readOnly: !(settings?.useOwnData??false),
                 onCountryChanged: (value) {
                   FocusScope.of(context).requestFocus(FocusNode());
                   country = value;
@@ -499,11 +480,10 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                   focusedBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.black87)),
                   isDense: true,
                   labelText: S.of(context).postalCode,
-                  labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                   contentPadding: EdgeInsets.zero,
                   errorText: _postalError
                 ),
-                enabled: !isLoading,
+                enabled: settings?.useOwnData??false,
                 keyboardType: TextInputType.text,
                 textInputAction: TextInputAction.next,
                 maxLines: 1,
@@ -520,10 +500,9 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                   focusedBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.black87)),
                   isDense: true,
                   labelText: S.of(context).phoneNumber,
-                  labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                   contentPadding: EdgeInsets.zero,
                 ),
-                enabled: !isLoading,
+                enabled: settings?.useOwnData??false,
                 keyboardType: TextInputType.phone,
                 textInputAction: TextInputAction.next,
                 maxLines: 1,
@@ -537,11 +516,10 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                   focusedBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.black87)),
                   isDense: true,
                   labelText: "${S.of(context).employee}#",
-                  labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                   contentPadding: EdgeInsets.zero,
                   errorText: _codeError
                 ),
-                enabled: !isLoading,
+                enabled: settings?.useOwnData??false,
                 keyboardType: TextInputType.number,
                 textInputAction: TextInputAction.next,
                 maxLines: 1,
@@ -554,30 +532,11 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                     enabledBorder: UnderlineInputBorder(),
                     focusedBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.black87)),
                     isDense: true,
-                    labelText: S.of(context).employeeFunction,
-                    labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
-                    contentPadding: EdgeInsets.zero,
-                    errorText: _functionError
-                ),
-                enabled: !isLoading,
-                keyboardType: TextInputType.text,
-                textInputAction: TextInputAction.next,
-                maxLines: 1,
-                controller: _function,
-              ),
-              SizedBox(height: 4,),
-              TextField(
-                decoration: InputDecoration(
-                    border: UnderlineInputBorder(),
-                    enabledBorder: UnderlineInputBorder(),
-                    focusedBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.black87)),
-                    isDense: true,
                     labelText: S.of(context).startDate,
-                    labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                     contentPadding: EdgeInsets.zero,
                     errorText: _startDateError
                 ),
-                enabled: !isLoading,
+                enabled: settings?.useOwnData??false,
                 readOnly: true,
                 onTap: (){
                   _selectStartDate(context);
@@ -593,13 +552,12 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                     focusedBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.black87)),
                     isDense: true,
                     labelText: S.of(context).salary,
-                    labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                     suffixIconConstraints: BoxConstraints(maxHeight: 20),
-                    suffixIcon: Text("\$/h",style: TextStyle(fontSize: 18),),
+                    suffixIcon: Text("\$/h",style: TextStyle(),),
                     contentPadding: EdgeInsets.zero,
                     errorText: _salaryError
                 ),
-                enabled: !isLoading,
+                enabled: settings?.useOwnData??false,
                 keyboardType: TextInputType.number,
                 textInputAction: TextInputAction.next,
                 maxLines: 1,
@@ -613,11 +571,10 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                     focusedBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.black87)),
                     isDense: true,
                     labelText: S.of(context).birthday,
-                    labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                     contentPadding: EdgeInsets.zero,
                     errorText: _birthDayError
                 ),
-                enabled: !isLoading,
+                enabled: settings?.useOwnData??false,
                 readOnly: true,
                 onTap: (){
                   _selectBirthDate(context);
@@ -633,93 +590,77 @@ class _CreateEditEmployeeState extends State<CreateEditEmployee> {
                     focusedBorder: UnderlineInputBorder(borderSide: BorderSide(color: Colors.black87)),
                     isDense: true,
                     labelText: "NFC",
-                    labelStyle: TextStyle(color: Colors.grey,fontSize: 18),
                     contentPadding: EdgeInsets.zero,
                 ),
-                enabled: !isLoading,
-                onTap: (){
-                  FlutterNfcReader.read().then((NfcData data)async{
-                    if(data!=null){
-                      if(data.id!=null && data.id.isNotEmpty){
-                        _nfc.text = data.id;
-                      }else{
-                        _nfc.text = data.content;
-                      }
-                    }
-                  }).catchError((e){
-                    showMessage(e.toString());
-                  });
+                onTap: ()async{
+                  if(await NfcManager.instance.isAvailable()){
+                    NfcManager.instance.startSession(
+                      onDiscovered: (NfcTag tag) async {
+                        Tools.playSound();
+                        NfcManager.instance.stopSession();
+                        Tools.consoleLog('[NFC Scanned][${tag.data}]');
+                        String? nfc = Tools.getNFCIdentifier(tag.data);
+                        if(nfc != null){
+                          _nfc.text = nfc;
+                        }else{
+                          Tools.showErrorMessage(context, S.of(context).invalidNFC);
+                        }
+                      },
+                      alertMessage: 'NFC Scanned!',
+                      onError: (NfcError error)async{
+                        Tools.showErrorMessage(context, error.message);
+                      },
+                    ).catchError((e){
+                      Tools.consoleLog('[CreateEditEmployee.NfcManager.startSession.err]$e');
+                    });
+                  }else{
+                    Tools.showErrorMessage(context, S.of(context).notAllowedNFC);
+                  }
                 },
                 maxLines: 1,
                 controller: _nfc,
               ),
               SizedBox(height: 4,),
               DropdownButton<String>(
-                items: ['English','Spanish','French'].map((String value) {
+                items: ['English', 'Spanish', 'French'].map((String value) {
                   return DropdownMenuItem<String>(
                     value: value,
                     child: Padding(
                       padding: const EdgeInsets.symmetric(vertical: 2,horizontal: 8),
-                      child: Text(value),
+                      child: Text(value, style: TextStyle(fontSize: 14),),
                     ),
                   );
                 }).toList(),
                 underline: Container(color: Colors.black87,width: double.infinity,height: 1,),
                 style: TextStyle(fontSize: 20, color: Colors.black87),
-                hint: Text(S.of(context).chooseLanguage),
+                hint: Text(S.of(context).chooseLanguage, style: TextStyle(fontSize: 14),),
                 isExpanded: true,
-                onChanged: (v) {
+                onChanged: !(settings?.useOwnData??false)? null :(v) {
                   setState(() { language = v; });
                   FocusScope.of(context).requestFocus(FocusNode());
                 },
                 value: language,
+                disabledHint: Text('$language', style: TextStyle(fontSize: 14),),
               ),
-              Row(
-                children: [
-                  Checkbox(
-                      value: lunchTime>0,
-                      onChanged: (v){
-                        setState(() {
-                          if(v){
-                            lunchTime=30;
-                          }else{
-                            lunchTime=0;
-                          }
-                        });
-                      },
-                    materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                  ),
-                  Text(S.of(context).hasLunchBreak)
-                ],
-              ),
-              ButtonTheme(
-                minWidth: MediaQuery.of(context).size.width-60,
-                padding: EdgeInsets.all(8),
-                splashColor: Color(primaryColor),
-                child: RaisedButton(
-                  child: isLoading?
-                  SizedBox(
-                    height: 28,
-                    width: 28,
-                    child: CircularProgressIndicator(),
-                  )
-                  :Padding(
-                    padding: const EdgeInsets.all(4.0),
-                    child: Text(S.of(context).save.toUpperCase(),style: TextStyle(fontSize: 18,fontWeight: FontWeight.bold,color: Colors.white),),
-                  ),
-                  onPressed: ()async{
-                    FocusScope.of(context).requestFocus(FocusNode());
-                    bool valid = validator();
+
+              MaterialButton(
+                minWidth: MediaQuery.of(context).size.width,
+                height: 40,
+                color: Colors.black87,
+                shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
+                child: Padding(
+                      padding: const EdgeInsets.all(4.0),
+                      child: Text(S.of(context).save.toUpperCase(),style: TextStyle(fontWeight: FontWeight.bold,color: Colors.white),),
+                    ),
+                onPressed: ()async{
+                  FocusScope.of(context).requestFocus(FocusNode());
+                  bool valid = validator();
+                  if(valid){
+                    await createEditEmployee();
+                  }else{
                     setState(() {});
-                    if(valid){
-                      setState(() {isLoading = true;});
-                      await createEditEmployee();
-                      setState(() {isLoading = false;});
-                    }
-                  },
-                  color: Colors.black87,
-                  shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
-                ),
+                  }
+                },
               ),
             ],
           ),
