@@ -1,12 +1,15 @@
 import 'dart:async';
-
-import 'package:facepunch/lang/l10n.dart';
-import 'package:facepunch/models/app_const.dart';
-import 'package:facepunch/models/company_model.dart';
-import 'package:facepunch/models/harvest_model.dart';
 import 'package:flutter/material.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
 import 'package:provider/provider.dart';
+
+import '/widgets/utils.dart';
+import '/lang/l10n.dart';
+import '/config/app_const.dart';
+import '/models/company_model.dart';
+import '/models/harvest_model.dart';
+import '/providers/company_provider.dart';
+import '/providers/harvest_provider.dart';
 
 class HarvestReportScreen extends StatefulWidget{
 
@@ -15,14 +18,14 @@ class HarvestReportScreen extends StatefulWidget{
 }
 
 class _HarvestReportScreenState extends State<HarvestReportScreen>{
+
   DateTime selectedDate = DateTime.now();
   RefreshController _refreshController = RefreshController(initialRefresh: false);
-  HTask selectedTask;
+  HTask? selectedTask;
   List<HarvestEmployeeStats> employeeStats = [];
-  HarvestCompanyStats companyStats;
+  HarvestCompanyStats? companyStats;
   List<HarvestDateStats> dateStats = [];
-  final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
-  Timer _timer;
+  late Timer _timer;
   bool isFetchingDateStats = false;
   bool isShowEmployeeStats = true;
 
@@ -40,7 +43,7 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
   }
 
   _selectHarvestDate() async {
-    final DateTime picked = await showDatePicker(
+    final DateTime? picked = await showDatePicker(
         context: context,
         initialDate: selectedDate,
         initialDatePickerMode: DatePickerMode.day,
@@ -63,10 +66,10 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
   }
 
   _getEmployeeStats()async{
-    if(selectedTask!=null && selectedDate !=null){
-      final result = await context.read<HarvestModel>().getEmployeeHarvestStats(date: selectedDate.toString(),field: selectedTask.fieldId);
+    if(selectedTask!=null){
+      final result = await context.read<HarvestProvider>().getEmployeeHarvestStats(selectedDate.toString(), selectedTask?.fieldId);
       if(result is String){
-        showMessage(result);
+        Tools.showErrorMessage(context, result);
       }else{
         employeeStats = result;
       }
@@ -75,9 +78,9 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
   }
 
   _getDateStats()async{
-    final result = await context.read<HarvestModel>().getDateHarvestStats(selectedDate.toString());
+    final result = await context.read<HarvestProvider>().getDateHarvestStats(selectedDate.toString());
     if(result is String){
-      showMessage(result);
+      Tools.showErrorMessage(context, result);
     }else{
       dateStats = result;
     }
@@ -94,27 +97,15 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
 
 
   _getCompanyStats()async{
-    if(selectedTask!=null && selectedDate !=null){
-      final result = await context.read<HarvestModel>().getCompanyHarvestStats(date: selectedDate.toString(),field: selectedTask.fieldId);
+    if(selectedTask!=null){
+      final result = await context.read<HarvestProvider>().getCompanyHarvestStats(selectedDate.toString(), selectedTask?.fieldId);
       if(result is String){
-        showMessage(result);
+        Tools.showErrorMessage(context, result);
       }else{
         companyStats = result;
       }
       if(mounted)setState(() {});
     }
-  }
-
-  showMessage(String message){
-    _scaffoldKey.currentState.hideCurrentSnackBar();
-    _scaffoldKey.currentState.showSnackBar(
-        SnackBar(
-          content: Text(message),
-          backgroundColor: Colors.red,
-          duration: Duration(seconds: 2),
-          action: SnackBarAction(onPressed: (){},label: S.of(context).close,textColor: Colors.white,),
-        )
-    );
   }
 
   @override
@@ -124,33 +115,34 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
     super.dispose();
   }
 
-  Color getLogColor(double avg, CompanySettings settings){
+  Color getLogColor(double avg, CompanySettings? settings){
     try{
       double lowValue = 2.5;
       double highValue = 3.0;
       Color lowColor = Color(0xFFe24c3c);
       Color mediumColor = Color(0xFFe2e03c);
       Color highColor = Color(0xFF6fe23c);
-      if(settings.lowValue!=null)lowValue = double.parse(settings.lowValue);
-      if(settings.highValue!=null)highValue = double.parse(settings.highValue);
-      if(settings.lowColor!=null)lowColor = Color(int.parse(settings.lowColor,radix: 16));
-      if(settings.mediumColor!=null)mediumColor = Color(int.parse(settings.mediumColor,radix: 16));
-      if(settings.highColor!=null)highColor = Color(int.parse(settings.highColor,radix: 16));
-      if(avg>=highValue) return highColor;
-      if(avg>=lowValue) return mediumColor;
+      if(settings != null){
+        if(settings.lowValue != null)lowValue = double.parse(settings.lowValue!);
+        if(settings.highValue != null)highValue = double.parse(settings.highValue!);
+        if(settings.lowColor != null)lowColor = Color(int.parse(settings.lowColor!, radix: 16));
+        if(settings.mediumColor != null)mediumColor = Color(int.parse(settings.mediumColor!, radix: 16));
+        if(settings.highColor != null)highColor = Color(int.parse(settings.highColor!, radix: 16));
+      }
+      if(avg >= highValue) return highColor;
+      if(avg >= lowValue) return mediumColor;
       return lowColor;
     }catch(e){
-      print('[getLogColor]$e');
+      Tools.consoleLog('[getLogColor]$e');
       return Colors.white;
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    List<HTask> tasks = context.watch<HarvestModel>().tasks;
-    final settings = context.watch<CompanyModel>().myCompanySettings;
+    List<HTask> tasks = context.watch<HarvestProvider>().tasks;
+    final settings = context.watch<CompanyProvider>().myCompanySettings;
     return Scaffold(
-      key: _scaffoldKey,
       appBar: AppBar(
         backgroundColor: Color(primaryColor),
         title: Text(S.of(context).harvestReport),
@@ -163,18 +155,14 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
             child: Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                FlatButton(
+                TextButton(
                   onPressed: (){
                     setState(() {
                       selectedDate = selectedDate.subtract(Duration(days: 1));
                     });
                     _refreshController.requestRefresh();
                   },
-                  shape: CircleBorder(),
-                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
                   child: Icon(Icons.arrow_back_ios_outlined,color: Colors.black,size: 30,),
-                  padding: EdgeInsets.all(4),
-                  minWidth: 0,
                 ),
                 SizedBox(width: 10,),
                 MaterialButton(
@@ -185,18 +173,14 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
                   child: Text(selectedDate.toString().split(' ')[0],style: TextStyle(fontSize: 16,fontWeight: FontWeight.w500),),
                 ),
                 SizedBox(width: 10,),
-                FlatButton(
+                TextButton(
                   onPressed: (){
                     setState(() {
                       selectedDate = selectedDate.add(Duration(days: 1));
                     });
                     _refreshController.requestRefresh();
                   },
-                  shape: CircleBorder(),
-                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
                   child: Icon(Icons.arrow_forward_ios_outlined,color: Colors.black,size: 30,),
-                  padding: EdgeInsets.all(4),
-                  minWidth: 0,
                 ),
               ],
             ),
@@ -234,12 +218,26 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
                                   child: Column(
                                     mainAxisAlignment: MainAxisAlignment.center,
                                     children: [
-                                      Text(task.field.name.substring(0, task.field.name.length>1?2:task.field.name.length).toUpperCase(),style: TextStyle(fontSize: 16,fontWeight: FontWeight.w500),),
+                                      Text('${task.field?.shortName()}',
+                                        style: TextStyle(
+                                          color: selectedTask==task?Colors.white:Colors.black87
+                                        ),
+                                      ),
                                       Padding(
                                         padding: const EdgeInsets.symmetric(horizontal: 2),
-                                        child: FittedBox(child: Text(task.field.crop,),),
+                                        child: FittedBox(
+                                          child: Text('${task.field?.crop}',
+                                            style: TextStyle(
+                                                color: selectedTask==task?Colors.white:Colors.black87
+                                            ),
+                                          ),
+                                        )
                                       ),
-                                      Text(task.container.name[0].toUpperCase(),style: TextStyle(fontSize: 16,fontWeight: FontWeight.w500),),
+                                      Text('${task.container?.shortName()}',
+                                        style: TextStyle(
+                                            color: selectedTask==task?Colors.white:Colors.black87
+                                        ),
+                                      ),
                                     ],
                                   ),
                                 ),
@@ -270,123 +268,123 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
                     child: isFetchingDateStats?SizedBox(
                       width: 24,
                       height: 24,
-                      child: CircularProgressIndicator(backgroundColor: Colors.white,)
+                      child: CircularProgressIndicator(backgroundColor: Colors.white, strokeWidth: 2,)
                     ):Icon(Icons.title),
                   ),
                 )
               ],
             ),
           ),
-          if(companyStats!=null && isShowEmployeeStats)
-          Container(
-            padding: EdgeInsets.all(16),
-            child: Column(
-              children: [
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Expanded(
-                      flex: 1,
-                      child: Text(S.of(context).totalOfTheDay)
-                    ),
-                    Expanded(
+          if(companyStats != null && isShowEmployeeStats)
+            Container(
+              padding: EdgeInsets.all(16),
+              child: Column(
+                children: [
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      Expanded(
                         flex: 1,
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            for(var container in companyStats.containersOfDate)
-                              Text('${container['quantity']} ${container['name']}'),
-                          ],
-                        )
-                    ),
-                    Flexible(
-                        flex: 1,
-                        child: Text('${companyStats.harvestTimeOfDate==0?'0.00':(companyStats.quantityOfDate/companyStats.harvestTimeOfDate).toStringAsFixed(2)} ${S.of(context).containerHour}'),
-                    ),
-                  ],
-                ),
-                SizedBox(height: 8,),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Expanded(
-                        flex: 1,
-                        child: Text(S.of(context).totalOfTheSeason)
-                    ),
-                    Expanded(
-                        flex: 1,
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            for(var container in companyStats.containersOfYear)
-                              Text('${container['quantity']} ${container['name']}'),
-                          ],
-                        )
-                    ),
-                    Flexible(
-                        child: Text('${companyStats.harvestTimeOfYear==0?'0.00':(companyStats.quantityOfYear/companyStats.harvestTimeOfYear).toStringAsFixed(2)} ${S.of(context).containerHour}')
-                    ),
-                  ],
-                ),
-              ],
+                        child: Text(S.of(context).totalOfTheDay)
+                      ),
+                      Expanded(
+                          flex: 1,
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              for(var container in companyStats!.containersOfDate??[])
+                                Text('${container['quantity']} ${container['name']}'),
+                            ],
+                          )
+                      ),
+                      Flexible(
+                          flex: 1,
+                          child: Text('${companyStats!.dateAvg()} ${S.of(context).containerHour}'),
+                      ),
+                    ],
+                  ),
+                  SizedBox(height: 8,),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    children: [
+                      Expanded(
+                          flex: 1,
+                          child: Text(S.of(context).totalOfTheSeason)
+                      ),
+                      Expanded(
+                          flex: 1,
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              for(var container in companyStats!.containersOfYear??[])
+                                Text('${container['quantity']} ${container['name']}'),
+                            ],
+                          )
+                      ),
+                      Flexible(
+                          child: Text('${companyStats!.yearAvg()} ${S.of(context).containerHour}')
+                      ),
+                    ],
+                  ),
+                ],
+              ),
             ),
-          ),
           if(isShowEmployeeStats)
-          Container(
-            color: Color(primaryColor),
-            padding: EdgeInsets.symmetric(vertical: 4),
-            child: Row(
+            Container(
+              color: Color(primaryColor),
+              padding: EdgeInsets.symmetric(vertical: 4),
+              child: Row(
+                children: [
+                  Expanded(
+                      flex: 1,
+                      child: Text(S.of(context).employee,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                  ),
+                  Expanded(
+                      flex: 1,
+                      child: Text(S.of(context).quantity,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                  ),
+                  Expanded(
+                      flex: 1,
+                      child: Text(S.of(context).containerHour,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                  ),
+                ],
+              ),
+            ),
+          if(!isShowEmployeeStats)
+            Column(
               children: [
-                Expanded(
-                    flex: 1,
-                    child: Text(S.of(context).employee,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                Padding(
+                  padding: const EdgeInsets.all(8.0),
+                  child: Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                    children: [
+                      Text(S.of(context).totalOfTheDay),
+                      Text('${harvestTotalOfDate()} ${S.of(context).containers}'),
+                    ],
+                  ),
                 ),
-                Expanded(
-                    flex: 1,
-                    child: Text(S.of(context).quantity,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
-                ),
-                Expanded(
-                    flex: 1,
-                    child: Text(S.of(context).containerHour,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                Container(
+                  color: Color(primaryColor),
+                  padding: EdgeInsets.symmetric(vertical: 4),
+                  child: Row(
+                    children: [
+                      Expanded(
+                          flex: 1,
+                          child: Text(S.of(context).field,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                      ),
+                      Expanded(
+                          flex: 1,
+                          child: Text(S.of(context).container,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                      ),
+                      Expanded(
+                          flex: 1,
+                          child: Text(S.of(context).quantity,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                      ),
+                    ],
+                  ),
                 ),
               ],
             ),
-          ),
-          if(!isShowEmployeeStats)
-          Column(
-            children: [
-              Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: [
-                    Text(S.of(context).totalOfTheDay),
-                    Text('${harvestTotalOfDate()} ${S.of(context).containers}'),
-                  ],
-                ),
-              ),
-              Container(
-                color: Color(primaryColor),
-                padding: EdgeInsets.symmetric(vertical: 4),
-                child: Row(
-                  children: [
-                    Expanded(
-                        flex: 1,
-                        child: Text(S.of(context).field,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
-                    ),
-                    Expanded(
-                        flex: 1,
-                        child: Text(S.of(context).container,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
-                    ),
-                    Expanded(
-                        flex: 1,
-                        child: Text(S.of(context).quantity,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
           Expanded(
             child: SmartRefresher(
               enablePullDown: true,
@@ -400,14 +398,14 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
                     Container(
                       padding: EdgeInsets.symmetric(vertical: 4),
                       decoration: BoxDecoration(
-                          color: getLogColor(stats.time==0?0:(stats.quantity/stats.time), settings),
+                        color: getLogColor(stats.avg(), settings),
                         border: Border(bottom: BorderSide(color: Colors.grey)),
                       ),
                       child: Row(
                         children: [
                           Expanded(
                               flex: 1,
-                              child: Text(stats.employeeName,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                              child: Text('${stats.employeeName}',style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
                           ),
                           Expanded(
                               flex: 1,
@@ -415,7 +413,7 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
                           ),
                           Expanded(
                               flex: 1,
-                              child: Text(stats.time==0?'0.00':'${(stats.quantity/stats.time).toStringAsFixed(2)}',style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                              child: Text('${stats.avg().toStringAsFixed(2)}',style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
                           ),
                         ],
                       ),
@@ -431,13 +429,13 @@ class _HarvestReportScreenState extends State<HarvestReportScreen>{
                         children: [
                           Expanded(
                             flex: 1,
-                            child: Text(stats.fieldName,textAlign: TextAlign.center,)
+                            child: Text('${stats.fieldName}', textAlign: TextAlign.center,)
                           ),
                           Expanded(
                               flex: 2,
                               child: Column(
                                 children: [
-                                  for(var container in stats.containers)
+                                  for(var container in stats.containers??[])
                                     Row(
                                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                                       children: [
