@@ -1,15 +1,18 @@
-import 'package:audioplayers/audio_cache.dart';
-import 'package:facepunch/lang/l10n.dart';
-import 'package:facepunch/models/app_const.dart';
-import 'package:facepunch/models/harvest_model.dart';
-import 'package:facepunch/screens/admin/nfc/nfc_settings.dart';
-import 'package:facepunch/widgets/popover/cool_ui.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_nfc_reader/flutter_nfc_reader.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
+import 'package:nfc_manager/nfc_manager.dart';
 import 'package:provider/provider.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
+import 'package:collection/collection.dart';
+import 'package:loader_overlay/loader_overlay.dart';
+
+import '/lang/l10n.dart';
+import '/config/app_const.dart';
+import '/models/harvest_model.dart';
+import '/screens/admin/nfc/nfc_settings.dart';
+import '/widgets/utils.dart';
+import '/widgets/popover/cool_ui.dart';
+import '/providers/harvest_provider.dart';
 
 class NFCScanPage extends StatefulWidget{
 
@@ -18,17 +21,17 @@ class NFCScanPage extends StatefulWidget{
 }
 
 class _NFCScanPageState extends State<NFCScanPage>{
-  final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
+
   bool isBigNFCImage = true;
   DateTime selectedDate = DateTime.now();
-  HTask selectedTask;
+  HTask? selectedTask;
   List<Harvest> harvests = [];
   RefreshController _refreshController = RefreshController(initialRefresh: true);
   bool isLoading = false;
-  final AudioCache player = AudioCache();
+  List<HTask> tasks = [];
 
   void _onRefresh() async{
-    harvests = await context.read<HarvestModel>().getHarvestsOfDate(selectedDate.toString());
+    harvests = await context.read<HarvestProvider>().getHarvestsOfDate(selectedDate.toString());
     _refreshController.refreshCompleted();
     if(mounted)setState(() {});
   }
@@ -41,30 +44,31 @@ class _NFCScanPageState extends State<NFCScanPage>{
 
   @override
   void dispose() {
-    FlutterNfcReader.stop().catchError(print);
+    NfcManager.instance.stopSession().catchError((e){
+      Tools.consoleLog('[NFCScanPage.NfcManager.instance.stopSession]$e');
+    });
     super.dispose();
   }
 
-  showHarvestTaskDialog({HTask task}){
-    List<HContainer> containers = context.read<HarvestModel>().containers;
+  _showHarvestTaskDialog(HTask? task){
+    List<HContainer> containers = context.read<HarvestProvider>().containers;
     if(containers.isEmpty){
-      showMessage(S.of(context).addContainers);
+      Tools.showErrorMessage(context, S.of(context).addContainers);
       return null;
     }
-    List<Field> fields = context.read<HarvestModel>().fields;
+    List<Field> fields = context.read<HarvestProvider>().fields;
     if(fields.isEmpty){
-      showMessage(S.of(context).addFields);
+      Tools.showErrorMessage(context, S.of(context).addFields);
       return null;
     }
     HContainer selectedContainer;
     Field selectedField;
-    bool isUpdating = false;
-    if(task==null){
+    if(task == null){
       selectedContainer = containers[0];
       selectedField = fields[0];
     }else{
-      selectedContainer = task.container;
-      selectedField = task.field;
+      selectedContainer = task.container!;
+      selectedField = task.field!;
     }
     showDialog(
       context: context,
@@ -72,8 +76,11 @@ class _NFCScanPageState extends State<NFCScanPage>{
       builder: (_context){
         return AlertDialog(
             title: Text(task==null?S.of(context).createNewTask:S.of(context).updateTask,textAlign: TextAlign.center,),
+            contentPadding: EdgeInsets.all(8),
+            insetPadding: EdgeInsets.zero,
             content:StatefulBuilder(
               builder: (_,setState)=>Container(
+                width: MediaQuery.of(context).size.width-40,
                 child: Column(
                   mainAxisSize: MainAxisSize.min,
                   crossAxisAlignment: CrossAxisAlignment.start,
@@ -82,7 +89,7 @@ class _NFCScanPageState extends State<NFCScanPage>{
                     Container(
                       decoration: BoxDecoration(
                         border: Border.all(color: Colors.black54),
-                        borderRadius: BorderRadius.circular(30),
+                        borderRadius: BorderRadius.circular(4),
                       ),
                       padding: EdgeInsets.symmetric(horizontal: 8.0,vertical: 4),
                       clipBehavior: Clip.hardEdge,
@@ -93,13 +100,15 @@ class _NFCScanPageState extends State<NFCScanPage>{
                             child: Text('${value.name}, ${value.crop} , ${value.cropVariety}',style: TextStyle(fontSize: 18),),
                           );
                         }).toList(),
-                        value: fields.firstWhere((f) => f.id == selectedField.id,orElse: ()=>null),
+                        value: fields.firstWhereOrNull((f) => f.id == selectedField.id),
                         isExpanded: true,
                         isDense: true,
                         underline: SizedBox(),
                         onChanged: (f) {
                           setState((){
-                            selectedField = f;
+                            if(f != null){
+                              selectedField = f;
+                            }
                           });
                         },
                       ),
@@ -109,7 +118,7 @@ class _NFCScanPageState extends State<NFCScanPage>{
                     Container(
                       decoration: BoxDecoration(
                         border: Border.all(color: Colors.black54),
-                        borderRadius: BorderRadius.circular(30),
+                        borderRadius: BorderRadius.circular(4),
                       ),
                       padding: EdgeInsets.symmetric(horizontal: 8.0,vertical: 4),
                       clipBehavior: Clip.hardEdge,
@@ -117,61 +126,49 @@ class _NFCScanPageState extends State<NFCScanPage>{
                         items: containers.map((HContainer value) {
                           return new DropdownMenuItem<HContainer>(
                             value: value,
-                            child: new Text(value.name,style: TextStyle(fontSize: 18),),
+                            child: new Text('${value.name}',style: TextStyle(fontSize: 18),),
                           );
                         }).toList(),
                         isExpanded: true,
                         isDense: true,
                         underline: SizedBox(),
-                        value: containers.firstWhere((c) =>c.id == selectedContainer.id,orElse: ()=>null),
+                        value: containers.firstWhereOrNull((c) =>c.id == selectedContainer.id),
                         onChanged: (c) {
                           setState((){
-                            selectedContainer = c;
+                            if(c != null){
+                              selectedContainer = c;
+                            }
                           });
                         },
                       ),
                     ),
                     SizedBox(height: 16,),
                     Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      mainAxisAlignment: MainAxisAlignment.end,
                       children: [
-                        RaisedButton(
-                          child: Padding(
-                            padding: const EdgeInsets.all(4.0),
-                            child: isUpdating
-                                ?SizedBox( height: 14,width: 14,child: CircularProgressIndicator(backgroundColor: Colors.white,))
-                                :Text(S.of(context).save.toUpperCase(),style: TextStyle(color: Colors.white),),
-                          ),
-                          color: Color(primaryColor),
-                          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
+                        TextButton(
+                          child: Text(S.of(context).close,style: TextStyle(color: Colors.red),),
+                          onPressed: ()=>Navigator.pop(_context),
+                        ),
+                        TextButton(
+                          child: Text(S.of(context).save, style: TextStyle(color: Color(primaryColor)),),
                           onPressed: ()async{
-                            if(!isUpdating){
-                              if(task!=null){
-                                task.field = selectedField;
-                                task.container = selectedContainer;
-                              }else{
-                                task = HTask(field: selectedField,container: selectedContainer, containerId: selectedContainer.id,fieldId: selectedField.id);
-                              }
-                              setState((){isUpdating = true;});
-                              String result = await context.read<HarvestModel>().createOrUpdateTask(task);
-                              setState((){isUpdating = false;});
-                              if(result!=null){
-                                showMessage(result);
-                              }
-                              Navigator.pop(_context);
+                            if(task != null){
+                              task!.field = selectedField;
+                              task!.fieldId = selectedField.id;
+                              task!.container = selectedContainer;
+                              task!.containerId = selectedContainer.id;
+                            }else{
+                              task = HTask(field: selectedField, container: selectedContainer, containerId: selectedContainer.id, fieldId: selectedField.id);
+                            }
+                            Navigator.pop(_context);
+                            context.loaderOverlay.show();
+                            String? result = await context.read<HarvestProvider>().createOrUpdateTask(task!);
+                            context.loaderOverlay.hide();
+                            if(result != null){
+                              Tools.showErrorMessage(context, result);
                             }
                           },
-                        ),
-                        RaisedButton(
-                          child: Padding(
-                            padding: const EdgeInsets.all(4.0),
-                            child: Text(S.of(context).close.toUpperCase(),style: TextStyle(color: Colors.white),),
-                          ),
-                          onPressed: ()async{
-                            if(!isUpdating)Navigator.pop(_context);
-                          },
-                          color: Colors.red,
-                          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
                         ),
                       ],
                     )
@@ -181,83 +178,11 @@ class _NFCScanPageState extends State<NFCScanPage>{
             )
         );
       },
-    );
-  }
-
-  confirmDeleteTaskDialog(HTask task){
-    bool isDeleting = false;
-
-    showDialog(
-      context: context,
-      barrierDismissible: false,
-      builder: (_context){
-        return AlertDialog(
-            title: Text(S.of(context).deleteTaskConfirm,textAlign: TextAlign.center,),
-            content:StatefulBuilder(
-              builder: (_,setState)=>Container(
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        RaisedButton(
-                          child: Padding(
-                            padding: const EdgeInsets.all(4.0),
-                            child: isDeleting?SizedBox(
-                              width: 14,
-                              height: 14,
-                              child: CircularProgressIndicator(),
-                            ):Text(S.of(context).delete,style: TextStyle(color: Colors.white),),
-                          ),
-                          color: Color(primaryColor),
-                          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
-                          onPressed: ()async{
-                            if(!isDeleting){
-                              setState((){isDeleting = true;});
-                              String result = await context.read<HarvestModel>().deleteTask(task);
-                              setState((){isDeleting = true;});
-                              Navigator.pop(_context);
-                              if(result!=null)showMessage(result);
-                            }
-                          },
-                        ),
-                        RaisedButton(
-                          child: Padding(
-                            padding: const EdgeInsets.all(4.0),
-                            child: Text(S.of(context).close.toUpperCase(),style: TextStyle(color: Colors.white),),
-                          ),
-                          onPressed: ()async{
-                            if(!isDeleting)Navigator.pop(_context);
-                          },
-                          color: Colors.red,
-                          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
-                        ),
-                      ],
-                    )
-                  ],
-                ),
-              ),
-            )
-        );
-      },
-    );
-  }
-
-  showMessage(String message){
-    _scaffoldKey.currentState.hideCurrentSnackBar();
-    _scaffoldKey.currentState.showSnackBar(
-        SnackBar(
-          content: Text(message),
-          backgroundColor: Colors.red,
-          duration: Duration(seconds: 2),
-          action: SnackBarAction(onPressed: (){},label: S.of(context).close,textColor: Colors.white,),
-        )
     );
   }
 
   _selectHarvestDate() async {
-    final DateTime picked = await showDatePicker(
+    final DateTime? picked = await showDatePicker(
         context: context,
         initialDate: selectedDate,
         initialDatePickerMode: DatePickerMode.day,
@@ -273,60 +198,226 @@ class _NFCScanPageState extends State<NFCScanPage>{
     try{
       if(!mounted)return;
       setState(() {isLoading=true;});
-      var result = await context.read<HarvestModel>().addHarvest(task: selectedTask, date: selectedDate.toString(),nfc: nfc);
+      var result = await context.read<HarvestProvider>().addHarvest(task: selectedTask!, date: selectedDate.toString(),nfc: nfc);
+      if(!mounted)return;
       setState(() {isLoading=false;});
-      if(result!=null){
+      if(result != null){
         if(result is Harvest){
-          harvests.insert(0, result);
-          if(mounted)setState(() {});
+          setState(() {
+            harvests.insert(0, result);
+          });
         }else{
-          showMessage(result.toString());
+          Tools.showErrorMessage(context, result.toString());
         }
       }
       _startNfcRead();
     }catch(e){
-      print('[_addHarvest]$e');
-      showMessage(e.toString());
-      setState(() {isLoading=false;});
+      Tools.consoleLog('[_addHarvest]$e');
+      Tools.showErrorMessage(context, e.toString());
+      if(mounted)setState(() {isLoading=false;});
     }
   }
 
   _startNfcRead()async{
-    if(isLoading)return;
-    // await FlutterNfcReader.stop();
-    FlutterNfcReader.read().then((NfcData data)async{
-      if(data!=null){
-        player.play('sound/sound.mp3').catchError(print);
-        print([data.id,data.content]);
-        if(data.id!=null && data.id.isNotEmpty){
-          _addHarvest(data.id);
-        }else if(data.content!=null && data.content.isNotEmpty){
-          _addHarvest(data.content);
-        }
-      }
-    }).catchError((e){
-      showMessage(e.message);
-    });
+    if(isLoading) return;
+    if(await NfcManager.instance.isAvailable()){
+      NfcManager.instance.startSession(
+        onDiscovered: (NfcTag tag) async {
+          Tools.playSound();
+          NfcManager.instance.stopSession();
+          Tools.consoleLog('[NFC Scanned][${tag.data}]');
+          String? nfc = Tools.getNFCIdentifier(tag.data);
+          if(nfc != null){
+            _addHarvest(nfc);
+          }else{
+            Tools.showErrorMessage(context, S.of(context).invalidNFC);
+          }
+        },
+        alertMessage: 'NFC Scanned!',
+        onError: (NfcError error)async{
+          Tools.consoleLog('[NFCScanPage.onError][${error.message}][${error.details}]');
+          Tools.showErrorMessage(context, error.message);
+        },
+      ).catchError((e){
+        Tools.consoleLog('[NFCScanPage._startNfcRead]$e');
+      });
+    }else{
+      Tools.showErrorMessage(context, S.of(context).notAllowedNFC);
+    }
   }
 
+  _deleteHarvestTask(HTask task)async{
+    if(await Tools.confirmDeleting(context, S.of(context).deleteTaskConfirm)){
+      context.loaderOverlay.show();
+      String? result = await context.read<HarvestProvider>().deleteTask(task);
+      context.loaderOverlay.hide();
+      if(result != null){
+        Tools.showErrorMessage(context, result);
+      }else if(mounted)setState(() {
+        if(selectedTask == task) selectedTask = null;
+      });
+    }
+  }
+
+  Widget _harvestTaskItem(HTask task){
+    Widget child = Padding(
+      padding: const EdgeInsets.all(4.0),
+      child: MaterialButton(
+        onPressed: (){
+          setState(() {selectedTask = task;});
+          _startNfcRead();
+        },
+        minWidth: 0,
+        shape: CircleBorder(),
+        color: selectedTask==task?Color(primaryColor):Colors.white,
+        clipBehavior: Clip.antiAlias,
+        padding: EdgeInsets.zero,
+        materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
+        child: AspectRatio(
+          aspectRatio: 1.0,
+          child: Padding(
+            padding: const EdgeInsets.all(2.0),
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Text('${task.field?.shortName()}', style: TextStyle(fontSize: 10),),
+                Text('${task.field?.crop}',style: TextStyle(fontSize: 10), maxLines: 1,),
+                Text('${task.field?.cropVariety}',style: TextStyle(fontSize: 10), maxLines: 1,),
+                Text('${task.container?.shortName()}',style: TextStyle(fontSize: 10),),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+    return CupertinoPopoverButton(
+      popoverBoxShadow: [
+        BoxShadow(color: Colors.black54,blurRadius: 5.0)
+      ],
+      popoverWidth: 180,
+      popoverBuild: (_context){
+        return CupertinoPopoverMenuList(
+          children: <Widget>[
+            CupertinoPopoverMenuItem(
+              child: Padding(
+                padding: const EdgeInsets.all(4.0),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Text(S.of(context).updateTask, style: TextStyle(color: Colors.black87),),
+                    Icon(Icons.edit, color: Colors.black87,),
+                  ],
+                ),
+              ),
+              onTap: (){
+                Navigator.pop(_context);
+                _showHarvestTaskDialog(task);
+                return true;
+              },
+            ),
+            CupertinoPopoverMenuItem(
+              child: Padding(
+                padding: const EdgeInsets.all(4.0),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Text(S.of(context).delete, style: TextStyle(color: Colors.red),),
+                    Icon(Icons.delete, color: Colors.red,),
+                  ],
+                ),
+              ),
+              onTap: (){
+                Navigator.pop(_context);
+                _deleteHarvestTask(task);
+                return true;
+              },
+            )
+          ],
+        );
+      },
+      onTap: (){
+        return true;
+      },
+      child: child,
+    );
+  }
+
+  _showHarvestEditDialog(Harvest harvest){
+    TextEditingController _quantity = TextEditingController(text: harvest.quantity.toString());
+    showDialog(
+      context: context,
+      builder: (_context){
+        return AlertDialog(
+            title: Text(S.of(context).editHarvest, textAlign: TextAlign.center,),
+            content: Container(
+              width: MediaQuery.of(context).size.width-40,
+              child: Column(
+                mainAxisSize: MainAxisSize.min,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text("${S.of(context).employee} : ${harvest.user?.name}"),
+                  Text("${S.of(context).container} : ${harvest.container?.name}"),
+                  Text("${S.of(context).field} : ${harvest.field?.name}"),
+                  SizedBox(height: 8,),
+                  TextField(
+                    keyboardType: TextInputType.number,
+                    decoration: InputDecoration(
+                      border: OutlineInputBorder(),
+                      labelText: S.of(context).quantity,
+                      isDense: true,
+                    ),
+                    controller: _quantity,
+                  ),
+                ],
+              ),
+            ),
+          actions: [
+            TextButton(
+                onPressed: (){
+                  Navigator.pop(_context);
+                },
+                child: Text(S.of(context).close, style: TextStyle(color: Colors.red),)
+            ),
+            TextButton(
+                onPressed: ()async{
+                  try{
+                    Navigator.pop(_context);
+                    context.loaderOverlay.show();
+                    String? result = await harvest.update(double.parse(_quantity.text));
+                    context.loaderOverlay.hide();
+                    if(result!=null){
+                      Tools.showErrorMessage(context, result);
+                    }else{
+                      if(mounted)setState(() {});
+                    }
+                  }catch(e){
+                    Tools.showErrorMessage(context, e.toString());
+                    Tools.consoleLog("[NFCScanPage._showHarvestEditDialog.err] $e");
+                    context.loaderOverlay.hide();
+                  }
+                },
+                child: Text(S.of(context).save, style: TextStyle(color: Colors.blue),)
+            ),
+          ],
+        );
+      },
+    );
+  }
 
   @override
   Widget build(BuildContext context) {
     double width = MediaQuery.of(context).size.width;
-    List<HTask> tasks = context.watch<HarvestModel>().tasks;
+    tasks = context.watch<HarvestProvider>().tasks;
+
     return Scaffold(
-      key: _scaffoldKey,
       appBar: AppBar(
         title: Text(S.of(context).harvestTracking),
         backgroundColor: Color(primaryColor),
         centerTitle: true,
         actions: [
-          FlatButton(
+          TextButton(
             onPressed: ()=>Navigator.push(context,MaterialPageRoute(builder: (context)=>NFCSettingPage())),
-            shape: CircleBorder(),
-            materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
             child: Icon(Icons.settings,color: Colors.white,size: 30,),
-            minWidth: 0,
           )
         ],
       ),
@@ -337,18 +428,14 @@ class _NFCScanPageState extends State<NFCScanPage>{
             child: Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                FlatButton(
+                TextButton(
                   onPressed: (){
                     setState(() {
                       selectedDate = selectedDate.subtract(Duration(days: 1));
                     });
                     _refreshController.requestRefresh();
                   },
-                  shape: CircleBorder(),
-                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
                   child: Icon(Icons.arrow_back_ios_outlined,color: Colors.black,size: 30,),
-                  padding: EdgeInsets.all(4),
-                  minWidth: 0,
                 ),
                 SizedBox(width: 10,),
                 MaterialButton(
@@ -359,18 +446,14 @@ class _NFCScanPageState extends State<NFCScanPage>{
                   child: Text(selectedDate.toString().split(' ')[0],style: TextStyle(fontSize: 16,fontWeight: FontWeight.w500),),
                 ),
                 SizedBox(width: 10,),
-                FlatButton(
+                TextButton(
                   onPressed: (){
                     setState(() {
                       selectedDate = selectedDate.add(Duration(days: 1));
                     });
                     _refreshController.requestRefresh();
                   },
-                  shape: CircleBorder(),
-                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
                   child: Icon(Icons.arrow_forward_ios_outlined,color: Colors.black,size: 30,),
-                  padding: EdgeInsets.all(4),
-                  minWidth: 0,
                 ),
               ],
             ),
@@ -387,84 +470,7 @@ class _NFCScanPageState extends State<NFCScanPage>{
                       child: ListView(
                         children: [
                           for(var task in tasks)
-                            CupertinoPopoverButton(
-                              popoverBoxShadow: [
-                                BoxShadow(color: Colors.black54,blurRadius: 5.0)
-                              ],
-                              popoverWidth: 180,
-                              popoverBuild: (_context){
-                                return CupertinoPopoverMenuList(
-                                  children: <Widget>[
-                                    CupertinoPopoverMenuItem(
-                                      child: Padding(
-                                        padding: const EdgeInsets.all(4.0),
-                                        child: Row(
-                                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                          children: [
-                                            Text(S.of(context).editContainer, style: TextStyle(color: Colors.black87),),
-                                            Icon(Icons.edit, color: Colors.black87,),
-                                          ],
-                                        ),
-                                      ),
-                                      onTap: (){
-                                        Navigator.pop(_context);
-                                        showHarvestTaskDialog(task: task);
-                                        return true;
-                                      },
-                                    ),
-                                    CupertinoPopoverMenuItem(
-                                      child: Padding(
-                                        padding: const EdgeInsets.all(4.0),
-                                        child: Row(
-                                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                          children: [
-                                            Text(S.of(context).delete, style: TextStyle(color: Colors.red),),
-                                            Icon(Icons.delete, color: Colors.red,),
-                                          ],
-                                        ),
-                                      ),
-                                      onTap: (){
-                                        Navigator.pop(_context);
-                                        confirmDeleteTaskDialog(task);
-                                        return true;
-                                      },
-                                    )
-                                  ],
-                                );
-                              },
-                              onTap: (){
-                                return true;
-                              },
-                              child: Padding(
-                                padding: const EdgeInsets.all(4.0),
-                                child: MaterialButton(
-                                  onPressed: (){
-                                    setState(() {selectedTask = task;});
-                                    _startNfcRead();
-                                  },
-                                  minWidth: 0,
-                                  shape: CircleBorder(),
-                                  color: selectedTask==task?Color(primaryColor):Colors.white,
-                                  clipBehavior: Clip.antiAlias,
-                                  padding: EdgeInsets.zero,
-                                  materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                                  child: AspectRatio(
-                                    aspectRatio: 1.0,
-                                    child: Column(
-                                      mainAxisAlignment: MainAxisAlignment.center,
-                                      children: [
-                                        Text(task.field.name.substring(0, task.field.name.length>1?2:task.field.name.length).toUpperCase(),style: TextStyle(fontSize: 16,fontWeight: FontWeight.w500),),
-                                        Padding(
-                                          padding: const EdgeInsets.symmetric(horizontal: 2),
-                                          child: FittedBox(child: Text(task.field.crop,),),
-                                        ),
-                                        Text(task.container.name[0].toUpperCase(),style: TextStyle(fontSize: 16,fontWeight: FontWeight.w500),),
-                                      ],
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ),
+                            _harvestTaskItem(task)
                         ],
                         scrollDirection: Axis.horizontal,
                       ),
@@ -473,13 +479,13 @@ class _NFCScanPageState extends State<NFCScanPage>{
                 Padding(
                   padding: const EdgeInsets.symmetric(horizontal: 4),
                   child: MaterialButton(
-                    onPressed: (){showHarvestTaskDialog();},
+                    onPressed: ()=>_showHarvestTaskDialog(null),
                     height: 70,
                     minWidth: 0,
                     shape: CircleBorder(),
                     color: Color(primaryColor),
                     materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                    child: Icon(Icons.add),
+                    child: Icon(Icons.add, color: Colors.white,),
                   ),
                 )
               ],
@@ -504,19 +510,19 @@ class _NFCScanPageState extends State<NFCScanPage>{
                     children: [
                       Expanded(
                         flex: 1,
-                          child: Text(S.of(context).employee,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                          child: Text(S.of(context).employee,style: TextStyle(fontWeight: FontWeight.w500, color: Colors.white),textAlign: TextAlign.center,)
                       ),
                       Expanded(
                           flex: 1,
-                          child: Text(S.of(context).field,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                          child: Text(S.of(context).field,style: TextStyle(fontWeight: FontWeight.w500, color: Colors.white),textAlign: TextAlign.center,)
                       ),
                       Expanded(
                           flex: 1,
-                          child: Text(S.of(context).container,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                          child: Text(S.of(context).container,style: TextStyle(fontWeight: FontWeight.w500, color: Colors.white),textAlign: TextAlign.center,)
                       ),
                       Expanded(
                           flex: 1,
-                          child: Text(S.of(context).quantity,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                          child: Text(S.of(context).quantity,style: TextStyle(fontWeight: FontWeight.w500, color: Colors.white),textAlign: TextAlign.center,)
                       ),
                     ],
                   ),
@@ -532,24 +538,35 @@ class _NFCScanPageState extends State<NFCScanPage>{
                       children: [
                         for(var harvest in harvests)
                           Slidable(
-                            actionPane: SlidableDrawerActionPane(),
-                            actionExtentRatio: 0.15,
-                            secondaryActions: <Widget>[
-                              IconSlideAction(
-                                caption: S.of(context).delete,
-                                color: Colors.red,
-                                foregroundColor: Colors.white,
-                                iconWidget: Icon(Icons.delete,color: Colors.white,size: 20,),
-                                onTap: ()async{
-                                  setState(() {harvests.remove(harvest);});
-                                  context.read<HarvestModel>().deleteHarvest(harvest.id).then((value){
-                                    if(mounted && value!=null && (value is String))showMessage(value);
-                                  });
-                                },
-                              ),
-                            ],
+                            endActionPane: ActionPane(
+                              motion: ScrollMotion(),
+                              extentRatio: 0.3,
+                              children: [
+                                SlidableAction(
+                                  backgroundColor: Colors.blue,
+                                  foregroundColor: Colors.white,
+                                  icon: Icons.edit,
+                                  onPressed: (c)=>_showHarvestEditDialog(harvest),
+                                ),
+                                SlidableAction(
+                                  backgroundColor: Colors.red,
+                                  foregroundColor: Colors.white,
+                                  icon: Icons.delete,
+                                  onPressed: (c)async{
+                                    if(await Tools.confirmDeleting(context, S.of(context).deleteHarvestConfirm)){
+                                      setState(() {harvests.remove(harvest);});
+                                      context.read<HarvestProvider>().deleteHarvest(harvest.id).then((message){
+                                        if(mounted && message!=null && (message is String)){
+                                          Tools.showErrorMessage(context, message);
+                                        }
+                                      });
+                                    }
+                                  },
+                                )
+                              ],
+                            ),
                             child: Container(
-                              padding: EdgeInsets.symmetric(vertical: 6),
+                              padding: EdgeInsets.symmetric(vertical: 8),
                               decoration: BoxDecoration(
                                 color: Colors.white,
                                 border: Border(bottom: BorderSide(color: Colors.grey))
@@ -558,19 +575,19 @@ class _NFCScanPageState extends State<NFCScanPage>{
                                 children: [
                                   Expanded(
                                       flex: 1,
-                                      child: Text(harvest.user.name,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                                      child: Text('${harvest.user?.name}', style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
                                   ),
                                   Expanded(
                                       flex: 1,
-                                      child: Text(harvest.field.name,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                                      child: Text('${harvest.field?.name}', style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
                                   ),
                                   Expanded(
                                       flex: 1,
-                                      child: Text(harvest.container.name,style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                                      child: Text('${harvest.container?.name}', style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
                                   ),
                                   Expanded(
                                       flex: 1,
-                                      child: Text(harvest.quantity.toString(),style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
+                                      child: Text('${harvest.quantity}', style: TextStyle(fontWeight: FontWeight.w500),textAlign: TextAlign.center,)
                                   ),
                                 ],
                               ),
