@@ -1,11 +1,15 @@
-import 'package:facepunch/lang/l10n.dart';
-import 'package:facepunch/models/app_const.dart';
-import 'package:facepunch/models/company_model.dart';
-import 'package:facepunch/models/harvest_model.dart';
-import 'package:facepunch/widgets/popover/cool_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_colorpicker/flutter_colorpicker.dart';
 import 'package:provider/provider.dart';
+
+import '/lang/l10n.dart';
+import '/config/app_const.dart';
+import '/models/company_model.dart';
+import '/models/harvest_model.dart';
+import '/widgets/popover/cool_ui.dart';
+import '/widgets/utils.dart';
+import '/providers/company_provider.dart';
+import '/providers/harvest_provider.dart';
 
 class NFCSettingPage extends StatefulWidget{
 
@@ -24,37 +28,52 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
   Color lowColor = Colors.red;
   Color mediumColor = Colors.yellow;
   Color highColor = Colors.green;
-  CompanySettings companySettings;
+  CompanySettings? companySettings;
+  List<Field> fields = [];
+  List<HContainer> containers = [];
+  Field? selectedField;
+  HContainer? selectedContainer;
 
   @override
   void initState() {
-    companySettings = context.read<CompanyModel>().myCompanySettings;
-    if(companySettings!=null){
-      if(companySettings.lowValue!=null)_lowValue = TextEditingController(text: companySettings.lowValue);
-      if(companySettings.highValue!=null)_highValue = TextEditingController(text: companySettings.highValue);
-      if(companySettings.lowColor!=null)lowColor = Color(int.parse('0x${companySettings.lowColor}'));
-      if(companySettings.mediumColor!=null)mediumColor = Color(int.parse('0x${companySettings.mediumColor}'));
-      if(companySettings.highColor!=null)highColor = Color(int.parse('0x${companySettings.highColor}'));
-      if(companySettings.reportTime!=null)_reportTime = TimeOfDay.fromDateTime(DateTime.parse(companySettings.reportTime));
-      if(companySettings.lastUpdated!=null)_lastUpdated = DateTime.parse(companySettings.lastUpdated);
-    }
+    companySettings = context.read<CompanyProvider>().myCompanySettings;
+    _init();
     super.initState();
   }
 
-  showFieldDialog({Field field}){
+  _init(){
+    try{
+      if(companySettings != null){
+        if(companySettings?.lowValue != null)_lowValue = TextEditingController(text: companySettings?.lowValue);
+        if(companySettings?.highValue != null)_highValue = TextEditingController(text: companySettings?.highValue);
+        if(companySettings?.lowColor != null)lowColor = Color(int.parse('0x${companySettings?.lowColor}'));
+        if(companySettings?.mediumColor != null)mediumColor = Color(int.parse('0x${companySettings?.mediumColor}'));
+        if(companySettings?.highColor != null)highColor = Color(int.parse('0x${companySettings?.highColor}'));
+        if(companySettings?.reportTime != null)_reportTime = TimeOfDay.fromDateTime(DateTime.parse(companySettings!.reportTime!));
+        if(companySettings?.lastUpdated != null)_lastUpdated = DateTime.parse(companySettings!.lastUpdated!);
+      }
+    }catch(e){
+      Tools.consoleLog('[NFCSettingPage._init]$e');
+    }
+  }
+
+  showFieldDialog(Field? field){
     TextEditingController _fieldName = TextEditingController(text: field?.name);
     TextEditingController _fieldCrop = TextEditingController(text: field?.crop);
     TextEditingController _fieldCropVariety = TextEditingController(text: field?.cropVariety);
     bool isSavingField = false;
-    String _fieldNameError, _fieldCropError,_fieldVarietyError;
+    String? _fieldNameError, _fieldCropError,_fieldVarietyError;
     showDialog(
       context: context,
       barrierDismissible: false,
       builder: (_context){
         return AlertDialog(
           title: Text(field==null?S.of(context).createNewField:S.of(context).updateField,textAlign: TextAlign.center,),
-          content:StatefulBuilder(
-            builder: (_,setState)=>Container(
+          insetPadding: EdgeInsets.zero,
+          contentPadding: EdgeInsets.all(8),
+          content: StatefulBuilder(
+            builder: (_, setState)=>Container(
+              width: MediaQuery.of(context).size.width - 40,
               child: Column(
                 mainAxisSize: MainAxisSize.min,
                 children: [
@@ -69,10 +88,10 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                       labelStyle: TextStyle(color: Colors.black54),
                       isDense: true,
                       contentPadding: EdgeInsets.zero,
-                      errorText: _fieldNameError
+                      errorText: _fieldNameError,
                     ),
+                    maxLength: 5,
                   ),
-                  SizedBox(height: 8,),
                   TextField(
                     controller: _fieldCrop,
                     textCapitalization: TextCapitalization.words,
@@ -86,8 +105,8 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                       contentPadding: EdgeInsets.zero,
                       errorText: _fieldCropError
                     ),
+                    maxLength: 12,
                   ),
-                  SizedBox(height: 8,),
                   TextField(
                     controller: _fieldCropVariety,
                     textCapitalization: TextCapitalization.words,
@@ -101,23 +120,28 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                       contentPadding: EdgeInsets.zero,
                       errorText: _fieldVarietyError
                     ),
+                    maxLength: 12,
                   ),
-                  SizedBox(height: 16,),
                   Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    mainAxisAlignment: MainAxisAlignment.end,
                     children: [
-                      RaisedButton(
-                        child: isSavingField
-                        ?SizedBox(
-                          height: 28,
-                          width: 28,
-                          child: CircularProgressIndicator(backgroundColor: Colors.white,),
-                        ):Padding(
-                          padding: const EdgeInsets.all(4.0),
-                          child: Text(S.of(context).save.toUpperCase(),style: TextStyle(fontSize: 18,fontWeight: FontWeight.bold,color: Colors.white),),
+                      TextButton(
+                        child: Text(S.of(context).close
+                          ,style: TextStyle(color: Colors.red),
                         ),
-                        color: Color(primaryColor),
-                        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
+                        onPressed: ()async{
+                          Navigator.pop(_context);
+                        },
+                      ),
+                      TextButton(
+                        child: isSavingField
+                            ?SizedBox(
+                          height: 20,
+                          width: 20,
+                          child: CircularProgressIndicator(strokeWidth: 2,),
+                        ):Text(S.of(context).save,
+                          style: TextStyle(color: Color(primaryColor)),
+                        ),
                         onPressed: ()async{
                           try{
                             _fieldVarietyError = null; _fieldCropError = null; _fieldNameError = null;
@@ -134,15 +158,15 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                               return;
                             }
                             setState((){isSavingField=true;});
-                            if(field!=null){
-                              field.name = _fieldName.text;
-                              field.crop = _fieldCrop.text;
-                              field.cropVariety = _fieldCropVariety.text;
+                            if(field != null){
+                              field!.name = _fieldName.text;
+                              field!.crop = _fieldCrop.text;
+                              field!.cropVariety = _fieldCropVariety.text;
                             }else{
                               field = Field(name: _fieldName.text,crop: _fieldCrop.text,cropVariety: _fieldCropVariety.text);
                             }
-                            String result = await context.read<HarvestModel>().createOrUpdateField(field);
-                            if(result!=null){
+                            String? result = await context.read<HarvestProvider>().createOrUpdateField(field!);
+                            if(result != null){
                               _fieldVarietyError = result;
                             }else{
                               Navigator.pop(_context);
@@ -152,17 +176,6 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                             Navigator.pop(_context);
                           }
                         },
-                      ),
-                      RaisedButton(
-                        child: Padding(
-                          padding: const EdgeInsets.all(4.0),
-                          child: Text(S.of(context).close.toUpperCase(),style: TextStyle(fontSize: 18,fontWeight: FontWeight.bold,color: Colors.white),),
-                        ),
-                        onPressed: ()async{
-                          Navigator.pop(_context);
-                        },
-                        color: Colors.red,
-                        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
                       ),
                     ],
                   )
@@ -175,88 +188,21 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
     );
   }
 
-  confirmDeleteFieldDialog(Field field){
-    bool isDeletingField = false;
-    showDialog(
-      context: context,
-      barrierDismissible: false,
-      builder: (_context){
-        return AlertDialog(
-            title: Text(S.of(context).deleteFieldConfirm,textAlign: TextAlign.center,),
-            content:StatefulBuilder(
-              builder: (_,setState)=>Container(
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        RaisedButton(
-                          child: isDeletingField
-                              ?SizedBox(
-                            height: 28,
-                            width: 28,
-                            child: CircularProgressIndicator(backgroundColor: Colors.white,),
-                          ):Padding(
-                            padding: const EdgeInsets.all(4.0),
-                            child: Text(S.of(context).delete,style: TextStyle(fontSize: 18,fontWeight: FontWeight.bold,color: Colors.white),),
-                          ),
-                          color: Color(primaryColor),
-                          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
-                          onPressed: ()async{
-                            setState((){isDeletingField = true;});
-                            String result = await context.read<HarvestModel>().deleteField(field);
-                            setState((){isDeletingField = false;});
-                            Navigator.pop(_context);
-                            if(result!=null)showMessage(result);
-                          },
-                        ),
-                        RaisedButton(
-                          child: Padding(
-                            padding: const EdgeInsets.all(4.0),
-                            child: Text(S.of(context).close.toUpperCase(),style: TextStyle(fontSize: 18,fontWeight: FontWeight.bold, color: Colors.white),),
-                          ),
-                          onPressed: ()async{
-                            Navigator.pop(_context);
-                          },
-                          color: Colors.red,
-                          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
-                        ),
-                      ],
-                    )
-                  ],
-                ),
-              ),
-            )
-        );
-      },
-    );
-  }
-
-  showMessage(String message){
-    _scaffoldKey.currentState.hideCurrentSnackBar();
-    _scaffoldKey.currentState.showSnackBar(
-        SnackBar(
-          content: Text(message),
-          backgroundColor: Colors.red,
-          duration: Duration(seconds: 2),
-          action: SnackBarAction(onPressed: (){},label: S.of(context).close,textColor: Colors.white,),
-        )
-    );
-  }
-
-  showContainerDialog({HContainer container}){
+  showContainerDialog(HContainer? container){
     TextEditingController _containerName = TextEditingController(text: container?.name);
     bool isSavingContainer = false;
-    String _containerNameError;
+    String? _containerNameError;
     showDialog(
       context: context,
       barrierDismissible: false,
       builder: (_context){
         return AlertDialog(
-            title: Text(container==null?S.of(context).createNewContainer:S.of(context).updateContainer,textAlign: TextAlign.center,),
+            title: Text(container==null?S.of(context).createNewContainer:S.of(context).updateContainer, textAlign: TextAlign.center,),
+            insetPadding: EdgeInsets.zero,
+            contentPadding: EdgeInsets.all(8),
             content:StatefulBuilder(
               builder: (_,setState)=>Container(
+                width: MediaQuery.of(context).size.width - 40,
                 child: Column(
                   mainAxisSize: MainAxisSize.min,
                   children: [
@@ -273,23 +219,24 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                           contentPadding: EdgeInsets.zero,
                           errorText: _containerNameError
                       ),
+                      maxLength: 5,
                     ),
-                    SizedBox(height: 16,),
                     Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      mainAxisAlignment: MainAxisAlignment.end,
                       children: [
-                        RaisedButton(
+                        TextButton(
+                          child: Text(S.of(context).close, style: TextStyle(color: Colors.red),),
+                          onPressed: ()async{
+                            Navigator.pop(_context);
+                          },
+                        ),
+                        TextButton(
                           child: isSavingContainer
-                          ?SizedBox(
-                            height: 28,
-                            width: 28,
-                            child: CircularProgressIndicator(backgroundColor: Colors.white,),
-                          ):Padding(
-                            padding: const EdgeInsets.all(4.0),
-                            child: Text(S.of(context).save.toUpperCase(),style: TextStyle(fontSize: 18,fontWeight: FontWeight.bold,color: Colors.white),),
-                          ),
-                          color: Color(primaryColor),
-                          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
+                              ?SizedBox(
+                            height: 20,
+                            width: 20,
+                            child: CircularProgressIndicator(strokeWidth: 2,),
+                          ):Text(S.of(context).save, style: TextStyle(color: Color(primaryColor)),),
                           onPressed: ()async{
                             try{
                               _containerNameError = null;
@@ -298,91 +245,23 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                                 return;
                               }
                               setState((){isSavingContainer=true;});
-                              if(container!=null){
-                                container.name = _containerName.text;
+                              if(container != null){
+                                container!.name = _containerName.text;
                               }else{
                                 container = HContainer(name: _containerName.text);
                               }
-                              String result = await context.read<HarvestModel>().createOrUpdateContainer(container);
-                              if(result!=null){
+                              String? result = await context.read<HarvestProvider>().createOrUpdateContainer(container!);
+                              if(result != null){
                                 _containerNameError = result;
                               }else{
                                 Navigator.pop(_context);
                               }
                               setState((){isSavingContainer=false;});
                             }catch(e){
+                              Tools.consoleLog('[showContainerDialog.onPressed]$e');
                               Navigator.pop(_context);
                             }
                           },
-                        ),
-                        RaisedButton(
-                          child: Padding(
-                            padding: const EdgeInsets.all(4.0),
-                            child: Text(S.of(context).close.toUpperCase(),style: TextStyle(fontSize: 18,fontWeight: FontWeight.bold,color: Colors.white),),
-                          ),
-                          onPressed: ()async{
-                            Navigator.pop(_context);
-                          },
-                          color: Colors.red,
-                          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
-                        ),
-                      ],
-                    )
-                  ],
-                ),
-              ),
-            )
-        );
-      },
-    );
-  }
-
-  confirmDeleteContainerDialog(HContainer container){
-    bool isDeletingContainer = false;
-    showDialog(
-      context: context,
-      barrierDismissible: false,
-      builder: (_context){
-        return AlertDialog(
-            title: Text(S.of(context).deleteContainerConfirm,textAlign: TextAlign.center,),
-            content:StatefulBuilder(
-              builder: (_,setState)=>Container(
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        RaisedButton(
-                          child: isDeletingContainer
-                              ?SizedBox(
-                            height: 28,
-                            width: 28,
-                            child: CircularProgressIndicator(backgroundColor: Colors.white,),
-                          ):Padding(
-                            padding: const EdgeInsets.all(4.0),
-                            child: Text(S.of(context).delete,style: TextStyle(fontSize: 18,fontWeight: FontWeight.bold,color: Colors.white),),
-                          ),
-                          color: Color(primaryColor),
-                          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
-                          onPressed: ()async{
-                            setState((){isDeletingContainer = true;});
-                            String result = await context.read<HarvestModel>().deleteContainer(container);
-                            setState((){isDeletingContainer = false;});
-                            Navigator.pop(_context);
-                            if(result!=null)showMessage(result);
-                          },
-                        ),
-                        RaisedButton(
-                          child: Padding(
-                            padding: const EdgeInsets.all(4.0),
-                            child: Text(S.of(context).close.toUpperCase(),style: TextStyle(fontSize: 18,fontWeight: FontWeight.bold, color: Colors.white),),
-                          ),
-                          onPressed: ()async{
-                            Navigator.pop(_context);
-                          },
-                          color: Colors.red,
-                          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
                         ),
                       ],
                     )
@@ -399,33 +278,36 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
     Color pickedColor = Color(color.value);
     await showDialog(
       context: context,
-      child: AlertDialog(
-        title: Text(S.of(context).chooseColor),
-        content: SingleChildScrollView(
-          child: ColorPicker(
-            pickerColor: pickedColor,
-            onColorChanged: (c){pickedColor = c;},
-            pickerAreaHeightPercent: 0.8,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: Text(S.of(context).chooseColor),
+          content: SingleChildScrollView(
+            child: ColorPicker(
+              pickerColor: pickedColor,
+              onColorChanged: (c){pickedColor = c;},
+              pickerAreaHeightPercent: 0.8,
+            ),
           ),
-        ),
-        actions: <Widget>[
-          FlatButton(
-            child: Text(S.of(context).ok),
-            onPressed: () {
-              color = pickedColor;
-              Navigator.of(context).pop();
-            },
-          ),
-        ],
-      ),
+          actions: <Widget>[
+            TextButton(
+              child: Text(S.of(context).ok),
+              onPressed: () {
+                color = pickedColor;
+                Navigator.of(context).pop();
+              },
+            ),
+          ],
+        );
+      },
     );
     return color;
   }
 
   changeReportTime()async{
-    final TimeOfDay picked = await showTimePicker(
+    final TimeOfDay? picked = await showTimePicker(
       context: context,
       initialTime: _reportTime,
+      initialEntryMode: TimePickerEntryMode.input
     );
     if(picked!=null){
       setState(() {
@@ -436,23 +318,23 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
 
   bool validator(){
     if(_lowValue.text.isEmpty){
-      showMessage(S.of(context).lowValueIsEmpty);
+      Tools.showErrorMessage(context, S.of(context).lowValueIsEmpty);
       return false;
     }
     if(double.tryParse(_lowValue.text)==null){
-      showMessage(S.of(context).lowValueShouldBeNumber);
+      Tools.showErrorMessage(context, S.of(context).lowValueShouldBeNumber);
       return false;
     }
     if(_highValue.text.isEmpty){
-      showMessage(S.of(context).highValueIsEmpty);
+      Tools.showErrorMessage(context, S.of(context).highValueIsEmpty);
       return false;
     }
-    if(double.tryParse(_highValue.text)==null){
-      showMessage(S.of(context).highValueShouldBeNumber);
+    if(double.tryParse(_highValue.text) ==null ){
+      Tools.showErrorMessage(context, S.of(context).highValueShouldBeNumber);
       return false;
     }
-    if(double.tryParse(_highValue.text)<double.tryParse(_lowValue.text)){
-      showMessage(S.of(context).highValueShouldBeBigger);
+    if(double.parse(_highValue.text) < double.parse(_lowValue.text)){
+      Tools.showErrorMessage(context, S.of(context).highValueShouldBeBigger);
       return false;
     }
     return true;
@@ -465,10 +347,217 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
     return v;
   }
 
+  Widget _fieldItem(Field field){
+    Widget child = Container(
+      height: 70,
+      width: 70,
+      decoration: BoxDecoration(
+          shape: BoxShape.circle,
+          color: Colors.white,
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withOpacity(0.2),
+              spreadRadius: 1,
+              blurRadius: 1,
+
+            )
+          ]
+      ),
+      clipBehavior: Clip.hardEdge,
+      padding: EdgeInsets.all(2),
+      margin: EdgeInsets.all(2),
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(field.shortName(),style: TextStyle(fontSize: 20,fontWeight: FontWeight.bold),),
+          FittedBox(
+              child: Text('${field.crop}', style: TextStyle(fontSize: 12),)
+          ),
+          FittedBox(
+              child: Text('${field.cropVariety}', style: TextStyle(fontSize: 12),)
+          )
+        ],
+      ),
+    );
+    if(selectedField == field){
+      return Stack(
+        alignment: Alignment.center,
+        children: [
+          child,
+          CircularProgressIndicator()
+        ],
+      );
+    }
+    return CupertinoPopoverButton(
+      popoverBoxShadow: [
+        BoxShadow(color: Colors.black54,blurRadius: 5.0)
+      ],
+      popoverWidth: 180,
+      popoverBuild: (_context){
+        return CupertinoPopoverMenuList(
+          children: <Widget>[
+            CupertinoPopoverMenuItem(
+              child: Padding(
+                padding: const EdgeInsets.all(4.0),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Text(S.of(context).editField,style: TextStyle(color: Colors.black87),),
+                    Icon(Icons.edit,color: Colors.black87,),
+                  ],
+                ),
+              ),
+              onTap: (){
+                Navigator.pop(_context);
+                showFieldDialog(field);
+                return true;
+              },
+            ),
+            CupertinoPopoverMenuItem(
+              child: Padding(
+                padding: const EdgeInsets.all(4.0),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Text(S.of(context).delete,style: TextStyle(color: Colors.red),),
+                    Icon(Icons.delete,color: Colors.red,),
+                  ],
+                ),
+              ),
+              onTap: (){
+                Navigator.pop(_context);
+                _deleteField(field);
+                return true;
+              },
+            )
+          ],
+        );
+      },
+      onTap: (){
+        return true;
+      },
+      child: child,
+    );
+  }
+
+  _deleteField(Field field)async{
+    if(await Tools.confirmDeleting(context, S.of(context).deleteFieldConfirm)){
+      setState(() { selectedField = field; });
+      String? result = await context.read<HarvestProvider>().deleteField(field);
+      if(result != null) Tools.showErrorMessage(context, result);
+      if(mounted)setState(() {
+        selectedField = null;
+      });
+    }
+  }
+
+  Widget _containerItem(HContainer container){
+    Widget child = Container(
+      height: 70,
+      width: 70,
+      decoration: BoxDecoration(
+          shape: BoxShape.circle,
+          color: Colors.white,
+          boxShadow: [
+            BoxShadow(
+              color: Colors.black.withOpacity(0.2),
+              spreadRadius: 1,
+              blurRadius: 1,
+
+            )
+          ]
+      ),
+      clipBehavior: Clip.hardEdge,
+      padding: EdgeInsets.all(2),
+      margin: EdgeInsets.all(2),
+      alignment: Alignment.center,
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(container.shortName(),style: TextStyle(fontSize: 20,fontWeight: FontWeight.bold),),
+          FittedBox(
+              child: Text('${container.name}',style: TextStyle(fontSize: 12),)
+          ),
+        ],
+      ),
+    );
+    if(selectedContainer == container){
+      return Stack(
+        alignment: Alignment.center,
+        children: [
+          child,
+          CircularProgressIndicator()
+        ],
+      );
+    }
+    return CupertinoPopoverButton(
+      popoverBoxShadow: [
+        BoxShadow(color: Colors.black54,blurRadius: 5.0)
+      ],
+      popoverWidth: 180,
+      popoverBuild: (_context){
+        return CupertinoPopoverMenuList(
+          children: <Widget>[
+            CupertinoPopoverMenuItem(
+              child: Padding(
+                padding: const EdgeInsets.all(4.0),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Text(S.of(context).editContainer,style: TextStyle(color: Colors.black87),),
+                    Icon(Icons.edit,color: Colors.black87,),
+                  ],
+                ),
+              ),
+              onTap: (){
+                Navigator.pop(_context);
+                showContainerDialog(container);
+                return true;
+              },
+            ),
+            CupertinoPopoverMenuItem(
+              child: Padding(
+                padding: const EdgeInsets.all(4.0),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Text(S.of(context).delete,style: TextStyle(color: Colors.red),),
+                    Icon(Icons.delete,color: Colors.red,),
+                  ],
+                ),
+              ),
+              onTap: (){
+                Navigator.pop(_context);
+                _deleteContainer(container);
+                return true;
+              },
+            )
+          ],
+        );
+      },
+      onTap: (){
+        return true;
+      },
+      child: child,
+    );
+  }
+
+  _deleteContainer(HContainer container)async{
+    if(await Tools.confirmDeleting(context, S.of(context).deleteContainerConfirm)){
+      setState(() { selectedContainer = container; });
+      String? result = await context.read<HarvestProvider>().deleteContainer(container);
+      if(result != null) Tools.showErrorMessage(context, result);
+      if(mounted)setState(() {
+        selectedContainer = null;
+      });
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
-    List<Field> fields = context.watch<HarvestModel>().fields;
-    List<HContainer> containers = context.watch<HarvestModel>().containers;
+    fields = context.watch<HarvestProvider>().fields;
+    containers = context.watch<HarvestProvider>().containers;
+
     return Scaffold(
       key: _scaffoldKey,
       appBar: AppBar(
@@ -488,86 +577,7 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                   crossAxisAlignment: WrapCrossAlignment.center,
                   children: [
                     for(var field in fields)
-                      CupertinoPopoverButton(
-                        popoverBoxShadow: [
-                          BoxShadow(color: Colors.black54,blurRadius: 5.0)
-                        ],
-                        popoverWidth: 180,
-                        popoverBuild: (_context){
-                          return CupertinoPopoverMenuList(
-                            children: <Widget>[
-                              CupertinoPopoverMenuItem(
-                                child: Padding(
-                                  padding: const EdgeInsets.all(4.0),
-                                  child: Row(
-                                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                    children: [
-                                      Text(S.of(context).editField,style: TextStyle(color: Colors.black87),),
-                                      Icon(Icons.edit,color: Colors.black87,),
-                                    ],
-                                  ),
-                                ),
-                                onTap: (){
-                                  Navigator.pop(_context);
-                                  showFieldDialog(field: field);
-                                  return true;
-                                },
-                              ),
-                              CupertinoPopoverMenuItem(
-                                child: Padding(
-                                  padding: const EdgeInsets.all(4.0),
-                                  child: Row(
-                                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                    children: [
-                                      Text(S.of(context).delete,style: TextStyle(color: Colors.red),),
-                                      Icon(Icons.delete,color: Colors.red,),
-                                    ],
-                                  ),
-                                ),
-                                onTap: (){
-                                  Navigator.pop(_context);
-                                  confirmDeleteFieldDialog(field);
-                                  return true;
-                                },
-                              )
-                            ],
-                          );
-                        },
-                        onTap: (){
-                          return true;
-                        },
-                        child: Container(
-                          height: 70,
-                          width: 70,
-                          decoration: BoxDecoration(
-                              shape: BoxShape.circle,
-                              color: Color(primaryColor),
-                              boxShadow: [
-                                BoxShadow(
-                                  color: Colors.black.withOpacity(0.2),
-                                  spreadRadius: 1,
-                                  blurRadius: 1,
-                                  
-                                )
-                              ]
-                          ),
-                          clipBehavior: Clip.hardEdge,
-                          padding: EdgeInsets.all(2),
-                          margin: EdgeInsets.all(2),
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: [
-                              Text(field.name.substring(0, field.name.length>1?2:field.name.length).toUpperCase(),style: TextStyle(fontSize: 20,fontWeight: FontWeight.bold),),
-                              FittedBox(
-                                  child: Text(field.crop, style: TextStyle(fontSize: 12),)
-                              ),
-                              FittedBox(
-                                  child: Text(field.cropVariety, style: TextStyle(fontSize: 12),)
-                              )
-                            ],
-                          ),
-                        ),
-                      ),
+                      _fieldItem(field),
                     Container(
                       height: 70,
                       width: 70,
@@ -587,9 +597,9 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                       padding: EdgeInsets.all(2),
                       margin: EdgeInsets.all(2),
                       child: InkWell(
-                        onTap: (){showFieldDialog();},
+                        onTap: (){showFieldDialog(null);},
                           borderRadius: BorderRadius.circular(30),
-                          child: Icon(Icons.add,size: 30,)
+                          child: Icon(Icons.add, size: 30, color: Colors.white,)
                       ),
                     ),
                   ],
@@ -603,84 +613,7 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                   crossAxisAlignment: WrapCrossAlignment.center,
                   children: [
                     for(var container in containers)
-                      CupertinoPopoverButton(
-                        popoverBoxShadow: [
-                          BoxShadow(color: Colors.black54,blurRadius: 5.0)
-                        ],
-                        popoverWidth: 180,
-                        popoverBuild: (_context){
-                          return CupertinoPopoverMenuList(
-                            children: <Widget>[
-                              CupertinoPopoverMenuItem(
-                                child: Padding(
-                                  padding: const EdgeInsets.all(4.0),
-                                  child: Row(
-                                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                    children: [
-                                      Text(S.of(context).editContainer,style: TextStyle(color: Colors.black87),),
-                                      Icon(Icons.edit,color: Colors.black87,),
-                                    ],
-                                  ),
-                                ),
-                                onTap: (){
-                                  Navigator.pop(_context);
-                                  showContainerDialog(container: container);
-                                  return true;
-                                },
-                              ),
-                              CupertinoPopoverMenuItem(
-                                child: Padding(
-                                  padding: const EdgeInsets.all(4.0),
-                                  child: Row(
-                                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                                    children: [
-                                      Text(S.of(context).delete,style: TextStyle(color: Colors.red),),
-                                      Icon(Icons.delete,color: Colors.red,),
-                                    ],
-                                  ),
-                                ),
-                                onTap: (){
-                                  Navigator.pop(_context);
-                                  confirmDeleteContainerDialog(container);
-                                  return true;
-                                },
-                              )
-                            ],
-                          );
-                        },
-                        onTap: (){
-                          return true;
-                        },
-                        child: Container(
-                          height: 70,
-                          width: 70,
-                          decoration: BoxDecoration(
-                              shape: BoxShape.circle,
-                              color: Color(primaryColor),
-                              boxShadow: [
-                                BoxShadow(
-                                  color: Colors.black.withOpacity(0.2),
-                                  spreadRadius: 1,
-                                  blurRadius: 1,
-                                  
-                                )
-                              ]
-                          ),
-                          clipBehavior: Clip.hardEdge,
-                          padding: EdgeInsets.all(2),
-                          margin: EdgeInsets.all(2),
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: [
-                              Text(container.name[0].toUpperCase(),style: TextStyle(fontSize: 20,fontWeight: FontWeight.bold),),
-                              FittedBox(
-                                  child: Text('${container.name}',style: TextStyle(fontSize: 12),)
-                              ),
-                              SizedBox(height: 12,),
-                            ],
-                          ),
-                        ),
-                      ),
+                      _containerItem(container),
                     Container(
                       height: 70,
                       width: 70,
@@ -700,9 +633,9 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                       padding: EdgeInsets.all(2),
                       margin: EdgeInsets.all(2),
                       child: InkWell(
-                          onTap: (){showContainerDialog();},
+                          onTap: (){ showContainerDialog(null); },
                           borderRadius: BorderRadius.circular(30),
-                          child: Icon(Icons.add,size: 30,)
+                          child: Icon(Icons.add,size: 30,color: Colors.white,)
                       ),
                     ),
                   ],
@@ -868,13 +801,13 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
               SizedBox(height: 8,),
               Padding(
                 padding: const EdgeInsets.all(8.0),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                child: Column(
                   children: [
                     Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
                       children: [
-                        Text('${S.of(context).reportTime}:\n${_reportTime==null?'':_reportTime.format(context)}',style: TextStyle(fontWeight: FontWeight.w500,fontSize: 16),),
-                        SizedBox(width: 8,),
+                        Text('${S.of(context).reportTime}: ${_reportTime.format(context)}', style: TextStyle(fontWeight: FontWeight.w500,fontSize: 16),),
+                        SizedBox(width: 16,),
                         InkWell(
                           child: Padding(
                             padding: const EdgeInsets.all(4.0),
@@ -885,51 +818,52 @@ class _NFCSettingPageState extends State<NFCSettingPage>{
                         )
                       ],
                     ),
-                    Text('${S.of(context).lastUpdated}: ${_lastUpdated==null?'':_lastUpdated.toString().split(' ')[0]}',style: TextStyle(fontWeight: FontWeight.w500,fontSize: 16),),
+                    Text('${S.of(context).lastUpdated}: ${_lastUpdated.toString().split(' ')[0]}',style: TextStyle(fontWeight: FontWeight.w500,fontSize: 16),),
                   ],
                 ),
               ),
               SizedBox(height: 8,),
               Center(
-                child: ButtonTheme(
+                child: MaterialButton(
                   minWidth: MediaQuery.of(context).size.width-60,
                   padding: EdgeInsets.all(8),
                   splashColor: Color(primaryColor),
-                  child: RaisedButton(
-                    child: isLoading
-                        ?SizedBox(
-                      height: 28,
-                      width: 28,
-                      child: CircularProgressIndicator(),
-                    )
-                        :Padding(
-                      padding: const EdgeInsets.all(4.0),
-                      child: Text(S.of(context).save,style: TextStyle(fontSize: 18,fontWeight: FontWeight.bold,color: Colors.white),),
-                    ),
-                    onPressed: ()async{
-                      try{
-                        if(validator()){
-                          final now = DateTime.now();
-                          companySettings.reportTime = DateTime(now.year, now.month, now.day, _reportTime.hour, _reportTime.minute).toString();
-                          companySettings.lastUpdated = now.toString().split(' ')[0];
-                          companySettings.highColor = highColor.value.toRadixString(16);
-                          companySettings.mediumColor = mediumColor.value.toRadixString(16);
-                          companySettings.lowColor = lowColor.value.toRadixString(16);
-                          companySettings.highValue = _highValue.text;
-                          companySettings.lowValue = _lowValue.text;
-                          setState(() {isLoading=true;});
-                          String result = await context.read<CompanyModel>().updateCompanySetting(companySettings);
-                          setState(() {isLoading=false;});
-                          if(result!=null)showMessage(result);
+                  color: Colors.black87,
+                  shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
+                  onPressed: ()async{
+                    try{
+                      if(validator()){
+                        final now = DateTime.now();
+                        if(companySettings != null){
+                          companySettings!.reportTime = DateTime(now.year, now.month, now.day, _reportTime.hour, _reportTime.minute).toString();
+                          companySettings!.lastUpdated = now.toString().split(' ')[0];
+                          companySettings!.highColor = highColor.value.toRadixString(16);
+                          companySettings!.mediumColor = mediumColor.value.toRadixString(16);
+                          companySettings!.lowColor = lowColor.value.toRadixString(16);
+                          companySettings!.highValue = _highValue.text;
+                          companySettings!.lowValue = _lowValue.text;
+                          setState(() { isLoading=true; });
+                          String? result = await context.read<CompanyProvider>().updateCompanySetting(companySettings!);
+                          setState(() { isLoading=false; });
+                          if(result != null){
+                            Tools.showErrorMessage(context, result);
+                          }
                         }
-                      }catch(e){
-                        print('[NFCSettingPage.onSave] $e');
-                        showMessage(e.toString());
                       }
-                    },
-                    color: Colors.black87,
-                    shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(30)),
-                  ),
+                    }catch(e){
+                      Tools.consoleLog('[NFCSettingPage.onSave] $e');
+                      Tools.showErrorMessage(context, e.toString());
+                    }
+                  },
+                  child: isLoading
+                      ?SizedBox(
+                        height: 28,
+                        width: 28,
+                        child: CircularProgressIndicator(),
+                      ):Padding(
+                        padding: const EdgeInsets.all(4.0),
+                        child: Text(S.of(context).save, style: TextStyle(fontSize: 18,fontWeight: FontWeight.bold,color: Colors.white),),
+                      ),
                 ),
               ),
             ],
