@@ -1,10 +1,12 @@
-import 'package:facepunch/lang/l10n.dart';
+import 'package:facepunch/config/app_const.dart';
+
+import '/lang/l10n.dart';
 import 'package:flutter/material.dart';
 import 'package:syncfusion_flutter_pdfviewer/pdfviewer.dart';
 
 class PDFFullScreen extends StatefulWidget{
   final String url;
-  PDFFullScreen({this.url});
+  PDFFullScreen({required this.url});
 
   @override
   _PDFFullScreenState createState() => _PDFFullScreenState();
@@ -12,7 +14,7 @@ class PDFFullScreen extends StatefulWidget{
 
 class _PDFFullScreenState extends State<PDFFullScreen> {
 
-  String pdfError;
+  String? pdfError;
 
 
   @override
@@ -21,6 +23,8 @@ class _PDFFullScreenState extends State<PDFFullScreen> {
       appBar: AppBar(
         title: Text(S.of(context).timeSheet),
         centerTitle: true,
+        backgroundColor: Color(primaryColor),
+        elevation: 0,
       ),
       body: pdfError!=null?Center(
         child: Text(S.of(context).pdfNotGenerated),
