@@ -12,40 +12,34 @@ typedef StringCallback(String data);
 
 class AutoCompleteTextField<T> extends StatefulWidget {
   final List<T> suggestions;
-  final Filter<T> itemFilter;
-  final Comparator<T> itemSorter;
-  final StringCallback textChanged, textSubmitted;
-  final ValueSetter<bool> onFocusChanged;
-  final InputEventCallback<T> itemSubmitted;
-  final AutoCompleteOverlayItemBuilder<T> itemBuilder;
+  final Filter<T>? itemFilter;
+  final Comparator<T>? itemSorter;
+  final StringCallback? textChanged, textSubmitted;
+  final ValueSetter<bool>? onFocusChanged;
+  final InputEventCallback<T>? itemSubmitted;
+  final AutoCompleteOverlayItemBuilder<T>? itemBuilder;
   final int suggestionsAmount;
   final GlobalKey<AutoCompleteTextFieldState<T>> key;
   final bool submitOnSuggestionTap, clearOnSubmit;
-  final List<TextInputFormatter> inputFormatters;
+  final List<TextInputFormatter>? inputFormatters;
   final int minLength;
 
   final InputDecoration decoration;
-  final TextStyle style;
+  final TextStyle? style;
   final TextInputType keyboardType;
   final TextInputAction textInputAction;
   final TextCapitalization textCapitalization;
-  final TextEditingController controller;
-  final FocusNode focusNode;
-  final bool autoFocus;
+  final TextEditingController? controller;
+  final FocusNode? focusNode;
+  final bool? autoFocus;
 
   AutoCompleteTextField(
-      {@required
-      this.itemSubmitted, //Callback on item selected, this is the item selected of type <T>
-        @required
-        this.key, //GlobalKey used to enable addSuggestion etc
-        @required
-        this.suggestions, //Suggestions that will be displayed
-        @required
+      {required this.itemSubmitted, //Callback on item selected, this is the item selected of type <T>
+        required this.key, //GlobalKey used to enable addSuggestion etc
+        required this.suggestions, //Suggestions that will be displayed
         this.itemBuilder, //Callback to build each item, return a Widget
-        @required
-        this.itemSorter, //Callback to sort items in the form (a of type <T>, b of type <T>)
-        @required
-        this.itemFilter, //Callback to filter item: return true or false depending on input text
+        required this.itemSorter, //Callback to sort items in the form (a of type <T>, b of type <T>)
+        required this.itemFilter, //Callback to filter item: return true or false depending on input text
         this.inputFormatters,
         this.style,
         this.decoration: const InputDecoration(),
@@ -66,30 +60,30 @@ class AutoCompleteTextField<T> extends StatefulWidget {
       })
       : super(key: key);
 
-  void clear() => key.currentState.clear();
+  void clear() => key.currentState!.clear();
 
   void addSuggestion(T suggestion) =>
-      key.currentState.addSuggestion(suggestion);
+      key.currentState!.addSuggestion(suggestion);
 
   void removeSuggestion(T suggestion) =>
-      key.currentState.removeSuggestion(suggestion);
+      key.currentState!.removeSuggestion(suggestion);
 
   void updateSuggestions(List<T> suggestions) =>
-      key.currentState.updateSuggestions(suggestions);
+      key.currentState!.updateSuggestions(suggestions);
 
-  void triggerSubmitted() => key.currentState.triggerSubmitted();
+  void triggerSubmitted() => key.currentState!.triggerSubmitted();
 
   void updateDecoration(
-      {InputDecoration decoration,
-        List<TextInputFormatter> inputFormatters,
-        TextCapitalization textCapitalization,
-        TextStyle style,
-        TextInputType keyboardType,
-        TextInputAction textInputAction}) =>
-      key.currentState.updateDecoration(decoration, inputFormatters,
+      {InputDecoration? decoration,
+        List<TextInputFormatter>? inputFormatters,
+        TextCapitalization? textCapitalization,
+        TextStyle? style,
+        TextInputType? keyboardType,
+        TextInputAction? textInputAction}) =>
+      key.currentState!.updateDecoration(decoration, inputFormatters,
           textCapitalization, style, keyboardType, textInputAction);
 
-  TextField get textField => key.currentState.textField;
+  TextField get textField => key.currentState!.textField;
 
   @override
   State<StatefulWidget> createState() => new AutoCompleteTextFieldState<T>(
@@ -120,31 +114,32 @@ class AutoCompleteTextField<T> extends StatefulWidget {
 class AutoCompleteTextFieldState<T> extends State<AutoCompleteTextField> {
   final LayerLink _layerLink = LayerLink();
 
-  TextField textField;
+  late TextField textField;
   List<T> suggestions;
-  StringCallback textChanged, textSubmitted;
-  ValueSetter<bool> onFocusChanged;
-  InputEventCallback<T> itemSubmitted;
-  AutoCompleteOverlayItemBuilder<T> itemBuilder;
-  Comparator<T> itemSorter;
-  OverlayEntry listSuggestionsEntry;
-  List<T> filteredSuggestions;
-  Filter<T> itemFilter;
+  StringCallback? textChanged, textSubmitted;
+  ValueSetter<bool>? onFocusChanged;
+  InputEventCallback<T>? itemSubmitted;
+  AutoCompleteOverlayItemBuilder<T>? itemBuilder;
+  Comparator<T>? itemSorter;
+  OverlayEntry? listSuggestionsEntry;
+  List<T>? filteredSuggestions;
+  Filter<T>? itemFilter;
   int suggestionsAmount;
   int minLength;
   bool submitOnSuggestionTap, clearOnSubmit;
-  TextEditingController controller;
-  FocusNode focusNode;
+  TextEditingController? controller;
+  FocusNode? focusNode;
 
   String currentText = "";
 
   InputDecoration decoration;
-  List<TextInputFormatter> inputFormatters;
+  List<TextInputFormatter>? inputFormatters;
   TextCapitalization textCapitalization;
-  TextStyle style;
+  TextStyle? style;
   TextInputType keyboardType;
   TextInputAction textInputAction;
-  final bool autoFocus;
+  final bool? autoFocus;
+
   AutoCompleteTextFieldState(
       this.suggestions,
       this.textChanged,
@@ -182,7 +177,7 @@ class AutoCompleteTextFieldState<T> extends State<AutoCompleteTextField> {
         updateOverlay(newText);
 
         if (textChanged != null) {
-          textChanged(newText);
+          textChanged!(newText);
         }
       },
       onTap: () {
@@ -190,34 +185,34 @@ class AutoCompleteTextFieldState<T> extends State<AutoCompleteTextField> {
       },
       textAlignVertical: TextAlignVertical.center,
       onSubmitted: (submittedText) => triggerSubmitted(submittedText: submittedText),
-      autofocus: autoFocus,
+      autofocus: autoFocus??false,
     );
 
-    if (this.controller != null && this.controller.text != null) {
-      currentText = this.controller.text;
+    if (this.controller != null) {
+      currentText = this.controller!.text;
     }
 
-    textField.focusNode.addListener(() {
+    textField.focusNode?.addListener(() {
       if (onFocusChanged != null) {
-        onFocusChanged(textField.focusNode.hasFocus);
+        onFocusChanged!(textField.focusNode!.hasFocus);
       }
 
-      if (!textField.focusNode.hasFocus) {
+      if (!textField.focusNode!.hasFocus) {
         filteredSuggestions = [];
         updateOverlay();
-      } else if (!(currentText == "" || currentText == null)) {
+      } else if (!(currentText == "")) {
         updateOverlay(currentText);
       }
     });
   }
 
   void updateDecoration(
-      InputDecoration decoration,
-      List<TextInputFormatter> inputFormatters,
-      TextCapitalization textCapitalization,
-      TextStyle style,
-      TextInputType keyboardType,
-      TextInputAction textInputAction) {
+      InputDecoration? decoration,
+      List<TextInputFormatter>? inputFormatters,
+      TextCapitalization? textCapitalization,
+      TextStyle? style,
+      TextInputType? keyboardType,
+      TextInputAction? textInputAction) {
     if (decoration != null) {
       this.decoration = decoration;
     }
@@ -257,7 +252,7 @@ class AutoCompleteTextFieldState<T> extends State<AutoCompleteTextField> {
           updateOverlay(newText);
 
           if (textChanged != null) {
-            textChanged(newText);
+            textChanged!(newText);
           }
         },
         onTap: () {
@@ -271,8 +266,8 @@ class AutoCompleteTextFieldState<T> extends State<AutoCompleteTextField> {
 
   void triggerSubmitted({submittedText}) {
     submittedText == null
-        ? textSubmitted(currentText)
-        : textSubmitted(submittedText);
+        ? textSubmitted!(currentText)
+        : textSubmitted!(submittedText);
 
     if (clearOnSubmit) {
       clear();
@@ -280,7 +275,7 @@ class AutoCompleteTextFieldState<T> extends State<AutoCompleteTextField> {
   }
 
   void clear() {
-    textField.controller.clear();
+    textField.controller!.clear();
     currentText = "";
     updateOverlay();
   }
@@ -302,7 +297,7 @@ class AutoCompleteTextFieldState<T> extends State<AutoCompleteTextField> {
     updateOverlay(currentText);
   }
 
-  void updateOverlay([String query]) {
+  void updateOverlay([String? query]) {
     if (listSuggestionsEntry == null) {
       final Size textFieldSize = (context.findRenderObject() as RenderBox).size;
       final width = textFieldSize.width;
@@ -318,25 +313,25 @@ class AutoCompleteTextFieldState<T> extends State<AutoCompleteTextField> {
                     width: width,
                     child: new Card(
                         child: new Column(
-                          children: filteredSuggestions.map((suggestion) {
+                          children: filteredSuggestions!.map((suggestion) {
                             return new Row(children: [
                               new Expanded(
                                   child: new InkWell(
-                                      child: itemBuilder(context, suggestion),
+                                      child: itemBuilder!(context, suggestion),
                                       onTap: () {
                                         setState(() {
                                           if (submitOnSuggestionTap) {
                                             String newText = suggestion.toString();
-                                            textField.controller.text = newText;
-                                            textField.focusNode.unfocus();
-                                            itemSubmitted(suggestion);
+                                            textField.controller!.text = newText;
+                                            textField.focusNode!.unfocus();
+                                            itemSubmitted!(suggestion);
                                             if (clearOnSubmit) {
                                               clear();
                                             }
                                           } else {
                                             String newText = suggestion.toString();
-                                            textField.controller.text = newText;
-                                            textChanged(newText);
+                                            textField.controller!.text = newText;
+                                            textChanged!(newText);
                                           }
                                         });
                                       }))
@@ -344,22 +339,22 @@ class AutoCompleteTextFieldState<T> extends State<AutoCompleteTextField> {
                           }).toList(),
                         )))));
       });
-      Overlay.of(context).insert(listSuggestionsEntry);
+      Overlay.of(context)!.insert(listSuggestionsEntry!);
     }
 
     filteredSuggestions = getSuggestions(
         suggestions, itemSorter, itemFilter, suggestionsAmount, query);
 
-    listSuggestionsEntry.markNeedsBuild();
+    listSuggestionsEntry!.markNeedsBuild();
   }
 
-  List<T> getSuggestions(List<T> suggestions, Comparator<T> sorter,
-      Filter<T> filter, int maxAmount, String query) {
+  List<T> getSuggestions(List<T> suggestions, Comparator<T>? sorter,
+      Filter<T>? filter, int maxAmount, String? query) {
     if (null == query || query.length < minLength) {
       return [];
     }
 
-    suggestions = suggestions.where((item) => filter(item, query)).toList();
+    suggestions = suggestions.where((item) => filter!(item, query)).toList();
     suggestions.sort(sorter);
     maxAmount = 20;
     if (suggestions.length > maxAmount) {
@@ -373,10 +368,10 @@ class AutoCompleteTextFieldState<T> extends State<AutoCompleteTextField> {
     // if we created our own focus node and controller, dispose of them
     // otherwise, let the caller dispose of their own instances
     if (focusNode == null) {
-      textField.focusNode.dispose();
+      textField.focusNode!.dispose();
     }
     if (controller == null) {
-      textField.controller.dispose();
+      textField.controller!.dispose();
     }
     super.dispose();
   }
@@ -388,14 +383,16 @@ class AutoCompleteTextFieldState<T> extends State<AutoCompleteTextField> {
 }
 
 class SimpleAutoCompleteTextField extends AutoCompleteTextField<String> {
-  final StringCallback textChanged, textSubmitted;
+  final StringCallback? textChanged, textSubmitted;
   final int minLength;
-  final ValueSetter<bool> onFocusChanged;
-  final TextEditingController controller;
-  final FocusNode focusNode;
+  final ValueSetter<bool>? onFocusChanged;
+  final TextEditingController? controller;
+  final FocusNode? focusNode;
   final bool autoFocus;
 
-  SimpleAutoCompleteTextField({TextStyle style, InputDecoration decoration: const InputDecoration(),
+  SimpleAutoCompleteTextField({
+    TextStyle? style,
+    InputDecoration decoration: const InputDecoration(),
     this.onFocusChanged,
     this.textChanged,
     this.textSubmitted,
@@ -404,8 +401,8 @@ class SimpleAutoCompleteTextField extends AutoCompleteTextField<String> {
     this.focusNode,
     this.autoFocus = false,
     TextInputType keyboardType: TextInputType.text,
-    @required GlobalKey<AutoCompleteTextFieldState<String>> key,
-    @required List<String> suggestions,
+    required GlobalKey<AutoCompleteTextFieldState<String>> key,
+    required List<String> suggestions,
     int suggestionsAmount: 5,
     bool submitOnSuggestionTap: true,
     bool clearOnSubmit: true,
