@@ -21,20 +21,20 @@ class MessageLookup extends MessageLookupByLibrary {
 
   final messages = _notInlinedMessages(_notInlinedMessages);
   static _notInlinedMessages(_) => <String, Function> {
-    "timeSheetSystemForEmployee" : MessageLookupByLibrary.simpleMessage("Timesheet system for Employee"),
+    "timeSheetSystemForEmployee" : MessageLookupByLibrary.simpleMessage("EMPLOYEE TIME MANAGEMENT SYSTEM"),
     "yourEmailIsRequired" : MessageLookupByLibrary.simpleMessage("Your email is required."),
-    "emailIsInvalid" : MessageLookupByLibrary.simpleMessage("Invalid email"),
+    "emailIsInvalid" : MessageLookupByLibrary.simpleMessage("Email is invalid."),
     "passwordIsRequired" : MessageLookupByLibrary.simpleMessage("Password is required."),
-    "adminSignIn" : MessageLookupByLibrary.simpleMessage("Admin Sign In"),
     "email" : MessageLookupByLibrary.simpleMessage("Email"),
     "enterYourEmailAddress" : MessageLookupByLibrary.simpleMessage("Enter your email address."),
     "password" : MessageLookupByLibrary.simpleMessage("Password"),
     "enterYourPassword" : MessageLookupByLibrary.simpleMessage("Enter your password."),
+    "passwordTooShort" : MessageLookupByLibrary.simpleMessage("Input 8 characters minimum."),
     "rememberMe" : MessageLookupByLibrary.simpleMessage("Remember me"),
-    "cannotLogin" : MessageLookupByLibrary.simpleMessage("Can't log in?"),
+    "cannotLogin" : MessageLookupByLibrary.simpleMessage("Can't log in"),
     "login" : MessageLookupByLibrary.simpleMessage("Log in"),
     "facePunch" : MessageLookupByLibrary.simpleMessage("Facepunch"),
-    "weWillSendNewPasswordToYourEmail" : MessageLookupByLibrary.simpleMessage("A new password has been sent to your email."),
+    "weWillSendNewPasswordToYourEmail" : MessageLookupByLibrary.simpleMessage("A new password has been sent to your email"),
     "done" : MessageLookupByLibrary.simpleMessage("Done"),
     "firstNameIsRequired" : MessageLookupByLibrary.simpleMessage("First Name is required."),
     "lastNameIsRequired" : MessageLookupByLibrary.simpleMessage("Last Name is required."),
@@ -52,33 +52,25 @@ class MessageLookup extends MessageLookupByLibrary {
     "postalCodeIsRequired" : MessageLookupByLibrary.simpleMessage("Postal Code is required."),
     "pleaseEnterYourCompanyInformation" : MessageLookupByLibrary.simpleMessage("Please enter your Company information"),
     "companyName" : MessageLookupByLibrary.simpleMessage("Company Name"),
-    "streetAddress" : MessageLookupByLibrary.simpleMessage("Street Address"),
+    "streetAddress" : MessageLookupByLibrary.simpleMessage("Address"),
     "aptSuiteBuilding" : MessageLookupByLibrary.simpleMessage("Apt, Suite, Building"),
     "postalCode" : MessageLookupByLibrary.simpleMessage("Postal Code"),
     "phoneNumber" : MessageLookupByLibrary.simpleMessage("Phone Number"),
     "website" : MessageLookupByLibrary.simpleMessage("Website"),
-    "next" : MessageLookupByLibrary.simpleMessage("Next"),
-    "pleaseEnterYourEmployeeRange" : MessageLookupByLibrary.simpleMessage("Please enter your Employee Range"),
-    "free" : MessageLookupByLibrary.simpleMessage("Free"),
     "employees" : MessageLookupByLibrary.simpleMessage("Employees"),
-    "theNumberMustBe6Digits" : MessageLookupByLibrary.simpleMessage("The number must contain 6 digits."),
-    "thankYouForRegisteringWithUs" : MessageLookupByLibrary.simpleMessage("Thank you for registering with us"),
-    "pleaseEnterThe6DigitsConfirmationNumberSentToYouByEmail" : MessageLookupByLibrary.simpleMessage("Please enter the 6 digits confirmation number send to you by email"),
-    "didNotGetAVerificationCode" : MessageLookupByLibrary.simpleMessage("Did not get a verification code?"),
     "close" : MessageLookupByLibrary.simpleMessage("Close"),
     "registration" : MessageLookupByLibrary.simpleMessage("Registration"),
-    "allowFacePunchToTakePictures" : MessageLookupByLibrary.simpleMessage("Allow Facepunch to take pictures."),
-    "thereIsNotAnyFaces" : MessageLookupByLibrary.simpleMessage("No face detected."),
+    "allowFacePunchToTakePictures" : MessageLookupByLibrary.simpleMessage("Allow device to use the camera"),
+    "thereIsNotAnyFaces" : MessageLookupByLibrary.simpleMessage("No face detected"),
     "locationPermissionDenied" : MessageLookupByLibrary.simpleMessage("Location permissions are permanently denied, we cannot request permissions."),
-    "pinCodeNotCorrect" : MessageLookupByLibrary.simpleMessage("PIN Code is not correct."),
+    "pinCodeNotCorrect" : MessageLookupByLibrary.simpleMessage("Invalid PIN code"),
     "welcome" : MessageLookupByLibrary.simpleMessage("Welcome"),
     "bye" : MessageLookupByLibrary.simpleMessage("Bye"),
-    "takePicture" : MessageLookupByLibrary.simpleMessage("Take Picture"),
-    "tryAgain" : MessageLookupByLibrary.simpleMessage("Try gain"),
-    "selectCompany" : MessageLookupByLibrary.simpleMessage("Please select your company."),
-    "startFacePunch" : MessageLookupByLibrary.simpleMessage("Start Facepunch"),
-    "addContainers" : MessageLookupByLibrary.simpleMessage("No container entered"),
-    "addFields" : MessageLookupByLibrary.simpleMessage("No field entrered"),
+    "takePicture" : MessageLookupByLibrary.simpleMessage("Take a Picture"),
+    "tryAgain" : MessageLookupByLibrary.simpleMessage("Try again"),
+    "startFacePunch" : MessageLookupByLibrary.simpleMessage("Start FACEPUNCH"),
+    "addContainers" : MessageLookupByLibrary.simpleMessage("No container found. Please add a container."),
+    "addFields" : MessageLookupByLibrary.simpleMessage("No field found. Please add a field."),
     "createNewTask" : MessageLookupByLibrary.simpleMessage("Create a new task"),
     "updateTask" : MessageLookupByLibrary.simpleMessage("Update Task"),
     "field" : MessageLookupByLibrary.simpleMessage("Field"),
@@ -94,7 +86,7 @@ class MessageLookup extends MessageLookupByLibrary {
     "updateField" : MessageLookupByLibrary.simpleMessage("Update Field"),
     "fieldName" : MessageLookupByLibrary.simpleMessage("Field Name"),
     "fieldCrop" : MessageLookupByLibrary.simpleMessage("Field Crop"),
-    "fieldCropVariety" : MessageLookupByLibrary.simpleMessage("Field Crop Variety"),
+    "fieldCropVariety" : MessageLookupByLibrary.simpleMessage("Crop Variety"),
     "nameIsRequired" : MessageLookupByLibrary.simpleMessage("Name is required."),
     "cropIsRequired" : MessageLookupByLibrary.simpleMessage("Crop is required."),
     "fieldNameIsRequired" : MessageLookupByLibrary.simpleMessage("Field Name is required."),
@@ -127,7 +119,6 @@ class MessageLookup extends MessageLookupByLibrary {
     "oldPassword" : MessageLookupByLibrary.simpleMessage("Old password"),
     "newPassword" : MessageLookupByLibrary.simpleMessage("New Password"),
     "company" : MessageLookupByLibrary.simpleMessage("Company"),
-    "companyPlan" : MessageLookupByLibrary.simpleMessage("Subscription plan"),
     "receiveRevisionNotification" : MessageLookupByLibrary.simpleMessage("Receive Revision Notification"),
     "receivePunchNotification" : MessageLookupByLibrary.simpleMessage("Receive Punch Notification"),
     "notifications" : MessageLookupByLibrary.simpleMessage("Notifications"),
@@ -167,9 +158,7 @@ class MessageLookup extends MessageLookupByLibrary {
     "calender" : MessageLookupByLibrary.simpleMessage("Calender"),
     "document" : MessageLookupByLibrary.simpleMessage("Document"),
     "setting" : MessageLookupByLibrary.simpleMessage("Setting"),
-    "chooseYourCompany" : MessageLookupByLibrary.simpleMessage("Choose Your Company"),
-    "faceScanLogin" : MessageLookupByLibrary.simpleMessage("FACE SCAN LOGIN"),
-    "employeeSignIn" : MessageLookupByLibrary.simpleMessage("Employee Sign In"),
+    "faceScanLogin" : MessageLookupByLibrary.simpleMessage("FACESCAN LOGIN"),
     "selectYourCompany" : MessageLookupByLibrary.simpleMessage("Select Your Company"),
     "language" : MessageLookupByLibrary.simpleMessage("Language"),
     "country" : MessageLookupByLibrary.simpleMessage("Country"),
@@ -193,6 +182,9 @@ class MessageLookup extends MessageLookupByLibrary {
     "employeeName" : MessageLookupByLibrary.simpleMessage("Employee Name"),
     "accept" : MessageLookupByLibrary.simpleMessage("Accept"),
     "decline" : MessageLookupByLibrary.simpleMessage("Decline"),
+    "punchIn" : MessageLookupByLibrary.simpleMessage("Punch In"),
+    "punchOut" : MessageLookupByLibrary.simpleMessage("Punch Out"),
+    "pin" : MessageLookupByLibrary.simpleMessage("Pin"),
     "harvestReport" : MessageLookupByLibrary.simpleMessage("Harvest Report"),
     "totalOfTheDay" : MessageLookupByLibrary.simpleMessage("Total of the day"),
     "totalOfTheSeason" : MessageLookupByLibrary.simpleMessage("Total of the season"),
@@ -200,7 +192,8 @@ class MessageLookup extends MessageLookupByLibrary {
     "searchEmployee" : MessageLookupByLibrary.simpleMessage("Search Employee"),
     "selectProject" : MessageLookupByLibrary.simpleMessage("Select a project"),
     "selectTask" : MessageLookupByLibrary.simpleMessage("Select a task"),
-    "youAreNowWorkingOn" : MessageLookupByLibrary.simpleMessage("You Are Now Working On"),
+    "youAreWorkingOn" : MessageLookupByLibrary.simpleMessage("You are working on"),
+    "youAreWorkingOnCall" : MessageLookupByLibrary.simpleMessage("You are working on a call"),
     "project" : MessageLookupByLibrary.simpleMessage("Project"),
     "activity" : MessageLookupByLibrary.simpleMessage("Activity"),
     "startTime" : MessageLookupByLibrary.simpleMessage("Start Time"),
@@ -208,6 +201,7 @@ class MessageLookup extends MessageLookupByLibrary {
     "dailySchedule" : MessageLookupByLibrary.simpleMessage("Daily Schedule"),
     "editWorkHistory" : MessageLookupByLibrary.simpleMessage("Edit Work History"),
     "start" : MessageLookupByLibrary.simpleMessage("Start"),
+    "resume" : MessageLookupByLibrary.simpleMessage("Resume"),
     "end" : MessageLookupByLibrary.simpleMessage("End"),
     "pressToStart" : MessageLookupByLibrary.simpleMessage("Press to start"),
     "pressToEnd" : MessageLookupByLibrary.simpleMessage("Press to end"),
@@ -218,8 +212,84 @@ class MessageLookup extends MessageLookupByLibrary {
     "dispatch" : MessageLookupByLibrary.simpleMessage("Dispatch"),
     "addSchedule" : MessageLookupByLibrary.simpleMessage("Add Schedule"),
     "type" : MessageLookupByLibrary.simpleMessage("Type"),
+    "priority" : MessageLookupByLibrary.simpleMessage("Priority"),
     "call" : MessageLookupByLibrary.simpleMessage("Call"),
     "shop" : MessageLookupByLibrary.simpleMessage("Shop"),
     "schedule" : MessageLookupByLibrary.simpleMessage("Schedule"),
+    "selectCall" : MessageLookupByLibrary.simpleMessage("Select Call"),
+    "startCall" : MessageLookupByLibrary.simpleMessage("Start Call"),
+    "task" : MessageLookupByLibrary.simpleMessage("Task"),
+    "note" : MessageLookupByLibrary.simpleMessage("Note"),
+    "selectSchedule" : MessageLookupByLibrary.simpleMessage("Select Schedule"),
+    "shift" : MessageLookupByLibrary.simpleMessage("Shift"),
+    "startSchedule" : MessageLookupByLibrary.simpleMessage("Start Schedule"),
+    "description" : MessageLookupByLibrary.simpleMessage("Description"),
+    "youMustWriteDescription" : MessageLookupByLibrary.simpleMessage("You must write description."),
+    "startWorking" : MessageLookupByLibrary.simpleMessage("Start Working"),
+    "breaks" : MessageLookupByLibrary.simpleMessage("Breaks"),
+    "incorrectBreakTime" : MessageLookupByLibrary.simpleMessage("Incorrect Break Time"),
+    "correctBreakTime" : MessageLookupByLibrary.simpleMessage("Correct Break Time"),
+    "length" : MessageLookupByLibrary.simpleMessage("Length"),
+    "breakLengthCanNotBeZero" : MessageLookupByLibrary.simpleMessage("Break length can't equal 0"),
+    "invalidBreakLength" : MessageLookupByLibrary.simpleMessage("Invalid Break Length"),
+    "dailyTasks" : MessageLookupByLibrary.simpleMessage("Daily Tasks"),
+    "schedules" : MessageLookupByLibrary.simpleMessage("Schedules"),
+    "calls" : MessageLookupByLibrary.simpleMessage("Calls"),
+    "incorrect" : MessageLookupByLibrary.simpleMessage("Incorrect"),
+    "correct" : MessageLookupByLibrary.simpleMessage("Correct"),
+    "selectPriority" : MessageLookupByLibrary.simpleMessage("Select Priority"),
+    "deletePunchConfirm" : MessageLookupByLibrary.simpleMessage("Are you sure you want to delete this PUNCH?"),
+    "deleteWorkConfirm" : MessageLookupByLibrary.simpleMessage("Are you sure you want to delete this work?"),
+    "deleteBreakConfirm" : MessageLookupByLibrary.simpleMessage("Are you sure you want to delete this break ?"),
+    "breakTime" : MessageLookupByLibrary.simpleMessage("Break Time"),
+    "newVersionAvailable" : MessageLookupByLibrary.simpleMessage("A new version is now available. Would you like to update to the new version?"),
+    "update" : MessageLookupByLibrary.simpleMessage("Update"),
+    "callRevision" : MessageLookupByLibrary.simpleMessage("Call Revision"),
+    "addCall" : MessageLookupByLibrary.simpleMessage("Add Call"),
+    "bugReport" : MessageLookupByLibrary.simpleMessage("Bug Report"),
+    "make" : MessageLookupByLibrary.simpleMessage("Make"),
+    "model" : MessageLookupByLibrary.simpleMessage("Model"),
+    "os" : MessageLookupByLibrary.simpleMessage("OS (Operating system)"),
+    "version" : MessageLookupByLibrary.simpleMessage("Version"),
+    "brand" : MessageLookupByLibrary.simpleMessage("Brand"),
+    "name" : MessageLookupByLibrary.simpleMessage("Name"),
+    "system" : MessageLookupByLibrary.simpleMessage("System"),
+    "bugReportDescription" : MessageLookupByLibrary.simpleMessage("Insert a description of your errors here..."),
+    "punchTime" : MessageLookupByLibrary.simpleMessage("PUNCH Time"),
+    "revisionHasBeenSent" : MessageLookupByLibrary.simpleMessage("A revision request has been sent."),
+    "sent" : MessageLookupByLibrary.simpleMessage("Sent"),
+    "accepted" : MessageLookupByLibrary.simpleMessage("Accepted"),
+    "declined" : MessageLookupByLibrary.simpleMessage("Decline"),
+    "tapToSubmitDescription" : MessageLookupByLibrary.simpleMessage("Tap to submit a description."),
+    "invalidNFC" : MessageLookupByLibrary.simpleMessage("Can not find identifier from the NFC."),
+    "notAllowedNFC" : MessageLookupByLibrary.simpleMessage("Allow NFC permission."),
+    "deleteHarvestConfirm" : MessageLookupByLibrary.simpleMessage("Are you sure you want to delete this harvest ?"),
+    "thankYouForReporting" : MessageLookupByLibrary.simpleMessage("Thank you for reporting a bug"),
+    "open" : MessageLookupByLibrary.simpleMessage("Open"),
+    "callDetail" : MessageLookupByLibrary.simpleMessage("Call Detail"),
+    "date" : MessageLookupByLibrary.simpleMessage("Date"),
+    "thisPunchHasBeenSentAlready" : MessageLookupByLibrary.simpleMessage("This PUNCH has been sent already."),
+    "canNotSendRevisionAfterStart" : MessageLookupByLibrary.simpleMessage("Unable to send review after start"),
+    "canNotEditDeleteCall" : MessageLookupByLibrary.simpleMessage("Can neither edit nor delete a worked CALL."),
+    "workingNow" : MessageLookupByLibrary.simpleMessage("Working Now"),
+    "works" : MessageLookupByLibrary.simpleMessage("Work"),
+    "workRevision" : MessageLookupByLibrary.simpleMessage("Work Revision"),
+    "revisions" : MessageLookupByLibrary.simpleMessage("Revisions"),
+    "deviceIdForPunch" : MessageLookupByLibrary.simpleMessage("Device ID for Punch"),
+    "clickToCopy" : MessageLookupByLibrary.simpleMessage("Click to copy"),
+    "deviceIdCopied" : MessageLookupByLibrary.simpleMessage("Device Id copied"),
+    "startManualBreak" : MessageLookupByLibrary.simpleMessage("Start manual break"),
+    "endManualBreak" : MessageLookupByLibrary.simpleMessage("End manual break"),
+    "somethingWentWrong" : MessageLookupByLibrary.simpleMessage("Something went wrong"),
+    "signIn" : MessageLookupByLibrary.simpleMessage("Sign In"),
+    "revisionDescription" : MessageLookupByLibrary.simpleMessage("Description Revision"),
+    "revisionDescriptionSubmitted" : MessageLookupByLibrary.simpleMessage("Description Revision Submitted"),
+    "editHarvest" : MessageLookupByLibrary.simpleMessage("Edit Harvest"),
+    "welcomeToFacePunch" : MessageLookupByLibrary.simpleMessage("Welcome to Facepunch"),
+    "theBestEmployeeClockingSystem" : MessageLookupByLibrary.simpleMessage("The Best Employee Clocking System"),
+    "employeePortal" : MessageLookupByLibrary.simpleMessage("Employee Portal"),
+    "team" : MessageLookupByLibrary.simpleMessage("Team"),
+    "my" : MessageLookupByLibrary.simpleMessage("My"),
+    "day" : MessageLookupByLibrary.simpleMessage("Day"),
   };
 }
