@@ -10,177 +10,273 @@ class MessageLookup extends MessageLookupByLibrary {
   final messages = _notInlinedMessages(_notInlinedMessages);
 
   static _notInlinedMessages(_) => <String, Function> {
-    "timeSheetSystemForEmployee" : MessageLookupByLibrary.simpleMessage("Système de rapport pour employé"),
-    "yourEmailIsRequired" : MessageLookupByLibrary.simpleMessage("Une adresse courriel est requise"),
+    "timeSheetSystemForEmployee" : MessageLookupByLibrary.simpleMessage("SYSTÈME DE GESTION DE TEMPS DES EMPLOYÉS"),
+    "yourEmailIsRequired" : MessageLookupByLibrary.simpleMessage("Votre courriel est requis"),
     "emailIsInvalid" : MessageLookupByLibrary.simpleMessage("Courriel invalide"),
-    "passwordIsRequired" : MessageLookupByLibrary.simpleMessage("Un mot de passe est requis"),
-    "adminSignIn" : MessageLookupByLibrary.simpleMessage("Se Connecter"),
-    "email" : MessageLookupByLibrary.simpleMessage("Adresse courriel"),
-    "enterYourEmailAddress" : MessageLookupByLibrary.simpleMessage("Veuillez entrer votre adresse courriel"),
-    "password" : MessageLookupByLibrary.simpleMessage("Mot de passe"),
-    "enterYourPassword" : MessageLookupByLibrary.simpleMessage("Veuillez entrer votre mot de passe"),
-    "rememberMe" : MessageLookupByLibrary.simpleMessage("Se souvenir de moi"),
-    "cannotLogin" : MessageLookupByLibrary.simpleMessage("Impossible de se connecter ?"),
-    "login" : MessageLookupByLibrary.simpleMessage("Se connecter"),
+    "passwordIsRequired" : MessageLookupByLibrary.simpleMessage("Mot de passe requis"),
+    "email" : MessageLookupByLibrary.simpleMessage("Courriel"),
+    "enterYourEmailAddress" : MessageLookupByLibrary.simpleMessage("Entrez votre adresse courriel"),
+    "password" : MessageLookupByLibrary.simpleMessage("Mot de passe "),
+    "enterYourPassword" : MessageLookupByLibrary.simpleMessage("Entrez votre mot de passe"),
+    "passwordTooShort" : MessageLookupByLibrary.simpleMessage("Entrez 8 caractères minimum"),
+    "rememberMe" : MessageLookupByLibrary.simpleMessage("Mémoriser"),
+    "cannotLogin" : MessageLookupByLibrary.simpleMessage("Impossible de se connecter"),
+    "login" : MessageLookupByLibrary.simpleMessage("Ouverture de session"),
     "facePunch" : MessageLookupByLibrary.simpleMessage("Facepunch"),
     "weWillSendNewPasswordToYourEmail" : MessageLookupByLibrary.simpleMessage("Un nouveau mot de passe vous a été envoyé par courriel"),
     "done" : MessageLookupByLibrary.simpleMessage("Terminer"),
-    "firstNameIsRequired" : MessageLookupByLibrary.simpleMessage("Un prénom est requis"),
-    "lastNameIsRequired" : MessageLookupByLibrary.simpleMessage("Une nom de famille est requis"),
-    "signUp" : MessageLookupByLibrary.simpleMessage("Se connecter"),
+    "firstNameIsRequired" : MessageLookupByLibrary.simpleMessage("Votre prénom est requis"),
+    "lastNameIsRequired" : MessageLookupByLibrary.simpleMessage("Votre nom de famille est requis"),
+    "signUp" : MessageLookupByLibrary.simpleMessage("S'inscrire"),
     "firstName" : MessageLookupByLibrary.simpleMessage("Prénom"),
-    "enterYourFirstName" : MessageLookupByLibrary.simpleMessage("Veuillez entrer votre prénom"),
+    "enterYourFirstName" : MessageLookupByLibrary.simpleMessage("Entrez votre prénom"),
     "lastName" : MessageLookupByLibrary.simpleMessage("Nom de famille"),
-    "enterYourLastName" : MessageLookupByLibrary.simpleMessage("Veuillez entrer votre nom de famille"),
-    "register" : MessageLookupByLibrary.simpleMessage("S'enregistrer"),
-    "companyNameIsRequired" : MessageLookupByLibrary.simpleMessage("Un nom de compagnie est requis"),
-    "companyAddressIsRequired" : MessageLookupByLibrary.simpleMessage("Une adresse de compagnie est requise"),
-    "countryIsRequired" : MessageLookupByLibrary.simpleMessage("Un Pays/région est requis."),
-    "stateIsRequired" : MessageLookupByLibrary.simpleMessage("Un État/province/région est requis"),
-    "cityIsRequired" : MessageLookupByLibrary.simpleMessage("Une ville est requise"),
-    "postalCodeIsRequired" : MessageLookupByLibrary.simpleMessage("Un code postal est requis"),
-    "pleaseEnterYourCompanyInformation" : MessageLookupByLibrary.simpleMessage("Veuillez entrer vos informations de compagnie"),
-    "companyName" : MessageLookupByLibrary.simpleMessage("Nom de compagnie"),
+    "enterYourLastName" : MessageLookupByLibrary.simpleMessage("Entrez votre nom de famille"),
+    "register" : MessageLookupByLibrary.simpleMessage("Enregistrer"),
+    "companyNameIsRequired" : MessageLookupByLibrary.simpleMessage("Nom de la compagnie requis"),
+    "companyAddressIsRequired" : MessageLookupByLibrary.simpleMessage("Adresse de la compagnie requis"),
+    "countryIsRequired" : MessageLookupByLibrary.simpleMessage("La Province est requise"),
+    "stateIsRequired" : MessageLookupByLibrary.simpleMessage("Le Pays est requis"),
+    "cityIsRequired" : MessageLookupByLibrary.simpleMessage("La Ville est requise"),
+    "postalCodeIsRequired" : MessageLookupByLibrary.simpleMessage("Le Code postal est requis"),
+    "pleaseEnterYourCompanyInformation" : MessageLookupByLibrary.simpleMessage("Veuillez entrer les informations de la compagnie"),
+    "companyName" : MessageLookupByLibrary.simpleMessage("Nom de la compagnie"),
     "streetAddress" : MessageLookupByLibrary.simpleMessage("Adresse"),
-    "aptSuiteBuilding" : MessageLookupByLibrary.simpleMessage("App, suite, bâtiment"),
+    "aptSuiteBuilding" : MessageLookupByLibrary.simpleMessage("Adresse 2"),
     "postalCode" : MessageLookupByLibrary.simpleMessage("Code postal"),
-    "phoneNumber" : MessageLookupByLibrary.simpleMessage("Numéro de téléphone"),
-    "website" : MessageLookupByLibrary.simpleMessage("Site web"),
-    "next" : MessageLookupByLibrary.simpleMessage("Suivant"),
-    "pleaseEnterYourEmployeeRange" : MessageLookupByLibrary.simpleMessage("Veuillez sélectionner votre nombre d'employé"),
-    "free" : MessageLookupByLibrary.simpleMessage("Gratuit"),
-    "employees" : MessageLookupByLibrary.simpleMessage("Employés"),
-    "theNumberMustBe6Digits" : MessageLookupByLibrary.simpleMessage("Le numéro doit contenir 6 chiffres"),
-    "thankYouForRegisteringWithUs" : MessageLookupByLibrary.simpleMessage("Merci de vous être enregistrer avec nous !"),
-    "pleaseEnterThe6DigitsConfirmationNumberSentToYouByEmail" : MessageLookupByLibrary.simpleMessage("Veuillez entrer le numéro à 6 chiffres qui vous a été envoyépar courriel"),
-    "didNotGetAVerificationCode" : MessageLookupByLibrary.simpleMessage("Vous n'avez pas reçu de code de vérification?"),
+    "phoneNumber" : MessageLookupByLibrary.simpleMessage("Téléphone"),
+    "website" : MessageLookupByLibrary.simpleMessage("Site internet"),
+    "employees" : MessageLookupByLibrary.simpleMessage("Employé"),
     "close" : MessageLookupByLibrary.simpleMessage("Fermer"),
-    "registration" : MessageLookupByLibrary.simpleMessage("Enregistrement"),
-    "allowFacePunchToTakePictures" : MessageLookupByLibrary.simpleMessage("Permettre à FacePunch d'utiliser l'appareil photo"),
-    "thereIsNotAnyFaces" : MessageLookupByLibrary.simpleMessage("Aucun visage n'a pu être détecté"),
-    "locationPermissionDenied" : MessageLookupByLibrary.simpleMessage("Les autorisations de localisation sont désactivés, nous ne pouvons pas demander les autorisations"),
-    "pinCodeNotCorrect" : MessageLookupByLibrary.simpleMessage("Code PIN incorrect"),
+    "allowFacePunchToTakePictures" : MessageLookupByLibrary.simpleMessage("Autorisez l'appareil à utiliser la caméra"),
+    "thereIsNotAnyFaces" : MessageLookupByLibrary.simpleMessage("Aucun visage detecté"),
+    "locationPermissionDenied" : MessageLookupByLibrary.simpleMessage("Permission refusée"),
+    "pinCodeNotCorrect" : MessageLookupByLibrary.simpleMessage("Code PIN non valide"),
     "welcome" : MessageLookupByLibrary.simpleMessage("Bienvenue"),
-    "bye" : MessageLookupByLibrary.simpleMessage("Au revoir"),
+    "bye" : MessageLookupByLibrary.simpleMessage("Aurevoir"),
     "takePicture" : MessageLookupByLibrary.simpleMessage("Prendre une photo"),
-    "tryAgain" : MessageLookupByLibrary.simpleMessage("Veuillez essayer de nouveau"),
-    "selectCompany" : MessageLookupByLibrary.simpleMessage("Veuillez sélectionner votre compagnie"),
-    "startFacePunch" : MessageLookupByLibrary.simpleMessage("Commencer Facepunch"),
-    "addContainers" : MessageLookupByLibrary.simpleMessage("Aucun contenant enregistré"),
-    "addFields" : MessageLookupByLibrary.simpleMessage("Aucun champ enregistré"),
+    "tryAgain" : MessageLookupByLibrary.simpleMessage("Essayez à nouveau"),
+    "startFacePunch" : MessageLookupByLibrary.simpleMessage("Démarrer FACEPUNCH"),
+    "addContainers" : MessageLookupByLibrary.simpleMessage("Aucun contenant trouvé, veuillez ajouter un contenant."),
+    "addFields" : MessageLookupByLibrary.simpleMessage("Aucun champ trouvé, veuillez ajouter un champ"),
     "createNewTask" : MessageLookupByLibrary.simpleMessage("Créer une nouvelle tâche"),
-    "updateTask" : MessageLookupByLibrary.simpleMessage("Mettre à jour les tâches"),
-    "field" : MessageLookupByLibrary.simpleMessage("Champs"),
+    "updateTask" : MessageLookupByLibrary.simpleMessage("Mettre à jour la tâche"),
+    "field" : MessageLookupByLibrary.simpleMessage("Champ"),
     "container" : MessageLookupByLibrary.simpleMessage("Contenant"),
-    "save" : MessageLookupByLibrary.simpleMessage("Enregistrer"),
-    "deleteTaskConfirm" : MessageLookupByLibrary.simpleMessage("Êtes-vous certain de vouloir supprimer cette tâche ?"),
+    "save" : MessageLookupByLibrary.simpleMessage("Sauvegarder"),
+    "deleteTaskConfirm" : MessageLookupByLibrary.simpleMessage("Voulez-vous vraiment supprimer cette tâche?"),
     "delete" : MessageLookupByLibrary.simpleMessage("Supprimer"),
-    "harvestTracking" : MessageLookupByLibrary.simpleMessage("Suivi des récoltes"),
-    "editContainer" : MessageLookupByLibrary.simpleMessage("Modifier les contenants"),
+    "harvestTracking" : MessageLookupByLibrary.simpleMessage("Suivi des Récoltes"),
+    "editContainer" : MessageLookupByLibrary.simpleMessage("Modifier le conenant"),
     "employee" : MessageLookupByLibrary.simpleMessage("Employé"),
     "quantity" : MessageLookupByLibrary.simpleMessage("Quantité"),
-    "createNewField" : MessageLookupByLibrary.simpleMessage("Ajouter un nouveau champ"),
-    "updateField" : MessageLookupByLibrary.simpleMessage("Mettre à jour les champs"),
+    "createNewField" : MessageLookupByLibrary.simpleMessage("Créer un nouveau champ"),
+    "updateField" : MessageLookupByLibrary.simpleMessage("Mettre à jour un champ"),
     "fieldName" : MessageLookupByLibrary.simpleMessage("Nom du champ"),
-    "fieldCrop" : MessageLookupByLibrary.simpleMessage("Culture"),
-    "fieldCropVariety" : MessageLookupByLibrary.simpleMessage("Variété"),
-    "nameIsRequired" : MessageLookupByLibrary.simpleMessage("Un nom est requis"),
-    "cropIsRequired" : MessageLookupByLibrary.simpleMessage("Une culture est requise"),
-    "fieldNameIsRequired" : MessageLookupByLibrary.simpleMessage("Le nom du champ est requis"),
-    "varietyIsRequired" : MessageLookupByLibrary.simpleMessage("Une variété est requise"),
+    "fieldCrop" : MessageLookupByLibrary.simpleMessage("Culture du champ"),
+    "fieldCropVariety" : MessageLookupByLibrary.simpleMessage("Variété de la culture"),
+    "nameIsRequired" : MessageLookupByLibrary.simpleMessage("Le Nom est requis"),
+    "cropIsRequired" : MessageLookupByLibrary.simpleMessage("La culture est requis"),
+    "fieldNameIsRequired" : MessageLookupByLibrary.simpleMessage("Le nom du champs est requis"),
+    "varietyIsRequired" : MessageLookupByLibrary.simpleMessage("La variété est requis"),
     "deleteFieldConfirm" : MessageLookupByLibrary.simpleMessage("Êtes-vous certain de vouloir supprimer ce champ?"),
     "createNewContainer" : MessageLookupByLibrary.simpleMessage("Créer un nouveau contenant"),
-    "updateContainer" : MessageLookupByLibrary.simpleMessage("Mettre à jour les contenants"),
+    "updateContainer" : MessageLookupByLibrary.simpleMessage("Mettre à jour un contenant"),
     "containerName" : MessageLookupByLibrary.simpleMessage("Nom du contenant"),
     "deleteContainerConfirm" : MessageLookupByLibrary.simpleMessage("Êtes-vous certain de vouloir supprimer ce contenant?"),
-    "chooseColor" : MessageLookupByLibrary.simpleMessage("Veuillez choisir une couleur"),
+    "chooseColor" : MessageLookupByLibrary.simpleMessage("Choisir une couleur"),
     "ok" : MessageLookupByLibrary.simpleMessage("Ok"),
-    "lowValueIsEmpty" : MessageLookupByLibrary.simpleMessage("Veuillez entrer une valeur minimale"),
-    "lowValueShouldBeNumber" : MessageLookupByLibrary.simpleMessage("Veuillez entrer un chiffre"),
-    "highValueIsEmpty" : MessageLookupByLibrary.simpleMessage("Veuillez entrer une valeur élevée"),
-    "highValueShouldBeNumber" : MessageLookupByLibrary.simpleMessage("Veuillez entrer un chiffre"),
-    "highValueShouldBeBigger" : MessageLookupByLibrary.simpleMessage("La valeur élevé entré doit être plus grande que la valeur minimale entrée"),
+    "lowValueIsEmpty" : MessageLookupByLibrary.simpleMessage("La valeur basse est vide"),
+    "lowValueShouldBeNumber" : MessageLookupByLibrary.simpleMessage("La valeur basse devrait être un nombre"),
+    "highValueIsEmpty" : MessageLookupByLibrary.simpleMessage("La valeur haute est vide"),
+    "highValueShouldBeNumber" : MessageLookupByLibrary.simpleMessage("La valeur haute devrait être un nombre"),
+    "highValueShouldBeBigger" : MessageLookupByLibrary.simpleMessage("La valeur haute devrait être plus grande que la valeur basse"),
     "nfcSettings" : MessageLookupByLibrary.simpleMessage("Paramètres NFC"),
     "fields" : MessageLookupByLibrary.simpleMessage("Champs"),
     "editField" : MessageLookupByLibrary.simpleMessage("Modifier le champ"),
-    "containers" : MessageLookupByLibrary.simpleMessage("Contenants"),
+    "containers" : MessageLookupByLibrary.simpleMessage("Contenant"),
     "deleteContainer" : MessageLookupByLibrary.simpleMessage("Supprimer le contenant"),
-    "containerHour" : MessageLookupByLibrary.simpleMessage("Contenants/heure"),
-    "highDefault" : MessageLookupByLibrary.simpleMessage("Élevé (Défaut: 3+)"),
-    "mediumDefault" : MessageLookupByLibrary.simpleMessage("Moyenne (Défaut: 2.5+)"),
-    "lowDefault" : MessageLookupByLibrary.simpleMessage("Minimale (Défaut: 2.5-)"),
-    "reportTime" : MessageLookupByLibrary.simpleMessage("Heure du rapport"),
-    "lastUpdated" : MessageLookupByLibrary.simpleMessage("Dernières modifications"),
-    "aboutApp" : MessageLookupByLibrary.simpleMessage("À propos"),
+    "containerHour" : MessageLookupByLibrary.simpleMessage("Contenant/heure"),
+    "highDefault" : MessageLookupByLibrary.simpleMessage("Moyen (Défaut: 2.5+)"),
+    "mediumDefault" : MessageLookupByLibrary.simpleMessage("Medium (Default: 2.5+)"),
+    "lowDefault" : MessageLookupByLibrary.simpleMessage("Basse (Défaut: 2.5-)"),
+    "reportTime" : MessageLookupByLibrary.simpleMessage("Rapport des heures travaillées"),
+    "lastUpdated" : MessageLookupByLibrary.simpleMessage("Dernière mise à jour"),
+    "aboutApp" : MessageLookupByLibrary.simpleMessage("À propos de l'application"),
     "profile" : MessageLookupByLibrary.simpleMessage("Profil"),
     "oldPassword" : MessageLookupByLibrary.simpleMessage("Ancien mot de passe"),
     "newPassword" : MessageLookupByLibrary.simpleMessage("Nouveau mot de passe"),
     "company" : MessageLookupByLibrary.simpleMessage("Compagnie"),
-    "companyPlan" : MessageLookupByLibrary.simpleMessage("Abonnement"),
-    "receiveRevisionNotification" : MessageLookupByLibrary.simpleMessage("Recevoir les notifications de révision"),
-    "receivePunchNotification" : MessageLookupByLibrary.simpleMessage("Recevoir les notifications de punch"),
+    "receiveRevisionNotification" : MessageLookupByLibrary.simpleMessage("Recevoir une notification de révision"),
+    "receivePunchNotification" : MessageLookupByLibrary.simpleMessage("Recevoir une notification de PUNCH"),
     "notifications" : MessageLookupByLibrary.simpleMessage("Notifications"),
-    "editEmployee" : MessageLookupByLibrary.simpleMessage("Modifier l'employé"),
-    "deleteEmployee" : MessageLookupByLibrary.simpleMessage("Supprimer l'employé"),
-    "deleteEmployeeConfirm" : MessageLookupByLibrary.simpleMessage("Êtes-vous certain de vouloir supprimer cet employé?"),
-    "inOut" : MessageLookupByLibrary.simpleMessage("Entré/Sortie"),
-    "employeeLogIn" : MessageLookupByLibrary.simpleMessage("Portail d'employé"),
+    "editEmployee" : MessageLookupByLibrary.simpleMessage("Modifier la fiche d'employé"),
+    "deleteEmployee" : MessageLookupByLibrary.simpleMessage("Supprimer la fiche de l'employé"),
+    "deleteEmployeeConfirm" : MessageLookupByLibrary.simpleMessage("Êtes-vous certain de vouloir supprimer cette fiche d'employé ?"),
+    "inOut" : MessageLookupByLibrary.simpleMessage("IN/OUT"),
+    "employeeLogIn" : MessageLookupByLibrary.simpleMessage("Employé LOG IN"),
     "empty" : MessageLookupByLibrary.simpleMessage("Vide"),
-    "employeeLogOut" : MessageLookupByLibrary.simpleMessage("Se déconnecter"),
-    "createNewEmployee" : MessageLookupByLibrary.simpleMessage("Créer un nouvel employé"),
+    "employeeLogOut" : MessageLookupByLibrary.simpleMessage("Employé LOG OUT"),
+    "createNewEmployee" : MessageLookupByLibrary.simpleMessage("Créer une nouvelle fiche d'employé"),
     "photoCropper" : MessageLookupByLibrary.simpleMessage("Rogner la photo"),
     "success" : MessageLookupByLibrary.simpleMessage("Succès!"),
-    "camera" : MessageLookupByLibrary.simpleMessage("Caméra"),
+    "camera" : MessageLookupByLibrary.simpleMessage("Camera"),
     "gallery" : MessageLookupByLibrary.simpleMessage("Gallerie"),
-    "passwordPin" : MessageLookupByLibrary.simpleMessage("Mot de passe (PIN)"),
+    "passwordPin" : MessageLookupByLibrary.simpleMessage("Mot de passe (NIP)"),
     "address" : MessageLookupByLibrary.simpleMessage("Adresse"),
     "employeeFunction" : MessageLookupByLibrary.simpleMessage("Fonction de l'employé"),
-    "startDate" : MessageLookupByLibrary.simpleMessage("Date de début"),
-    "salary" : MessageLookupByLibrary.simpleMessage("Salaire"),
+    "startDate" : MessageLookupByLibrary.simpleMessage("Salaire"),
+    "salary" : MessageLookupByLibrary.simpleMessage("Salary"),
     "birthday" : MessageLookupByLibrary.simpleMessage("Date de naissance"),
     "chooseLanguage" : MessageLookupByLibrary.simpleMessage("Choisir une langue"),
-    "hasLunchBreak" : MessageLookupByLibrary.simpleMessage("Pause diner de 30 minutes non incluse"),
+    "hasLunchBreak" : MessageLookupByLibrary.simpleMessage("Pause de 30 minutes"),
     "edit" : MessageLookupByLibrary.simpleMessage("Modifier"),
-    "editEmployeePunch" : MessageLookupByLibrary.simpleMessage("Modifier l'heure d'entré de l'employé"),
-    "correctLunchTime" : MessageLookupByLibrary.simpleMessage("Temps de diner correcte"),
-    "incorrectLunchTime" : MessageLookupByLibrary.simpleMessage("Temps de diner incorrecte"),
-    "correctPunchTime" : MessageLookupByLibrary.simpleMessage("Entrée correct"),
-    "incorrectPunchTime" : MessageLookupByLibrary.simpleMessage("Entrée inccorect"),
-    "dailyLogs" : MessageLookupByLibrary.simpleMessage("Journal quotidien"),
-    "totalHours" : MessageLookupByLibrary.simpleMessage("Heures totales"),
-    "timeSheet" : MessageLookupByLibrary.simpleMessage("Rapport"),
-    "in" : MessageLookupByLibrary.simpleMessage("Entré"),
-    "out" : MessageLookupByLibrary.simpleMessage("Sortie"),
-    "pdfNotGenerated" : MessageLookupByLibrary.simpleMessage("Ce fichier PDF n'est pas encore disponible"),
-    "harvestReportNotGenerated" : MessageLookupByLibrary.simpleMessage("Le rapport de récolte n'est pas disponible"),
+    "editEmployeePunch" : MessageLookupByLibrary.simpleMessage("Modifier la feuille de temps de l'employé"),
+    "correctLunchTime" : MessageLookupByLibrary.simpleMessage("Modifier la pause diner d'un employé"),
+    "incorrectLunchTime" : MessageLookupByLibrary.simpleMessage("Pause diner incorrecte"),
+    "correctPunchTime" : MessageLookupByLibrary.simpleMessage("Modifier la feuille de temps d'un employé"),
+    "incorrectPunchTime" : MessageLookupByLibrary.simpleMessage("Feuille de temps incorrecte"),
+    "dailyLogs" : MessageLookupByLibrary.simpleMessage("Journal de bord"),
+    "totalHours" : MessageLookupByLibrary.simpleMessage("Total des heures"),
+    "timeSheet" : MessageLookupByLibrary.simpleMessage("Feuille de temps"),
+    "in" : MessageLookupByLibrary.simpleMessage("IN"),
+    "out" : MessageLookupByLibrary.simpleMessage("OUT"),
+    "pdfNotGenerated" : MessageLookupByLibrary.simpleMessage("Le fichier PDF n'a pas été généré"),
+    "harvestReportNotGenerated" : MessageLookupByLibrary.simpleMessage("Le rapport de récolte n'a pas été généré."),
     "calender" : MessageLookupByLibrary.simpleMessage("Calendrier"),
     "document" : MessageLookupByLibrary.simpleMessage("Document"),
     "setting" : MessageLookupByLibrary.simpleMessage("Paramètres"),
-    "chooseYourCompany" : MessageLookupByLibrary.simpleMessage("Sélectionner une compagnie"),
-    "faceScanLogin" : MessageLookupByLibrary.simpleMessage("Se connecter"),
-    "employeeSignIn" : MessageLookupByLibrary.simpleMessage("Portail d'employé"),
-    "selectYourCompany" : MessageLookupByLibrary.simpleMessage("Sélectionner une compagnie"),
-    "language" : MessageLookupByLibrary.simpleMessage("Language"),
-    "country" : MessageLookupByLibrary.simpleMessage("Pays/région"),
-    "state" : MessageLookupByLibrary.simpleMessage("État/province/région"),
+    "faceScanLogin" : MessageLookupByLibrary.simpleMessage("CONNEXION FACESCAN"),
+    "language" : MessageLookupByLibrary.simpleMessage("Langue"),
+    "country" : MessageLookupByLibrary.simpleMessage("Province"),
+    "state" : MessageLookupByLibrary.simpleMessage("Pays"),
     "city" : MessageLookupByLibrary.simpleMessage("Ville"),
-    "lunchBreakFrom" : MessageLookupByLibrary.simpleMessage("Période de diner de"),
-    "to" : MessageLookupByLibrary.simpleMessage("à"),
+    "lunchBreakFrom" : MessageLookupByLibrary.simpleMessage("Début de la pause dîner"),
+    "to" : MessageLookupByLibrary.simpleMessage("De"),
     "punch" : MessageLookupByLibrary.simpleMessage("Punch"),
-    "at" : MessageLookupByLibrary.simpleMessage("à"),
+    "at" : MessageLookupByLibrary.simpleMessage("À"),
     "total" : MessageLookupByLibrary.simpleMessage("Total"),
     "hours" : MessageLookupByLibrary.simpleMessage("Heures"),
-    "hoursForLunch" : MessageLookupByLibrary.simpleMessage("Heures de pause diner"),
+    "hoursForLunch" : MessageLookupByLibrary.simpleMessage("Heures de dîner"),
     "hourRevisionRequest" : MessageLookupByLibrary.simpleMessage("Demander une révision des heures"),
-    "submit" : MessageLookupByLibrary.simpleMessage("Envoyer"),
+    "submit" : MessageLookupByLibrary.simpleMessage("Soumettre"),
     "week" : MessageLookupByLibrary.simpleMessage("Semaine"),
-    "time" : MessageLookupByLibrary.simpleMessage("Temps"),
-    "noPunch" : MessageLookupByLibrary.simpleMessage("Aucun entrée disponible"),
+    "time" : MessageLookupByLibrary.simpleMessage("Heure"),
+    "noPunch" : MessageLookupByLibrary.simpleMessage("Aucun PUNCH"),
     "logs" : MessageLookupByLibrary.simpleMessage("Journal"),
-    "askRevision" : MessageLookupByLibrary.simpleMessage("Pour demander une révision, sélectionner la ligne avec l'erreur"),
+    "askRevision" : MessageLookupByLibrary.simpleMessage("Pour demander une révision, sélectionnez la ligne en erreur"),
     "enterPinCode" : MessageLookupByLibrary.simpleMessage("Veuillez entrer votre code PIN"),
     "employeeName" : MessageLookupByLibrary.simpleMessage("Nom de l'employé"),
+    "accept" : MessageLookupByLibrary.simpleMessage("Accepter"),
+    "decline" : MessageLookupByLibrary.simpleMessage("Refuser"),
+    "punchIn" : MessageLookupByLibrary.simpleMessage("Punch In"),
+    "punchOut" : MessageLookupByLibrary.simpleMessage("Punch Out"),
+    "pin" : MessageLookupByLibrary.simpleMessage("Code NIP"),
+    "harvestReport" : MessageLookupByLibrary.simpleMessage("Rapport de récolte"),
+    "totalOfTheDay" : MessageLookupByLibrary.simpleMessage("Total de la journée"),
+    "totalOfTheSeason" : MessageLookupByLibrary.simpleMessage("Total de la saison"),
     "nfc" : MessageLookupByLibrary.simpleMessage("NFC"),
-    "searchEmployee" : MessageLookupByLibrary.simpleMessage("Recherche d'employée"),
+    "searchEmployee" : MessageLookupByLibrary.simpleMessage("Rechercher un employé"),
+    "selectProject" : MessageLookupByLibrary.simpleMessage("Sélectionner un projet"),
+    "selectTask" : MessageLookupByLibrary.simpleMessage("Sélectionner une tâche"),
+    "youAreWorkingOn" : MessageLookupByLibrary.simpleMessage("Vous travaillez présentement sur"),
+    "youAreWorkingOnCall" : MessageLookupByLibrary.simpleMessage("Vous travaillez sur un APPEL."),
+    "project" : MessageLookupByLibrary.simpleMessage("Projet"),
+    "activity" : MessageLookupByLibrary.simpleMessage("Activité"),
+    "startTime" : MessageLookupByLibrary.simpleMessage("Temps de début"),
+    "endTime" : MessageLookupByLibrary.simpleMessage("Temps de fin"),
+    "dailySchedule" : MessageLookupByLibrary.simpleMessage("Horaire journalière"),
+    "editWorkHistory" : MessageLookupByLibrary.simpleMessage("Modifier l'historique de travail"),
+    "start" : MessageLookupByLibrary.simpleMessage("Débuter"),
+    "resume" : MessageLookupByLibrary.simpleMessage("Résumer"),
+    "end" : MessageLookupByLibrary.simpleMessage("Fin"),
+    "pressToStart" : MessageLookupByLibrary.simpleMessage("Appuyer pour débuter"),
+    "pressToEnd" : MessageLookupByLibrary.simpleMessage("Appuyer pour terminer"),
+    "pressToAskRevision" : MessageLookupByLibrary.simpleMessage("Appuyer pour demander une révision"),
+    "scheduleRevision" : MessageLookupByLibrary.simpleMessage("Révision de l'horaire"),
+    "todo" : MessageLookupByLibrary.simpleMessage("À faire"),
+    "notes" : MessageLookupByLibrary.simpleMessage("Notes"),
+    "dispatch" : MessageLookupByLibrary.simpleMessage("Dispatch"),
+    "addSchedule" : MessageLookupByLibrary.simpleMessage("Ajouter un horaire"),
+    "type" : MessageLookupByLibrary.simpleMessage("Type"),
+    "priority" : MessageLookupByLibrary.simpleMessage("Priorité"),
+    "call" : MessageLookupByLibrary.simpleMessage("Call"),
+    "shop" : MessageLookupByLibrary.simpleMessage("Shop"),
+    "schedule" : MessageLookupByLibrary.simpleMessage("Horraire"),
+    "selectCall" : MessageLookupByLibrary.simpleMessage("Selectionner un CALL"),
+    "startCall" : MessageLookupByLibrary.simpleMessage("Débuter un CALL"),
+    "task" : MessageLookupByLibrary.simpleMessage("Tâches"),
+    "note" : MessageLookupByLibrary.simpleMessage("Notes"),
+    "selectSchedule" : MessageLookupByLibrary.simpleMessage("Sélectionner un horaire"),
+    "shift" : MessageLookupByLibrary.simpleMessage("Quart de travail"),
+    "startSchedule" : MessageLookupByLibrary.simpleMessage("Débuter un horaire"),
+    "description" : MessageLookupByLibrary.simpleMessage("Description"),
+    "youMustWriteDescription" : MessageLookupByLibrary.simpleMessage("Veuillez entrer une description"),
+    "startWorking" : MessageLookupByLibrary.simpleMessage("Début de travail"),
+    "breaks" : MessageLookupByLibrary.simpleMessage("Pause"),
+    "incorrectBreakTime" : MessageLookupByLibrary.simpleMessage("Heure de pause incorrecte"),
+    "correctBreakTime" : MessageLookupByLibrary.simpleMessage("Heure de pause correcte"),
+    "length" : MessageLookupByLibrary.simpleMessage("Durée"),
+    "breakLengthCanNotBeZero" : MessageLookupByLibrary.simpleMessage("La pause ne peut pas être à 0"),
+    "invalidBreakLength" : MessageLookupByLibrary.simpleMessage("Durée de pause invalide"),
+    "dailyTasks" : MessageLookupByLibrary.simpleMessage("Tâches quotidiennes"),
+    "schedules" : MessageLookupByLibrary.simpleMessage("Horaire"),
+    "calls" : MessageLookupByLibrary.simpleMessage("Calls"),
+    "incorrect" : MessageLookupByLibrary.simpleMessage("Incorrecte"),
+    "correct" : MessageLookupByLibrary.simpleMessage("Correcte"),
+    "selectPriority" : MessageLookupByLibrary.simpleMessage("Sélectionnez la priorité"),
+    "deletePunchConfirm" : MessageLookupByLibrary.simpleMessage("Voulez-vous vraiment supprimer ce PUNCH?"),
+    "deleteWorkConfirm" : MessageLookupByLibrary.simpleMessage("Voulez-vous vraiment supprimer ce travail?"),
+    "deleteBreakConfirm" : MessageLookupByLibrary.simpleMessage("Voulez-vous vraiment supprimer cette pause?"),
+    "breakTime" : MessageLookupByLibrary.simpleMessage("Temps de pause"),
+    "newVersionAvailable" : MessageLookupByLibrary.simpleMessage("Une nouvelle version est maintenant disponible. Souhaitez-vous mettre à jour vers la nouvelle version ?"),
+    "update" : MessageLookupByLibrary.simpleMessage("Mise à jour"),
+    "callRevision" : MessageLookupByLibrary.simpleMessage("Révision des Appels"),
+    "addCall" : MessageLookupByLibrary.simpleMessage("Ajouter un APPEL"),
+    "bugReport" : MessageLookupByLibrary.simpleMessage("Rapport d'erreurs"),
+    "make" : MessageLookupByLibrary.simpleMessage("Make"),
+    "model" : MessageLookupByLibrary.simpleMessage("Modèle"),
+    "os" : MessageLookupByLibrary.simpleMessage("SO (Système d'opération)"),
+    "version" : MessageLookupByLibrary.simpleMessage("Version"),
+    "brand" : MessageLookupByLibrary.simpleMessage("Marque"),
+    "name" : MessageLookupByLibrary.simpleMessage("Nom"),
+    "system" : MessageLookupByLibrary.simpleMessage("Système"),
+    "bugReportDescription" : MessageLookupByLibrary.simpleMessage("Veuillez inclure une description de vos erreurs ici..."),
+    "punchTime" : MessageLookupByLibrary.simpleMessage("Heure de PUNCH"),
+    "revisionHasBeenSent" : MessageLookupByLibrary.simpleMessage("Une demande de révision a été envoyée"),
+    "sent" : MessageLookupByLibrary.simpleMessage("Envoyé"),
+    "accepted" : MessageLookupByLibrary.simpleMessage("Accepté"),
+    "declined" : MessageLookupByLibrary.simpleMessage("Refuser"),
+    "tapToSubmitDescription" : MessageLookupByLibrary.simpleMessage("Appuyez pour envoyer une description"),
+    "invalidNFC" : MessageLookupByLibrary.simpleMessage("Impossible de trouver le NFC"),
+    "notAllowedNFC" : MessageLookupByLibrary.simpleMessage("Autoriser l'appareil à utiliser le NFC"),
+    "deleteHarvestConfirm" : MessageLookupByLibrary.simpleMessage("Voulez-vous vraiment supprimer cette récolte?"),
+    "thankYouForReporting" : MessageLookupByLibrary.simpleMessage("Merci d'avoir signalé un bug"),
+    "open" : MessageLookupByLibrary.simpleMessage("Ouvrir"),
+    "callDetail" : MessageLookupByLibrary.simpleMessage("Détail de l'Appel"),
+    "date" : MessageLookupByLibrary.simpleMessage("Date"),
+    "thisPunchHasBeenSentAlready" : MessageLookupByLibrary.simpleMessage("Ce PUNCH a déjà été envoyé"),
+    "canNotSendRevisionAfterStart" : MessageLookupByLibrary.simpleMessage("Impossible d'envoyer une révision après le démarrage"),
+    "canNotEditDeleteCall" : MessageLookupByLibrary.simpleMessage("Impossible de modifier ou de supprimer un APPEL traité."),
+    "workingNow" : MessageLookupByLibrary.simpleMessage("En travail"),
+    "works" : MessageLookupByLibrary.simpleMessage("Travail"),
+    "workRevision" : MessageLookupByLibrary.simpleMessage("Révision du travail"),
+    "revisions" : MessageLookupByLibrary.simpleMessage("Révision"),
+    "deviceIdForPunch" : MessageLookupByLibrary.simpleMessage("NO de périphérique pour le PUNCH"),
+    "clickToCopy" : MessageLookupByLibrary.simpleMessage("Clic pour copier"),
+    "deviceIdCopied" : MessageLookupByLibrary.simpleMessage("Identifiant de l'appareil copié"),
+    "startManualBreak" : MessageLookupByLibrary.simpleMessage("Débuter la pause manuellement"),
+    "endManualBreak" : MessageLookupByLibrary.simpleMessage("Terminer la pause manuellement"),
+    "somethingWentWrong" : MessageLookupByLibrary.simpleMessage("Quelque chose s'est mal passé"),
+    "signIn" : MessageLookupByLibrary.simpleMessage("Se connecter"),
+    "revisionDescription" : MessageLookupByLibrary.simpleMessage("Révision de la description"),
+    "revisionDescriptionSubmitted" : MessageLookupByLibrary.simpleMessage("Révision de la description soumise"),
+    "editHarvest" : MessageLookupByLibrary.simpleMessage("modifier la récolte"),
+    "welcomeToFacePunch" : MessageLookupByLibrary.simpleMessage("Bienvenue sur Facepunch"),
+    "theBestEmployeeClockingSystem" : MessageLookupByLibrary.simpleMessage("Le meilleur système de pointage des employés"),
+    "employeePortal" : MessageLookupByLibrary.simpleMessage("Portail employé"),
+    "team" : MessageLookupByLibrary.simpleMessage("équipe"),
+    "my" : MessageLookupByLibrary.simpleMessage("Mes"),
+    "day" : MessageLookupByLibrary.simpleMessage("Journée"),
   };
 }
