@@ -93,7 +93,7 @@ class PunchDateUtils {
   static int calculateMaxWeeksNumberMonthly(DateTime startDate, DateTime endDate) {
     int monthsNumber = calculateMonthsDifference(startDate, endDate);
 
-    List<int> weeksNumbersMonthly = List();
+    List<int> weeksNumbersMonthly = [];
 
     if (monthsNumber == 0) {
       return calculateWeeksNumber(startDate, endDate);
@@ -166,24 +166,67 @@ class PunchDateUtils {
     return date.add(Duration(days: DateTime.daysPerWeek - date.weekday-1));
   }
 
-  static String getDateString(DateTime date){
+  static String getDateString(date){
+    if(date is String){
+      return DateFormat(DateFormat.YEAR_ABBR_MONTH_DAY).format(DateTime.parse(date));
+    }
     return DateFormat(DateFormat.YEAR_ABBR_MONTH_DAY).format(date);
   }
 
-  static String inputDateString(DateTime date){
+  static String inputDateString(date){
+    if(date is String){
+      return DateFormat('dd/MM/y').format(DateTime.parse(date));
+    }
     return DateFormat('dd/MM/y').format(date);
   }
 
-  static String getTimeString(DateTime date){
-    return DateFormat("kk:mm").format(date);
+  static String getTimeString(date){
+    if(date is String){
+      if(date.isEmpty){
+        return '';
+      }
+      return DateFormat("hh:mm").format(DateTime.parse(date));
+    }
+
+    return DateFormat("hh:mm").format(date);
   }
 
-  static String getTimeSecondString(DateTime date){
-    return DateFormat("kk:mm:ss").format(date);
+  static String getTimeSecondString(date){
+    if(date is String){
+      return DateFormat("hh:mm:ss").format(DateTime.parse(date));
+    }
+    return DateFormat("hh:mm:ss").format(date);
   }
 
-  static String get12TimeString(String time){
-    if(time==null || time.isEmpty)return '--:--';
-    return DateFormat("hh:mm a").format(DateTime.parse('2021-01-01 $time'));
+  static String toDateTime(time){
+    if(time==null)return '--:--';
+    if(time is String){
+      return DateFormat("MMM, d hh:mm a").format(DateTime.parse(time));
+    }
+    return DateFormat("MMM, d hh:mm a").format(time);
+  }
+
+  static String toStandardDateTime(date){
+    if(date is String){
+      return DateFormat("y-MM-dd hh:mm:ss").format(DateTime.parse(date));
+    }
+    return DateFormat("y-MM-dd hh:mm:ss").format(date);
+  }
+
+  static String toDateString(date){
+    if(date is String){
+      return DateFormat("y-MM-dd").format(DateTime.parse(date));
+    }
+    return DateFormat("y-MM-dd").format(date);
+  }
+
+  static String toDateHourMinute(date){
+    if(date is String){
+      if(date.isEmpty){
+        return '';
+      }
+      return DateFormat("y-MM-dd hh:mm").format(DateTime.parse(date));
+    }
+    return DateFormat("y-MM-dd hh:mm").format(date);
   }
 }
